Network.AWS.Route53

Description

Amazon Route 53 is a scalable Domain Name System (DNS) web service. It provides secure and reliable routing to your infrastructure that uses Amazon Web Services (AWS) products, such as Amazon Elastic Compute Cloud (Amazon EC2), Elastic Load Balancing, or Amazon Simple Storage Service (Amazon S3). You can also use Amazon Route 53 to route users to your infrastructure outside of AWS.

Documentation

module Network.AWS.Route53.AssociateVPCWithHostedZone

module Network.AWS.Route53.ChangeResourceRecordSets

module Network.AWS.Route53.ChangeTagsForResource

module Network.AWS.Route53.CreateHealthCheck

module Network.AWS.Route53.CreateHostedZone

module Network.AWS.Route53.CreateReusableDelegationSet

module Network.AWS.Route53.DeleteHealthCheck

module Network.AWS.Route53.DeleteHostedZone

module Network.AWS.Route53.DeleteReusableDelegationSet

module Network.AWS.Route53.DisassociateVPCFromHostedZone

module Network.AWS.Route53.GetChange

module Network.AWS.Route53.GetCheckerIpRanges

module Network.AWS.Route53.GetGeoLocation

module Network.AWS.Route53.GetHealthCheck

module Network.AWS.Route53.GetHealthCheckCount

module Network.AWS.Route53.GetHealthCheckLastFailureReason

module Network.AWS.Route53.GetHealthCheckStatus

module Network.AWS.Route53.GetHostedZone

module Network.AWS.Route53.GetHostedZoneCount

module Network.AWS.Route53.GetReusableDelegationSet

module Network.AWS.Route53.ListGeoLocations

module Network.AWS.Route53.ListHealthChecks

module Network.AWS.Route53.ListHostedZones

module Network.AWS.Route53.ListHostedZonesByName

module Network.AWS.Route53.ListResourceRecordSets

module Network.AWS.Route53.ListReusableDelegationSets

module Network.AWS.Route53.ListTagsForResource

module Network.AWS.Route53.ListTagsForResources

module Network.AWS.Route53.Types

module Network.AWS.Route53.UpdateHealthCheck

module Network.AWS.Route53.UpdateHostedZoneComment