Network.AWS.Route53.AssociateVPCWithHostedZone

Description

This action associates a VPC with an hosted zone.

To associate a VPC with an hosted zone, send a POST request to the '2013-04-01hostedzonehosted zone ID/associatevpc resource. The request body must include an XML document with a AssociateVPCWithHostedZoneRequest element. The response returns the AssociateVPCWithHostedZoneResponse element that contains ChangeInfo for you to track the progress of the AssociateVPCWithHostedZoneRequest you made. See GetChange operation for how to track the progress of your change.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_AssociateVPCWithHostedZone.html

Request

data AssociateVPCWithHostedZone

Request constructor

associateVPCWithHostedZone

Arguments

:: Text
avpcwhzHostedZoneId

-> VPC
avpcwhzVPC

-> AssociateVPCWithHostedZone

AssociateVPCWithHostedZone constructor.

The fields accessible through corresponding lenses are:

avpcwhzComment :: Maybe Text

avpcwhzHostedZoneId :: Text

avpcwhzVPC :: VPC

Request lenses

avpcwhzComment :: Lens' AssociateVPCWithHostedZone (Maybe Text)

Optional: Any comments you want to include about a AssociateVPCWithHostedZoneRequest.

avpcwhzHostedZoneId :: Lens' AssociateVPCWithHostedZone Text

The ID of the hosted zone you want to associate your VPC with.

Note that you cannot associate a VPC with a hosted zone that doesn't have an existing VPC association.

avpcwhzVPC :: Lens' AssociateVPCWithHostedZone VPC

The VPC that you want your hosted zone to be associated with.

Response

data AssociateVPCWithHostedZoneResponse

Response constructor

associateVPCWithHostedZoneResponse

Arguments

:: ChangeInfo
avpcwhzrChangeInfo

-> AssociateVPCWithHostedZoneResponse

AssociateVPCWithHostedZoneResponse constructor.

The fields accessible through corresponding lenses are:

avpcwhzrChangeInfo :: ChangeInfo

Response lenses

avpcwhzrChangeInfo :: Lens' AssociateVPCWithHostedZoneResponse ChangeInfo

A complex type that contains the ID, the status, and the date and time of your AssociateVPCWithHostedZoneRequest.