Network.AWS.Route53.ChangeTagsForResource

Description

http://docs.aws.amazon.com/Route53/latest/APIReference/API_ChangeTagsForResource.html

Request

data ChangeTagsForResource

Request constructor

changeTagsForResource

Arguments

:: TagResourceType
ctfrResourceType

-> Text
ctfrResourceId

-> NonEmpty Tag
ctfrAddTags

-> NonEmpty Text
ctfrRemoveTagKeys

-> ChangeTagsForResource

ChangeTagsForResource constructor.

The fields accessible through corresponding lenses are:

ctfrAddTags :: NonEmpty Tag

ctfrRemoveTagKeys :: NonEmpty Text

ctfrResourceId :: Text

ctfrResourceType :: TagResourceType

Request lenses

ctfrAddTags :: Lens' ChangeTagsForResource (NonEmpty Tag)

A complex type that contains a list of Tag elements. Each Tag element identifies a tag that you want to add or update for the specified resource.

ctfrRemoveTagKeys :: Lens' ChangeTagsForResource (NonEmpty Text)

A list of Tag keys that you want to remove from the specified resource.

ctfrResourceId :: Lens' ChangeTagsForResource Text

The ID of the resource for which you want to add, change, or delete tags.

ctfrResourceType :: Lens' ChangeTagsForResource TagResourceType

The type of the resource.

The resource type for health checks is healthcheck.

The resource type for hosted zones is hostedzone.

Response

data ChangeTagsForResourceResponse

Response constructor

changeTagsForResourceResponse :: ChangeTagsForResourceResponse

ChangeTagsForResourceResponse constructor.