Network.AWS.Route53.CreateReusableDelegationSet

Description

This action creates a reusable delegationSet.

To create a new reusable delegationSet, send a POST request to the '2013-04-01/delegationset' resource. The request body must include an XML document with a CreateReusableDelegationSetRequest element. The response returns the CreateReusableDelegationSetResponse element that contains metadata about the delegationSet.

If the optional parameter HostedZoneId is specified, it marks the delegationSet associated with that particular hosted zone as reusable.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_CreateReusableDelegationSet.html

Request

data CreateReusableDelegationSet

Request constructor

createReusableDelegationSet

Arguments

:: Text
crdsCallerReference

-> CreateReusableDelegationSet

CreateReusableDelegationSet constructor.

The fields accessible through corresponding lenses are:

crdsCallerReference :: Text

crdsHostedZoneId :: Maybe Text

Request lenses

crdsCallerReference :: Lens' CreateReusableDelegationSet Text

A unique string that identifies the request and that allows failed CreateReusableDelegationSet requests to be retried without the risk of executing the operation twice. You must use a unique CallerReference string every time you create a reusable delegation set. CallerReference can be any unique string; you might choose to use a string that identifies your project, such as DNSMigration_01.

Valid characters are any Unicode code points that are legal in an XML 1.0 document. The UTF-8 encoding of the value must be less than 128 bytes.

crdsHostedZoneId :: Lens' CreateReusableDelegationSet (Maybe Text)

The ID of the hosted zone whose delegation set you want to mark as reusable. It is an optional parameter.

Response

data CreateReusableDelegationSetResponse

Response constructor

createReusableDelegationSetResponse

Arguments

:: DelegationSet
crdsrDelegationSet

-> Text
crdsrLocation

-> CreateReusableDelegationSetResponse

CreateReusableDelegationSetResponse constructor.

The fields accessible through corresponding lenses are:

crdsrDelegationSet :: DelegationSet

crdsrLocation :: Text

Response lenses

crdsrDelegationSet :: Lens' CreateReusableDelegationSetResponse DelegationSet

A complex type that contains name server information.

crdsrLocation :: Lens' CreateReusableDelegationSetResponse Text

The unique URL representing the new reusbale delegation set.