Network.AWS.Route53.CreateTrafficPolicy

Description

Creates a traffic policy, which you use to create multiple DNS resource record sets for one domain name (such as example.com) or one subdomain name (such as www.example.com).

Creating a Request

createTrafficPolicy

Arguments

:: Text
ctpName

-> Text
ctpDocument

-> CreateTrafficPolicy

Creates a value of CreateTrafficPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctpComment - (Optional) Any comments that you want to include about the traffic policy.

ctpName - The name of the traffic policy.

ctpDocument - The definition of this traffic policy in JSON format. For more information, see Traffic Policy Document Format .

data CreateTrafficPolicy

A complex type that contains information about the traffic policy that you want to create.

See: createTrafficPolicy smart constructor.

Request Lenses

ctpComment :: Lens' CreateTrafficPolicy (Maybe Text)

(Optional) Any comments that you want to include about the traffic policy.

ctpName :: Lens' CreateTrafficPolicy Text

The name of the traffic policy.

ctpDocument :: Lens' CreateTrafficPolicy Text

The definition of this traffic policy in JSON format. For more information, see Traffic Policy Document Format .

Destructuring the Response

createTrafficPolicyResponse

Arguments

:: Int
ctprsResponseStatus

-> TrafficPolicy
ctprsTrafficPolicy

-> Text
ctprsLocation

-> CreateTrafficPolicyResponse

Creates a value of CreateTrafficPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctprsResponseStatus - -- | The response status code.

ctprsTrafficPolicy - A complex type that contains settings for the new traffic policy.

ctprsLocation - A unique URL that represents a new traffic policy.

data CreateTrafficPolicyResponse

A complex type that contains the response information for the CreateTrafficPolicy request.

See: createTrafficPolicyResponse smart constructor.

Response Lenses

ctprsResponseStatus :: Lens' CreateTrafficPolicyResponse Int

- | The response status code.

ctprsTrafficPolicy :: Lens' CreateTrafficPolicyResponse TrafficPolicy

A complex type that contains settings for the new traffic policy.

ctprsLocation :: Lens' CreateTrafficPolicyResponse Text

A unique URL that represents a new traffic policy.