Network.AWS.Route53.CreateTrafficPolicyInstance

Description

Creates resource record sets in a specified hosted zone based on the settings in a specified traffic policy version. In addition, CreateTrafficPolicyInstance associates the resource record sets with a specified domain name (such as example.com) or subdomain name (such as www.example.com). Amazon Route 53 responds to DNS queries for the domain or subdomain name by using the resource record sets that CreateTrafficPolicyInstance created.

Creating a Request

createTrafficPolicyInstance

Arguments

:: ResourceId
ctpiHostedZoneId

-> Text
ctpiName

-> Natural
ctpiTTL

-> Text
ctpiTrafficPolicyId

-> Natural
ctpiTrafficPolicyVersion

-> CreateTrafficPolicyInstance

Creates a value of CreateTrafficPolicyInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctpiHostedZoneId - The ID of the hosted zone in which you want Amazon Route 53 to create resource record sets by using the configuration in a traffic policy.

ctpiName - The domain name (such as example.com) or subdomain name (such as www.example.com) for which Amazon Route 53 responds to DNS queries by using the resource record sets that Amazon Route 53 creates for this traffic policy instance.

ctpiTTL - (Optional) The TTL that you want Amazon Route 53 to assign to all of the resource record sets that it creates in the specified hosted zone.

ctpiTrafficPolicyId - The ID of the traffic policy that you want to use to create resource record sets in the specified hosted zone.

ctpiTrafficPolicyVersion - The version of the traffic policy that you want to use to create resource record sets in the specified hosted zone.

data CreateTrafficPolicyInstance

A complex type that contains information about the resource record sets that you want to create based on a specified traffic policy.

See: createTrafficPolicyInstance smart constructor.

Request Lenses

ctpiHostedZoneId :: Lens' CreateTrafficPolicyInstance ResourceId

The ID of the hosted zone in which you want Amazon Route 53 to create resource record sets by using the configuration in a traffic policy.

ctpiName :: Lens' CreateTrafficPolicyInstance Text

The domain name (such as example.com) or subdomain name (such as www.example.com) for which Amazon Route 53 responds to DNS queries by using the resource record sets that Amazon Route 53 creates for this traffic policy instance.

ctpiTTL :: Lens' CreateTrafficPolicyInstance Natural

(Optional) The TTL that you want Amazon Route 53 to assign to all of the resource record sets that it creates in the specified hosted zone.

ctpiTrafficPolicyId :: Lens' CreateTrafficPolicyInstance Text

The ID of the traffic policy that you want to use to create resource record sets in the specified hosted zone.

ctpiTrafficPolicyVersion :: Lens' CreateTrafficPolicyInstance Natural

The version of the traffic policy that you want to use to create resource record sets in the specified hosted zone.

Destructuring the Response

createTrafficPolicyInstanceResponse

Arguments

:: Int
ctpirsResponseStatus

-> TrafficPolicyInstance
ctpirsTrafficPolicyInstance

-> Text
ctpirsLocation

-> CreateTrafficPolicyInstanceResponse

Creates a value of CreateTrafficPolicyInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctpirsResponseStatus - -- | The response status code.

ctpirsTrafficPolicyInstance - A complex type that contains settings for the new traffic policy instance.

ctpirsLocation - A unique URL that represents a new traffic policy instance.

data CreateTrafficPolicyInstanceResponse

A complex type that contains the response information for the CreateTrafficPolicyInstance request.

See: createTrafficPolicyInstanceResponse smart constructor.

Response Lenses

ctpirsResponseStatus :: Lens' CreateTrafficPolicyInstanceResponse Int

- | The response status code.

ctpirsTrafficPolicyInstance :: Lens' CreateTrafficPolicyInstanceResponse TrafficPolicyInstance

A complex type that contains settings for the new traffic policy instance.

ctpirsLocation :: Lens' CreateTrafficPolicyInstanceResponse Text

A unique URL that represents a new traffic policy instance.