Network.AWS.Route53.CreateTrafficPolicyVersion

Description

Creates a new version of an existing traffic policy. When you create a new version of a traffic policy, you specify the ID of the traffic policy that you want to update and a JSON-formatted document that describes the new version. You use traffic policies to create multiple DNS resource record sets for one domain name (such as example.com) or one subdomain name (such as www.example.com). You can create a maximum of 1000 versions of a traffic policy. If you reach the limit and need to create another version, you'll need to start a new traffic policy.

Creating a Request

createTrafficPolicyVersion

Arguments

:: Text
ctpvId

-> Text
ctpvDocument

-> CreateTrafficPolicyVersion

Creates a value of CreateTrafficPolicyVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctpvComment - The comment that you specified in the CreateTrafficPolicyVersion request, if any.

ctpvId - The ID of the traffic policy for which you want to create a new version.

ctpvDocument - The definition of this version of the traffic policy, in JSON format. You specified the JSON in the CreateTrafficPolicyVersion request. For more information about the JSON format, see CreateTrafficPolicy .

data CreateTrafficPolicyVersion

A complex type that contains information about the traffic policy that you want to create a new version for.

See: createTrafficPolicyVersion smart constructor.

Request Lenses

ctpvComment :: Lens' CreateTrafficPolicyVersion (Maybe Text)

The comment that you specified in the CreateTrafficPolicyVersion request, if any.

ctpvId :: Lens' CreateTrafficPolicyVersion Text

The ID of the traffic policy for which you want to create a new version.

ctpvDocument :: Lens' CreateTrafficPolicyVersion Text

The definition of this version of the traffic policy, in JSON format. You specified the JSON in the CreateTrafficPolicyVersion request. For more information about the JSON format, see CreateTrafficPolicy .

Destructuring the Response

createTrafficPolicyVersionResponse

Arguments

:: Int
ctpvrsResponseStatus

-> TrafficPolicy
ctpvrsTrafficPolicy

-> Text
ctpvrsLocation

-> CreateTrafficPolicyVersionResponse

Creates a value of CreateTrafficPolicyVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctpvrsResponseStatus - -- | The response status code.

ctpvrsTrafficPolicy - A complex type that contains settings for the new version of the traffic policy.

ctpvrsLocation - A unique URL that represents a new traffic policy version.

data CreateTrafficPolicyVersionResponse

A complex type that contains the response information for the CreateTrafficPolicyVersion request.

See: createTrafficPolicyVersionResponse smart constructor.

Response Lenses

ctpvrsResponseStatus :: Lens' CreateTrafficPolicyVersionResponse Int

- | The response status code.

ctpvrsTrafficPolicy :: Lens' CreateTrafficPolicyVersionResponse TrafficPolicy

A complex type that contains settings for the new version of the traffic policy.

ctpvrsLocation :: Lens' CreateTrafficPolicyVersionResponse Text

A unique URL that represents a new traffic policy version.