Network.AWS.Route53.CreateVPCAssociationAuthorization

Description

Authorizes the AWS account that created a specified VPC to submit an AssociateVPCWithHostedZone request to associate the VPC with a specified hosted zone that was created by a different account. To submit a CreateVPCAssociationAuthorization request, you must use the account that created the hosted zone. After you authorize the association, use the account that created the VPC to submit an AssociateVPCWithHostedZone request.

Creating a Request

createVPCAssociationAuthorization

Arguments

:: ResourceId
cvaaHostedZoneId

-> VPC
cvaaVPC

-> CreateVPCAssociationAuthorization

Creates a value of CreateVPCAssociationAuthorization with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvaaHostedZoneId - The ID of the private hosted zone that you want to authorize associating a VPC with.

cvaaVPC - A complex type that contains the VPC ID and region for the VPC that you want to authorize associating with your hosted zone.

data CreateVPCAssociationAuthorization

A complex type that contains information about the request to authorize associating a VPC with your private hosted zone. Authorization is only required when a private hosted zone and a VPC were created by using different accounts.

See: createVPCAssociationAuthorization smart constructor.

Request Lenses

cvaaHostedZoneId :: Lens' CreateVPCAssociationAuthorization ResourceId

The ID of the private hosted zone that you want to authorize associating a VPC with.

cvaaVPC :: Lens' CreateVPCAssociationAuthorization VPC

A complex type that contains the VPC ID and region for the VPC that you want to authorize associating with your hosted zone.

Destructuring the Response

createVPCAssociationAuthorizationResponse

Arguments

:: Int
cvaarsResponseStatus

-> ResourceId
cvaarsHostedZoneId

-> VPC
cvaarsVPC

-> CreateVPCAssociationAuthorizationResponse

Creates a value of CreateVPCAssociationAuthorizationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvaarsResponseStatus - -- | The response status code.

cvaarsHostedZoneId - The ID of the hosted zone that you authorized associating a VPC with.

cvaarsVPC - The VPC that you authorized associating with a hosted zone.

data CreateVPCAssociationAuthorizationResponse

A complex type that contains the response information from a CreateVPCAssociationAuthorization request.

See: createVPCAssociationAuthorizationResponse smart constructor.

Response Lenses

cvaarsResponseStatus :: Lens' CreateVPCAssociationAuthorizationResponse Int

- | The response status code.

cvaarsHostedZoneId :: Lens' CreateVPCAssociationAuthorizationResponse ResourceId

The ID of the hosted zone that you authorized associating a VPC with.

cvaarsVPC :: Lens' CreateVPCAssociationAuthorizationResponse VPC

The VPC that you authorized associating with a hosted zone.