Network.AWS.Route53.DeleteHostedZone

Description

This action deletes a hosted zone. To delete a hosted zone, send a DELETE request to the '2013-04-01hostedzonehosted zone ID resource.

For more information about deleting a hosted zone, see Deleting a Hosted Zone in the Amazon Route 53 Developer Guide.

You can delete a hosted zone only if there are no resource record sets other than the default SOA record and NS resource record sets. If your hosted zone contains other resource record sets, you must delete them before you can delete your hosted zone. If you try to delete a hosted zone that contains other resource record sets, Route 53 will deny your request with a HostedZoneNotEmpty error. For information about deleting records from your hosted zone, see ChangeResourceRecordSets.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_DeleteHostedZone.html

Request

data DeleteHostedZone

Request constructor

deleteHostedZone

Arguments

:: Text
dhzId

-> DeleteHostedZone

DeleteHostedZone constructor.

The fields accessible through corresponding lenses are:

dhzId :: Text

Request lenses

dhzId :: Lens' DeleteHostedZone Text

The ID of the hosted zone you want to delete.

Response

data DeleteHostedZoneResponse

Response constructor

deleteHostedZoneResponse

Arguments

:: ChangeInfo
dhzrChangeInfo

-> DeleteHostedZoneResponse

DeleteHostedZoneResponse constructor.

The fields accessible through corresponding lenses are:

dhzrChangeInfo :: ChangeInfo

Response lenses

dhzrChangeInfo :: Lens' DeleteHostedZoneResponse ChangeInfo

A complex type that contains the ID, the status, and the date and time of your delete request.