Network.AWS.Route53.DeleteReusableDelegationSet

Description

This action deletes a reusable delegation set. To delete a reusable delegation set, send a DELETE request to the '2013-04-01delegationsetdelegation set ID resource.

You can delete a reusable delegation set only if there are no associated hosted zones. If your reusable delegation set contains associated hosted zones, you must delete them before you can delete your reusable delegation set. If you try to delete a reusable delegation set that contains associated hosted zones, Route 53 will deny your request with a DelegationSetInUse error.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_DeleteReusableDelegationSet.html

Request

data DeleteReusableDelegationSet

Request constructor

deleteReusableDelegationSet

Arguments

:: Text
drdsId

-> DeleteReusableDelegationSet

DeleteReusableDelegationSet constructor.

The fields accessible through corresponding lenses are:

drdsId :: Text

Request lenses

drdsId :: Lens' DeleteReusableDelegationSet Text

The ID of the reusable delegation set you want to delete.

Response

data DeleteReusableDelegationSetResponse

Response constructor

deleteReusableDelegationSetResponse :: DeleteReusableDelegationSetResponse

DeleteReusableDelegationSetResponse constructor.