Network.AWS.Route53.DeleteTrafficPolicyInstance

Description

Deletes a traffic policy instance and all of the resource record sets that Amazon Route 53 created when you created the instance.

Creating a Request

deleteTrafficPolicyInstance

Arguments

:: Text
dtpiId

-> DeleteTrafficPolicyInstance

Creates a value of DeleteTrafficPolicyInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtpiId - The ID of the traffic policy instance that you want to delete. Important: When you delete a traffic policy instance, Amazon Route 53 also deletes all of the resource record sets that were created when you created the traffic policy instance.

data DeleteTrafficPolicyInstance

A request to delete a specified traffic policy instance.

See: deleteTrafficPolicyInstance smart constructor.

Request Lenses

dtpiId :: Lens' DeleteTrafficPolicyInstance Text

The ID of the traffic policy instance that you want to delete. Important: When you delete a traffic policy instance, Amazon Route 53 also deletes all of the resource record sets that were created when you created the traffic policy instance.

Destructuring the Response

deleteTrafficPolicyInstanceResponse

Arguments

:: Int
dtpirsResponseStatus

-> DeleteTrafficPolicyInstanceResponse

Creates a value of DeleteTrafficPolicyInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtpirsResponseStatus - -- | The response status code.

data DeleteTrafficPolicyInstanceResponse

An empty element.

See: deleteTrafficPolicyInstanceResponse smart constructor.

Response Lenses

dtpirsResponseStatus :: Lens' DeleteTrafficPolicyInstanceResponse Int

- | The response status code.