Network.AWS.Route53.DisassociateVPCFromHostedZone

Description

This action disassociates a VPC from an hosted zone.

To disassociate a VPC to a hosted zone, send a POST request to the '2013-04-01hostedzonehosted zone ID/disassociatevpc resource. The request body must include an XML document with a DisassociateVPCFromHostedZoneRequest element. The response returns the DisassociateVPCFromHostedZoneResponse element that contains ChangeInfo for you to track the progress of the DisassociateVPCFromHostedZoneRequest you made. See GetChange operation for how to track the progress of your change.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_DisassociateVPCFromHostedZone.html

Request

data DisassociateVPCFromHostedZone

Request constructor

disassociateVPCFromHostedZone

Arguments

:: Text
dvpcfhzHostedZoneId

-> VPC
dvpcfhzVPC

-> DisassociateVPCFromHostedZone

DisassociateVPCFromHostedZone constructor.

The fields accessible through corresponding lenses are:

dvpcfhzComment :: Maybe Text

dvpcfhzHostedZoneId :: Text

dvpcfhzVPC :: VPC

Request lenses

dvpcfhzComment :: Lens' DisassociateVPCFromHostedZone (Maybe Text)

Optional: Any comments you want to include about a DisassociateVPCFromHostedZoneRequest.

dvpcfhzHostedZoneId :: Lens' DisassociateVPCFromHostedZone Text

The ID of the hosted zone you want to disassociate your VPC from.

Note that you cannot disassociate the last VPC from a hosted zone.

dvpcfhzVPC :: Lens' DisassociateVPCFromHostedZone VPC

The VPC that you want your hosted zone to be disassociated from.

Response

data DisassociateVPCFromHostedZoneResponse

Response constructor

disassociateVPCFromHostedZoneResponse

Arguments

:: ChangeInfo
dvpcfhzrChangeInfo

-> DisassociateVPCFromHostedZoneResponse

DisassociateVPCFromHostedZoneResponse constructor.

The fields accessible through corresponding lenses are:

dvpcfhzrChangeInfo :: ChangeInfo

Response lenses

dvpcfhzrChangeInfo :: Lens' DisassociateVPCFromHostedZoneResponse ChangeInfo

A complex type that contains the ID, the status, and the date and time of your DisassociateVPCFromHostedZoneRequest.