Network.AWS.Route53.GetChange

Description

This action returns the current status of a change batch request. The status is one of the following values:

PENDING indicates that the changes in this request have not replicated to all Route 53 DNS servers. This is the initial status of all change batch requests.

INSYNC indicates that the changes have replicated to all Amazon Route 53 DNS servers.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_GetChange.html

Request

data GetChange

Request constructor

getChange

Arguments

:: Text
gcId

-> GetChange

GetChange constructor.

The fields accessible through corresponding lenses are:

gcId :: Text

Request lenses

gcId :: Lens' GetChange Text

The ID of the change batch request. The value that you specify here is the value that ChangeResourceRecordSets returned in the Id element when you submitted the request.

Response

data GetChangeResponse

Response constructor

getChangeResponse

Arguments

:: ChangeInfo
gcrChangeInfo

-> GetChangeResponse

GetChangeResponse constructor.

The fields accessible through corresponding lenses are:

gcrChangeInfo :: ChangeInfo

Response lenses

gcrChangeInfo :: Lens' GetChangeResponse ChangeInfo

A complex type that contains information about the specified change batch, including the change batch ID, the status of the change, and the date and time of the request.