Network.AWS.Route53.GetCheckerIpRanges

Description

To retrieve a list of the IP ranges used by Amazon Route 53 health checkers to check the health of your resources, send a GET request to the '2013-04-01/checkeripranges' resource. You can use these IP addresses to configure router and firewall rules to allow health checkers to check the health of your resources.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_GetCheckerIpRanges.html

Request

data GetCheckerIpRanges

Request constructor

getCheckerIpRanges :: GetCheckerIpRanges

GetCheckerIpRanges constructor.

Response

data GetCheckerIpRangesResponse

Response constructor

getCheckerIpRangesResponse :: GetCheckerIpRangesResponse

GetCheckerIpRangesResponse constructor.

The fields accessible through corresponding lenses are:

gcirrCheckerIpRanges :: [Text]

Response lenses

gcirrCheckerIpRanges :: Lens' GetCheckerIpRangesResponse [Text]

A complex type that contains sorted list of IP ranges in CIDR format for Amazon Route 53 health checkers.