Network.AWS.Route53.GetGeoLocation

Description

To retrieve a single geo location, send a GET request to the '2013-04-01/geolocation' resource with one of these options: continentcode | countrycode | countrycode and subdivisioncode.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_GetGeoLocation.html

Request

data GetGeoLocation

Request constructor

getGeoLocation :: GetGeoLocation

GetGeoLocation constructor.

The fields accessible through corresponding lenses are:

gglContinentCode :: Maybe Text

gglCountryCode :: Maybe Text

gglSubdivisionCode :: Maybe Text

Request lenses

gglContinentCode :: Lens' GetGeoLocation (Maybe Text)

The code for a continent geo location. Note: only continent locations have a continent code.

Valid values: AF | AN | AS | EU | OC | NA | SA

Constraint: Specifying ContinentCode with either CountryCode or SubdivisionCode returns an InvalidInput error.

gglCountryCode :: Lens' GetGeoLocation (Maybe Text)

The code for a country geo location. The default location uses * for the country code and will match all locations that are not matched by a geo location.

The default geo location uses a * for the country code. All other country codes follow the ISO 3166 two-character code.

gglSubdivisionCode :: Lens' GetGeoLocation (Maybe Text)

The code for a country's subdivision (e.g., a province of Canada). A subdivision code is only valid with the appropriate country code.

Constraint: Specifying SubdivisionCode without CountryCode returns an InvalidInput error.

Response

data GetGeoLocationResponse

Response constructor

getGeoLocationResponse

Arguments

:: GeoLocationDetails
gglrGeoLocationDetails

-> GetGeoLocationResponse

GetGeoLocationResponse constructor.

The fields accessible through corresponding lenses are:

gglrGeoLocationDetails :: GeoLocationDetails

Response lenses

gglrGeoLocationDetails :: Lens' GetGeoLocationResponse GeoLocationDetails

A complex type that contains the information about the specified geo location.