Network.AWS.Route53.GetHealthCheckCount

Description

To retrieve a count of all your health checks, send a GET request to the '2013-04-01/healthcheckcount' resource.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_GetHealthCheckCount.html

Request

data GetHealthCheckCount

Request constructor

getHealthCheckCount :: GetHealthCheckCount

GetHealthCheckCount constructor.

Response

data GetHealthCheckCountResponse

Response constructor

getHealthCheckCountResponse

Arguments

:: Integer
ghccrHealthCheckCount

-> GetHealthCheckCountResponse

GetHealthCheckCountResponse constructor.

The fields accessible through corresponding lenses are:

ghccrHealthCheckCount :: Integer

Response lenses

ghccrHealthCheckCount :: Lens' GetHealthCheckCountResponse Integer

The number of health checks associated with the current AWS account.