Network.AWS.Route53.GetHealthCheckLastFailureReason

Description

If you want to learn why a health check is currently failing or why it failed most recently (if at all), you can get the failure reason for the most recent failure. Send a GET request to the '2013-04-01healthcheckhealth check ID/lastfailurereason resource.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_GetHealthCheckLastFailureReason.html

Request

data GetHealthCheckLastFailureReason

Request constructor

getHealthCheckLastFailureReason

Arguments

:: Text
ghclfrHealthCheckId

-> GetHealthCheckLastFailureReason

GetHealthCheckLastFailureReason constructor.

The fields accessible through corresponding lenses are:

ghclfrHealthCheckId :: Text

Request lenses

ghclfrHealthCheckId :: Lens' GetHealthCheckLastFailureReason Text

The ID of the health check for which you want to retrieve the reason for the most recent failure.

Response

data GetHealthCheckLastFailureReasonResponse

Response constructor

getHealthCheckLastFailureReasonResponse :: GetHealthCheckLastFailureReasonResponse

GetHealthCheckLastFailureReasonResponse constructor.

The fields accessible through corresponding lenses are:

ghclfrrHealthCheckObservations :: [HealthCheckObservation]

Response lenses

ghclfrrHealthCheckObservations :: Lens' GetHealthCheckLastFailureReasonResponse [HealthCheckObservation]

A list that contains one HealthCheckObservation element for each Route 53 health checker.