Network.AWS.Route53.GetHealthCheckStatus

Description

To retrieve the health check status, send a GET request to the '2013-04-01healthcheckhealth check ID/status resource. You can use this call to get a health check's current status.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_GetHealthCheckStatus.html

Request

data GetHealthCheckStatus

Request constructor

getHealthCheckStatus

Arguments

:: Text
ghcsHealthCheckId

-> GetHealthCheckStatus

GetHealthCheckStatus constructor.

The fields accessible through corresponding lenses are:

ghcsHealthCheckId :: Text

Request lenses

ghcsHealthCheckId :: Lens' GetHealthCheckStatus Text

The ID of the health check for which you want to retrieve the most recent status.

Response

data GetHealthCheckStatusResponse

Response constructor

getHealthCheckStatusResponse :: GetHealthCheckStatusResponse

GetHealthCheckStatusResponse constructor.

The fields accessible through corresponding lenses are:

ghcsrHealthCheckObservations :: [HealthCheckObservation]

Response lenses

ghcsrHealthCheckObservations :: Lens' GetHealthCheckStatusResponse [HealthCheckObservation]

A list that contains one HealthCheckObservation element for each Route 53 health checker.