Network.AWS.Route53.GetHostedZone

Description

To retrieve the delegation set for a hosted zone, send a GET request to the 2013-04-01hostedzonehosted zone ID resource. The delegation set is the four Route 53 name servers that were assigned to the hosted zone when you created it.

Request

data GetHostedZone

The input for a GetHostedZone request.

Request constructor

getHostedZone

Arguments

:: Text
ghzId

-> GetHostedZone

Smart constructor for the minimum required parameters to construct a valid GetHostedZone request.

The fields accessible through corresponding lenses are:

Id :: Text

Request lenses

ghzId :: Lens' GetHostedZone Text

The ID of the hosted zone for which you want to get a list of the name servers in the delegation set.

Response

data GetHostedZoneResponse

A complex type containing information about the specified hosted zone.

Response constructor

getHostedZoneResponse

Arguments

:: HostedZone
ghzrHostedZone

-> DelegationSet
ghzrDelegationSet

-> GetHostedZoneResponse

Smart constructor for the minimum required parameters to construct a valid GetHostedZoneResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

HostedZone :: HostedZone

DelegationSet :: DelegationSet

Response lenses

ghzrHostedZone :: Lens' GetHostedZoneResponse HostedZone

A complex type that contains the information about the specified hosted zone.

ghzrDelegationSet :: Lens' GetHostedZoneResponse DelegationSet

A complex type that contains information about the name servers for the specified hosted zone.