Network.AWS.Route53.GetReusableDelegationSet

Description

To retrieve the reusable delegation set, send a GET request to the '2013-04-01delegationsetdelegation set ID resource.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_GetReusableDelegationSet.html

Request

data GetReusableDelegationSet

Request constructor

getReusableDelegationSet

Arguments

:: Text
grdsId

-> GetReusableDelegationSet

GetReusableDelegationSet constructor.

The fields accessible through corresponding lenses are:

grdsId :: Text

Request lenses

grdsId :: Lens' GetReusableDelegationSet Text

The ID of the reusable delegation set for which you want to get a list of the name server.

Response

data GetReusableDelegationSetResponse

Response constructor

getReusableDelegationSetResponse

Arguments

:: DelegationSet
grdsrDelegationSet

-> GetReusableDelegationSetResponse

GetReusableDelegationSetResponse constructor.

The fields accessible through corresponding lenses are:

grdsrDelegationSet :: DelegationSet

Response lenses

grdsrDelegationSet :: Lens' GetReusableDelegationSetResponse DelegationSet

A complex type that contains the information about the nameservers for the specified delegation set ID.