Network.AWS.Route53.GetTrafficPolicyInstance

Description

Gets information about a specified traffic policy instance.

Creating a Request

getTrafficPolicyInstance

Arguments

:: Text
gtpiId

-> GetTrafficPolicyInstance

Creates a value of GetTrafficPolicyInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtpiId - The ID of the traffic policy instance that you want to get information about.

data GetTrafficPolicyInstance

Gets information about a specified traffic policy instance.

See: getTrafficPolicyInstance smart constructor.

Request Lenses

gtpiId :: Lens' GetTrafficPolicyInstance Text

The ID of the traffic policy instance that you want to get information about.

Destructuring the Response

getTrafficPolicyInstanceResponse

Arguments

:: Int
gtpirsResponseStatus

-> TrafficPolicyInstance
gtpirsTrafficPolicyInstance

-> GetTrafficPolicyInstanceResponse

Creates a value of GetTrafficPolicyInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtpirsResponseStatus - -- | The response status code.

gtpirsTrafficPolicyInstance - A complex type that contains settings for the traffic policy instance.

data GetTrafficPolicyInstanceResponse

A complex type that contains information about the resource record sets that Amazon Route 53 created based on a specified traffic policy.

See: getTrafficPolicyInstanceResponse smart constructor.

Response Lenses

gtpirsResponseStatus :: Lens' GetTrafficPolicyInstanceResponse Int

- | The response status code.

gtpirsTrafficPolicyInstance :: Lens' GetTrafficPolicyInstanceResponse TrafficPolicyInstance

A complex type that contains settings for the traffic policy instance.