Network.AWS.Route53.ListGeoLocations

Description

To retrieve a list of supported geo locations, send a GET request to the '2013-04-01/geolocations' resource. The response to this request includes a GeoLocationDetailsList element with zero, one, or multiple GeoLocationDetails child elements. The list is sorted by country code, and then subdivision code, followed by continents at the end of the list.

By default, the list of geo locations is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. If the list is truncated, IsTruncated will be set to true and a combination of 'NextContinentCode, NextCountryCode, NextSubdivisionCode' will be populated. You can pass these as parameters to 'StartContinentCode,StartCountryCode, StartSubdivisionCode' to control the geo location that the list begins with.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_ListGeoLocations.html

Request

data ListGeoLocations

Request constructor

listGeoLocations :: ListGeoLocations

ListGeoLocations constructor.

The fields accessible through corresponding lenses are:

lglMaxItems :: Maybe Text

lglStartContinentCode :: Maybe Text

lglStartCountryCode :: Maybe Text

lglStartSubdivisionCode :: Maybe Text

Request lenses

lglMaxItems :: Lens' ListGeoLocations (Maybe Text)

The maximum number of geo locations you want in the response body.

lglStartContinentCode :: Lens' ListGeoLocations (Maybe Text)

The first continent code in the lexicographic ordering of geo locations that you want the ListGeoLocations request to list. For non-continent geo locations, this should be null.

Valid values: AF | AN | AS | EU | OC | NA | SA

Constraint: Specifying ContinentCode with either CountryCode or SubdivisionCode returns an InvalidInput error.

lglStartCountryCode :: Lens' ListGeoLocations (Maybe Text)

The first country code in the lexicographic ordering of geo locations that you want the ListGeoLocations request to list.

The default geo location uses a * for the country code. All other country codes follow the ISO 3166 two-character code.

lglStartSubdivisionCode :: Lens' ListGeoLocations (Maybe Text)

The first subdivision code in the lexicographic ordering of geo locations that you want the ListGeoLocations request to list.

Constraint: Specifying SubdivisionCode without CountryCode returns an InvalidInput error.

Response

data ListGeoLocationsResponse

Response constructor

listGeoLocationsResponse

Arguments

:: Bool
lglrIsTruncated

-> Text
lglrMaxItems

-> ListGeoLocationsResponse

ListGeoLocationsResponse constructor.

The fields accessible through corresponding lenses are:

lglrGeoLocationDetailsList :: [GeoLocationDetails]

lglrIsTruncated :: Bool

lglrMaxItems :: Text

lglrNextContinentCode :: Maybe Text

lglrNextCountryCode :: Maybe Text

lglrNextSubdivisionCode :: Maybe Text

Response lenses

lglrGeoLocationDetailsList :: Lens' ListGeoLocationsResponse [GeoLocationDetails]

A complex type that contains information about the geo locations that are returned by the request.

lglrIsTruncated :: Lens' ListGeoLocationsResponse Bool

A flag that indicates whether there are more geo locations to be listed. If your results were truncated, you can make a follow-up request for the next page of results by using the values included in the 'ListGeoLocationsResponse$NextContinentCode', 'ListGeoLocationsResponse$NextCountryCode' and 'ListGeoLocationsResponse$NextSubdivisionCode' elements.

Valid Values: true | false

lglrMaxItems :: Lens' ListGeoLocationsResponse Text

The maximum number of records you requested. The maximum value of MaxItems is 100.

lglrNextContinentCode :: Lens' ListGeoLocationsResponse (Maybe Text)

If the results were truncated, the continent code of the next geo location in the list. This element is present only if 'ListGeoLocationsResponse$IsTruncated' is true and the next geo location to list is a continent location.

lglrNextCountryCode :: Lens' ListGeoLocationsResponse (Maybe Text)

If the results were truncated, the country code of the next geo location in the list. This element is present only if 'ListGeoLocationsResponse$IsTruncated' is true and the next geo location to list is not a continent location.

lglrNextSubdivisionCode :: Lens' ListGeoLocationsResponse (Maybe Text)

If the results were truncated, the subdivision code of the next geo location in the list. This element is present only if 'ListGeoLocationsResponse$IsTruncated' is true and the next geo location has a subdivision.