Network.AWS.Route53.ListHealthChecks

Description

To retrieve a list of your health checks, send a GET request to the '2013-04-01/healthcheck' resource. The response to this request includes a HealthChecks element with zero, one, or multiple HealthCheck child elements. By default, the list of health checks is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. You can use the Marker parameter to control the health check that the list begins with.

Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Amazon Route 53 returns only the first 100.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_ListHealthChecks.html

Request

data ListHealthChecks

Request constructor

listHealthChecks :: ListHealthChecks

ListHealthChecks constructor.

The fields accessible through corresponding lenses are:

lhcMarker :: Maybe Text

lhcMaxItems :: Maybe Text

Request lenses

lhcMarker :: Lens' ListHealthChecks (Maybe Text)

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

lhcMaxItems :: Lens' ListHealthChecks (Maybe Text)

Specify the maximum number of health checks to return per page of results.

Response

data ListHealthChecksResponse

Response constructor

listHealthChecksResponse

Arguments

:: Text
lhcrMarker

-> Bool
lhcrIsTruncated

-> Text
lhcrMaxItems

-> ListHealthChecksResponse

ListHealthChecksResponse constructor.

The fields accessible through corresponding lenses are:

lhcrHealthChecks :: [HealthCheck]

lhcrIsTruncated :: Bool

lhcrMarker :: Text

lhcrMaxItems :: Text

lhcrNextMarker :: Maybe Text

Response lenses

lhcrHealthChecks :: Lens' ListHealthChecksResponse [HealthCheck]

A complex type that contains information about the health checks associated with the current AWS account.

lhcrIsTruncated :: Lens' ListHealthChecksResponse Bool

A flag indicating whether there are more health checks to be listed. If your results were truncated, you can make a follow-up request for the next page of results by using the Marker element.

Valid Values: true | false

lhcrMarker :: Lens' ListHealthChecksResponse Text

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

lhcrMaxItems :: Lens' ListHealthChecksResponse Text

The maximum number of health checks to be included in the response body. If the number of health checks associated with this AWS account exceeds MaxItems, the value of 'ListHealthChecksResponse$IsTruncated' in the response is true. Call ListHealthChecks again and specify the value of 'ListHealthChecksResponse$NextMarker' in the 'ListHostedZonesRequest$Marker' element to get the next page of results.

lhcrNextMarker :: Lens' ListHealthChecksResponse (Maybe Text)

Indicates where to continue listing health checks. If 'ListHealthChecksResponse$IsTruncated' is true, make another request to ListHealthChecks and include the value of the NextMarker element in the Marker element to get the next page of results.