Network.AWS.Route53.ListHostedZones

Description

To retrieve a list of your hosted zones, send a GET request to the '2013-04-01/hostedzone' resource. The response to this request includes a HostedZones element with zero, one, or multiple HostedZone child elements. By default, the list of hosted zones is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. You can use the Marker parameter to control the hosted zone that the list begins with.

Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Amazon Route 53 returns only the first 100.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_ListHostedZones.html

Request

data ListHostedZones

Request constructor

listHostedZones :: ListHostedZones

ListHostedZones constructor.

The fields accessible through corresponding lenses are:

lhzDelegationSetId :: Maybe Text

lhzMarker :: Maybe Text

lhzMaxItems :: Maybe Text

Request lenses

lhzDelegationSetId :: Lens' ListHostedZones (Maybe Text)

lhzMarker :: Lens' ListHostedZones (Maybe Text)

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

lhzMaxItems :: Lens' ListHostedZones (Maybe Text)

Specify the maximum number of hosted zones to return per page of results.

Response

data ListHostedZonesResponse

Response constructor

listHostedZonesResponse

Arguments

:: Bool
lhzrIsTruncated

-> Text
lhzrMaxItems

-> ListHostedZonesResponse

ListHostedZonesResponse constructor.

The fields accessible through corresponding lenses are:

lhzrHostedZones :: [HostedZone]

lhzrIsTruncated :: Bool

lhzrMarker :: Maybe Text

lhzrMaxItems :: Text

lhzrNextMarker :: Maybe Text

Response lenses

lhzrHostedZones :: Lens' ListHostedZonesResponse [HostedZone]

A complex type that contains information about the hosted zones associated with the current AWS account.

lhzrIsTruncated :: Lens' ListHostedZonesResponse Bool

A flag indicating whether there are more hosted zones to be listed. If your results were truncated, you can make a follow-up request for the next page of results by using the Marker element.

Valid Values: true | false

lhzrMarker :: Lens' ListHostedZonesResponse (Maybe Text)

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

lhzrMaxItems :: Lens' ListHostedZonesResponse Text

The maximum number of hosted zones to be included in the response body. If the number of hosted zones associated with this AWS account exceeds MaxItems, the value of 'ListHostedZonesResponse$IsTruncated' in the response is true. Call ListHostedZones again and specify the value of 'ListHostedZonesResponse$NextMarker' in the 'ListHostedZonesRequest$Marker' element to get the next page of results.

lhzrNextMarker :: Lens' ListHostedZonesResponse (Maybe Text)

Indicates where to continue listing hosted zones. If 'ListHostedZonesResponse$IsTruncated' is true, make another request to ListHostedZones and include the value of the NextMarker element in the Marker element to get the next page of results.