Network.AWS.Route53.ListHostedZonesByName

Description

To retrieve a list of your hosted zones in lexicographic order, send a GET request to the '2013-04-01/hostedzonesbyname' resource. The response to this request includes a HostedZones element with zero or more HostedZone child elements lexicographically ordered by DNS name. By default, the list of hosted zones is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. You can use the DNSName and HostedZoneId parameters to control the hosted zone that the list begins with.

Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Amazon Route 53 returns only the first 100.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_ListHostedZonesByName.html

Request

data ListHostedZonesByName

Request constructor

listHostedZonesByName :: ListHostedZonesByName

ListHostedZonesByName constructor.

The fields accessible through corresponding lenses are:

lhzbnDNSName :: Maybe Text

lhzbnHostedZoneId :: Maybe Text

lhzbnMaxItems :: Maybe Text

Request lenses

lhzbnDNSName :: Lens' ListHostedZonesByName (Maybe Text)

The first name in the lexicographic ordering of domain names that you want the ListHostedZonesByNameRequest request to list.

If the request returned more than one page of results, submit another request and specify the value of NextDNSName and NextHostedZoneId from the last response in the DNSName and HostedZoneId parameters to get the next page of results.

lhzbnHostedZoneId :: Lens' ListHostedZonesByName (Maybe Text)

If the request returned more than one page of results, submit another request and specify the value of NextDNSName and NextHostedZoneId from the last response in the DNSName and HostedZoneId parameters to get the next page of results.

lhzbnMaxItems :: Lens' ListHostedZonesByName (Maybe Text)

Specify the maximum number of hosted zones to return per page of results.

Response

data ListHostedZonesByNameResponse

Response constructor

listHostedZonesByNameResponse

Arguments

:: Bool
lhzbnrIsTruncated

-> Text
lhzbnrMaxItems

-> ListHostedZonesByNameResponse

ListHostedZonesByNameResponse constructor.

The fields accessible through corresponding lenses are:

lhzbnrDNSName :: Maybe Text

lhzbnrHostedZoneId :: Maybe Text

lhzbnrHostedZones :: [HostedZone]

lhzbnrIsTruncated :: Bool

lhzbnrMaxItems :: Text

lhzbnrNextDNSName :: Maybe Text

lhzbnrNextHostedZoneId :: Maybe Text

Response lenses

lhzbnrDNSName :: Lens' ListHostedZonesByNameResponse (Maybe Text)

The DNSName value sent in the request.

lhzbnrHostedZoneId :: Lens' ListHostedZonesByNameResponse (Maybe Text)

The HostedZoneId value sent in the request.

lhzbnrHostedZones :: Lens' ListHostedZonesByNameResponse [HostedZone]

A complex type that contains information about the hosted zones associated with the current AWS account.

lhzbnrIsTruncated :: Lens' ListHostedZonesByNameResponse Bool

A flag indicating whether there are more hosted zones to be listed. If your results were truncated, you can make a follow-up request for the next page of results by using the NextDNSName and NextHostedZoneId elements.

Valid Values: true | false

lhzbnrMaxItems :: Lens' ListHostedZonesByNameResponse Text

The maximum number of hosted zones to be included in the response body. If the number of hosted zones associated with this AWS account exceeds MaxItems, the value of 'ListHostedZonesByNameResponse$IsTruncated' in the response is true . Call ListHostedZonesByName again and specify the value of 'ListHostedZonesByNameResponse$NextDNSName' and 'ListHostedZonesByNameResponse$NextHostedZoneId' elements respectively to get the next page of results.

lhzbnrNextDNSName :: Lens' ListHostedZonesByNameResponse (Maybe Text)

If 'ListHostedZonesByNameResponse$IsTruncated' is true, there are more hosted zones associated with the current AWS account. To get the next page of results, make another request to ListHostedZonesByName. Specify the value of 'ListHostedZonesByNameResponse$NextDNSName' in the 'ListHostedZonesByNameRequest$DNSName' element and 'ListHostedZonesByNameResponse$NextHostedZoneId' in the 'ListHostedZonesByNameRequest$HostedZoneId' element.

lhzbnrNextHostedZoneId :: Lens' ListHostedZonesByNameResponse (Maybe Text)

If 'ListHostedZonesByNameResponse$IsTruncated' is true, there are more hosted zones associated with the current AWS account. To get the next page of results, make another request to ListHostedZonesByName. Specify the value of 'ListHostedZonesByNameResponse$NextDNSName' in the 'ListHostedZonesByNameRequest$DNSName' element and 'ListHostedZonesByNameResponse$NextHostedZoneId' in the 'ListHostedZonesByNameRequest$HostedZoneId' element.