Network.AWS.Route53.ListReusableDelegationSets

Description

To retrieve a list of your reusable delegation sets, send a GET request to the 2013-04-01/delegationset resource. The response to this request includes a DelegationSets element with zero, one, or multiple DelegationSet child elements. By default, the list of delegation sets is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. You can use the Marker parameter to control the delegation set that the list begins with. Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Amazon Route 53 returns only the first 100.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_ListReusableDelegationSets.html

Request

data ListReusableDelegationSets

Request constructor

listReusableDelegationSets :: ListReusableDelegationSets

ListReusableDelegationSets constructor.

The fields accessible through corresponding lenses are:

lrdsMarker :: Maybe Text

lrdsMaxItems :: Maybe Text

Request lenses

lrdsMarker :: Lens' ListReusableDelegationSets (Maybe Text)

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

lrdsMaxItems :: Lens' ListReusableDelegationSets (Maybe Text)

Specify the maximum number of reusable delegation sets to return per page of results.

Response

data ListReusableDelegationSetsResponse

Response constructor

listReusableDelegationSetsResponse

Arguments

:: Text
lrdsrMarker

-> Bool
lrdsrIsTruncated

-> Text
lrdsrMaxItems

-> ListReusableDelegationSetsResponse

ListReusableDelegationSetsResponse constructor.

The fields accessible through corresponding lenses are:

lrdsrDelegationSets :: [DelegationSet]

lrdsrIsTruncated :: Bool

lrdsrMarker :: Text

lrdsrMaxItems :: Text

lrdsrNextMarker :: Maybe Text

Response lenses

lrdsrDelegationSets :: Lens' ListReusableDelegationSetsResponse [DelegationSet]

A complex type that contains information about the reusable delegation sets associated with the current AWS account.

lrdsrIsTruncated :: Lens' ListReusableDelegationSetsResponse Bool

A flag indicating whether there are more reusable delegation sets to be listed. If your results were truncated, you can make a follow-up request for the next page of results by using the Marker element. Valid Values: true | false.

lrdsrMarker :: Lens' ListReusableDelegationSetsResponse Text

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

lrdsrMaxItems :: Lens' ListReusableDelegationSetsResponse Text

The maximum number of reusable delegation sets to be included in the response body. If the number of reusable delegation sets associated with this AWS account exceeds MaxItems, the value of ListReusablDelegationSetsResponse$IsTruncated in the response is true. Call ListReusableDelegationSets again and specify the value of ListReusableDelegationSetsResponse$NextMarker in the ListReusableDelegationSetsRequest$Marker element to get the next page of results.

lrdsrNextMarker :: Lens' ListReusableDelegationSetsResponse (Maybe Text)

Indicates where to continue listing reusable delegation sets. If ListReusableDelegationSetsResponse$IsTruncated is true, make another request to ListReusableDelegationSets and include the value of the NextMarker element in the Marker element to get the next page of results.