Network.AWS.Route53.ListTrafficPolicies

Description

Gets information about the latest version for every traffic policy that is associated with the current AWS account. Policies are listed in the order in which they were created.

Creating a Request

listTrafficPolicies :: ListTrafficPolicies

Creates a value of ListTrafficPolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltpTrafficPolicyIdMarker - (Conditional) For your first request to ListTrafficPolicies , don't include the TrafficPolicyIdMarker parameter. If you have more traffic policies than the value of MaxItems , ListTrafficPolicies returns only the first MaxItems traffic policies. To get the next group of policies, submit another request to ListTrafficPolicies . For the value of TrafficPolicyIdMarker , specify the value of TrafficPolicyIdMarker that was returned in the previous response.

ltpMaxItems - (Optional) The maximum number of traffic policies that you want Amazon Route 53 to return in response to this request. If you have more than MaxItems traffic policies, the value of IsTruncated in the response is true , and the value of TrafficPolicyIdMarker is the ID of the first traffic policy that Amazon Route 53 will return if you submit another request.

data ListTrafficPolicies

A complex type that contains the information about the request to list the traffic policies that are associated with the current AWS account.

See: listTrafficPolicies smart constructor.

Request Lenses

ltpTrafficPolicyIdMarker :: Lens' ListTrafficPolicies (Maybe Text)

(Conditional) For your first request to ListTrafficPolicies , don't include the TrafficPolicyIdMarker parameter. If you have more traffic policies than the value of MaxItems , ListTrafficPolicies returns only the first MaxItems traffic policies. To get the next group of policies, submit another request to ListTrafficPolicies . For the value of TrafficPolicyIdMarker , specify the value of TrafficPolicyIdMarker that was returned in the previous response.

ltpMaxItems :: Lens' ListTrafficPolicies (Maybe Text)

(Optional) The maximum number of traffic policies that you want Amazon Route 53 to return in response to this request. If you have more than MaxItems traffic policies, the value of IsTruncated in the response is true , and the value of TrafficPolicyIdMarker is the ID of the first traffic policy that Amazon Route 53 will return if you submit another request.

Destructuring the Response

listTrafficPoliciesResponse

Arguments

:: Int
ltprsResponseStatus

-> Bool
ltprsIsTruncated

-> Text
ltprsTrafficPolicyIdMarker

-> Text
ltprsMaxItems

-> ListTrafficPoliciesResponse

Creates a value of ListTrafficPoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltprsResponseStatus - -- | The response status code.

ltprsTrafficPolicySummaries - A list that contains one TrafficPolicySummary element for each traffic policy that was created by the current AWS account.

ltprsIsTruncated - A flag that indicates whether there are more traffic policies to be listed. If the response was truncated, you can get the next group of traffic policies by submitting another ListTrafficPolicies request and specifying the value of TrafficPolicyIdMarker in the TrafficPolicyIdMarker request parameter.

ltprsTrafficPolicyIdMarker - If the value of IsTruncated is true , TrafficPolicyIdMarker is the ID of the first traffic policy in the next group of MaxItems traffic policies.

ltprsMaxItems - The value that you specified for the MaxItems parameter in the ListTrafficPolicies request that produced the current response.

data ListTrafficPoliciesResponse

A complex type that contains the response information for the request.

See: listTrafficPoliciesResponse smart constructor.

Response Lenses

ltprsResponseStatus :: Lens' ListTrafficPoliciesResponse Int

- | The response status code.

ltprsTrafficPolicySummaries :: Lens' ListTrafficPoliciesResponse [TrafficPolicySummary]

A list that contains one TrafficPolicySummary element for each traffic policy that was created by the current AWS account.

ltprsIsTruncated :: Lens' ListTrafficPoliciesResponse Bool

A flag that indicates whether there are more traffic policies to be listed. If the response was truncated, you can get the next group of traffic policies by submitting another ListTrafficPolicies request and specifying the value of TrafficPolicyIdMarker in the TrafficPolicyIdMarker request parameter.

ltprsTrafficPolicyIdMarker :: Lens' ListTrafficPoliciesResponse Text

If the value of IsTruncated is true , TrafficPolicyIdMarker is the ID of the first traffic policy in the next group of MaxItems traffic policies.

ltprsMaxItems :: Lens' ListTrafficPoliciesResponse Text

The value that you specified for the MaxItems parameter in the ListTrafficPolicies request that produced the current response.