Network.AWS.Route53.ListTrafficPolicyInstances

Description

Gets information about the traffic policy instances that you created by using the current AWS account.

Amazon Route 53 returns a maximum of 100 items in each response. If you have a lot of traffic policy instances, you can use the MaxItems parameter to list them in groups of up to 100.

Creating a Request

listTrafficPolicyInstances :: ListTrafficPolicyInstances

Creates a value of ListTrafficPolicyInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltpiTrafficPolicyInstanceTypeMarker - If the value of IsTruncated in the previous response was true , you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancetype , specify the value of TrafficPolicyInstanceTypeMarker from the previous response, which is the type of the first traffic policy instance in the next group of traffic policy instances. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpiMaxItems - The maximum number of traffic policy instances that you want Amazon Route 53 to return in response to a ListTrafficPolicyInstances request. If you have more than MaxItems traffic policy instances, the value of the IsTruncated element in the response is true , and the values of HostedZoneIdMarker , TrafficPolicyInstanceNameMarker , and TrafficPolicyInstanceTypeMarker represent the first traffic policy instance in the next group of MaxItems traffic policy instances.

ltpiHostedZoneIdMarker - If the value of IsTruncated in the previous response was true , you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of HostedZoneId , specify the value of HostedZoneIdMarker from the previous response, which is the hosted zone ID of the first traffic policy instance in the next group of traffic policy instances. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpiTrafficPolicyInstanceNameMarker - If the value of IsTruncated in the previous response was true , you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancename , specify the value of TrafficPolicyInstanceNameMarker from the previous response, which is the name of the first traffic policy instance in the next group of traffic policy instances. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

data ListTrafficPolicyInstances

A request to get information about the traffic policy instances that you created by using the current AWS account.

See: listTrafficPolicyInstances smart constructor.

Request Lenses

ltpiTrafficPolicyInstanceTypeMarker :: Lens' ListTrafficPolicyInstances (Maybe RecordType)

If the value of IsTruncated in the previous response was true , you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancetype , specify the value of TrafficPolicyInstanceTypeMarker from the previous response, which is the type of the first traffic policy instance in the next group of traffic policy instances. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpiMaxItems :: Lens' ListTrafficPolicyInstances (Maybe Text)

The maximum number of traffic policy instances that you want Amazon Route 53 to return in response to a ListTrafficPolicyInstances request. If you have more than MaxItems traffic policy instances, the value of the IsTruncated element in the response is true , and the values of HostedZoneIdMarker , TrafficPolicyInstanceNameMarker , and TrafficPolicyInstanceTypeMarker represent the first traffic policy instance in the next group of MaxItems traffic policy instances.

ltpiHostedZoneIdMarker :: Lens' ListTrafficPolicyInstances (Maybe ResourceId)

If the value of IsTruncated in the previous response was true , you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of HostedZoneId , specify the value of HostedZoneIdMarker from the previous response, which is the hosted zone ID of the first traffic policy instance in the next group of traffic policy instances. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpiTrafficPolicyInstanceNameMarker :: Lens' ListTrafficPolicyInstances (Maybe Text)

If the value of IsTruncated in the previous response was true , you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancename , specify the value of TrafficPolicyInstanceNameMarker from the previous response, which is the name of the first traffic policy instance in the next group of traffic policy instances. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

Destructuring the Response

listTrafficPolicyInstancesResponse

Arguments

:: Int
ltpirsResponseStatus

-> Bool
ltpirsIsTruncated

-> Text
ltpirsMaxItems

-> ListTrafficPolicyInstancesResponse

Creates a value of ListTrafficPolicyInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltpirsTrafficPolicyInstanceTypeMarker - If IsTruncated is true , TrafficPolicyInstanceTypeMarker is the DNS type of the resource record sets that are associated with the first traffic policy instance that Amazon Route 53 will return if you submit another ListTrafficPolicyInstances request.

ltpirsHostedZoneIdMarker - If IsTruncated is true , HostedZoneIdMarker is the ID of the hosted zone of the first traffic policy instance that Amazon Route 53 will return if you submit another ListTrafficPolicyInstances request.

ltpirsTrafficPolicyInstanceNameMarker - If IsTruncated is true , TrafficPolicyInstanceNameMarker is the name of the first traffic policy instance that Amazon Route 53 will return if you submit another ListTrafficPolicyInstances request.

ltpirsResponseStatus - -- | The response status code.

ltpirsTrafficPolicyInstances - A list that contains one TrafficPolicyInstance element for each traffic policy instance that matches the elements in the request.

ltpirsIsTruncated - A flag that indicates whether there are more traffic policy instances to be listed. If the response was truncated, you can get more traffic policy instances by calling ListTrafficPolicyInstances again and specifying the values of the HostedZoneIdMarker , TrafficPolicyInstanceNameMarker , and TrafficPolicyInstanceTypeMarker in the corresponding request parameters.

ltpirsMaxItems - The value that you specified for the MaxItems parameter in the call to ListTrafficPolicyInstances that produced the current response.

data ListTrafficPolicyInstancesResponse

A complex type that contains the response information for the request.

See: listTrafficPolicyInstancesResponse smart constructor.

Response Lenses

ltpirsTrafficPolicyInstanceTypeMarker :: Lens' ListTrafficPolicyInstancesResponse (Maybe RecordType)

If IsTruncated is true , TrafficPolicyInstanceTypeMarker is the DNS type of the resource record sets that are associated with the first traffic policy instance that Amazon Route 53 will return if you submit another ListTrafficPolicyInstances request.

ltpirsHostedZoneIdMarker :: Lens' ListTrafficPolicyInstancesResponse (Maybe ResourceId)

If IsTruncated is true , HostedZoneIdMarker is the ID of the hosted zone of the first traffic policy instance that Amazon Route 53 will return if you submit another ListTrafficPolicyInstances request.

ltpirsTrafficPolicyInstanceNameMarker :: Lens' ListTrafficPolicyInstancesResponse (Maybe Text)

If IsTruncated is true , TrafficPolicyInstanceNameMarker is the name of the first traffic policy instance that Amazon Route 53 will return if you submit another ListTrafficPolicyInstances request.

ltpirsResponseStatus :: Lens' ListTrafficPolicyInstancesResponse Int

- | The response status code.

ltpirsTrafficPolicyInstances :: Lens' ListTrafficPolicyInstancesResponse [TrafficPolicyInstance]

A list that contains one TrafficPolicyInstance element for each traffic policy instance that matches the elements in the request.

ltpirsIsTruncated :: Lens' ListTrafficPolicyInstancesResponse Bool

A flag that indicates whether there are more traffic policy instances to be listed. If the response was truncated, you can get more traffic policy instances by calling ListTrafficPolicyInstances again and specifying the values of the HostedZoneIdMarker , TrafficPolicyInstanceNameMarker , and TrafficPolicyInstanceTypeMarker in the corresponding request parameters.

ltpirsMaxItems :: Lens' ListTrafficPolicyInstancesResponse Text

The value that you specified for the MaxItems parameter in the call to ListTrafficPolicyInstances that produced the current response.