Network.AWS.Route53.ListTrafficPolicyInstancesByHostedZone

Description

Gets information about the traffic policy instances that you created in a specified hosted zone.

Amazon Route 53 returns a maximum of 100 items in each response. If you have a lot of traffic policy instances, you can use the MaxItems parameter to list them in groups of up to 100.

Creating a Request

listTrafficPolicyInstancesByHostedZone

Arguments

:: ResourceId
ltpibhzHostedZoneId

-> ListTrafficPolicyInstancesByHostedZone

Creates a value of ListTrafficPolicyInstancesByHostedZone with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltpibhzTrafficPolicyInstanceTypeMarker - If the value of IsTruncated in the previous response is true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancetype , specify the value of TrafficPolicyInstanceTypeMarker from the previous response, which is the type of the first traffic policy instance in the next group of traffic policy instances. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpibhzMaxItems - The maximum number of traffic policy instances to be included in the response body for this request. If you have more than MaxItems traffic policy instances, the value of the IsTruncated element in the response is true , and the values of HostedZoneIdMarker , TrafficPolicyInstanceNameMarker , and TrafficPolicyInstanceTypeMarker represent the first traffic policy instance that Amazon Route 53 will return if you submit another request.

ltpibhzTrafficPolicyInstanceNameMarker - If the value of IsTruncated in the previous response is true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancename , specify the value of TrafficPolicyInstanceNameMarker from the previous response, which is the name of the first traffic policy instance in the next group of traffic policy instances. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpibhzHostedZoneId - The ID of the hosted zone that you want to list traffic policy instances for.

data ListTrafficPolicyInstancesByHostedZone

A request for the traffic policy instances that you created in a specified hosted zone.

See: listTrafficPolicyInstancesByHostedZone smart constructor.

Request Lenses

ltpibhzTrafficPolicyInstanceTypeMarker :: Lens' ListTrafficPolicyInstancesByHostedZone (Maybe RecordType)

If the value of IsTruncated in the previous response is true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancetype , specify the value of TrafficPolicyInstanceTypeMarker from the previous response, which is the type of the first traffic policy instance in the next group of traffic policy instances. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpibhzMaxItems :: Lens' ListTrafficPolicyInstancesByHostedZone (Maybe Text)

The maximum number of traffic policy instances to be included in the response body for this request. If you have more than MaxItems traffic policy instances, the value of the IsTruncated element in the response is true , and the values of HostedZoneIdMarker , TrafficPolicyInstanceNameMarker , and TrafficPolicyInstanceTypeMarker represent the first traffic policy instance that Amazon Route 53 will return if you submit another request.

ltpibhzTrafficPolicyInstanceNameMarker :: Lens' ListTrafficPolicyInstancesByHostedZone (Maybe Text)

If the value of IsTruncated in the previous response is true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancename , specify the value of TrafficPolicyInstanceNameMarker from the previous response, which is the name of the first traffic policy instance in the next group of traffic policy instances. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpibhzHostedZoneId :: Lens' ListTrafficPolicyInstancesByHostedZone ResourceId

The ID of the hosted zone that you want to list traffic policy instances for.

Destructuring the Response

listTrafficPolicyInstancesByHostedZoneResponse

Arguments

:: Int
ltpibhzrsResponseStatus

-> Bool
ltpibhzrsIsTruncated

-> Text
ltpibhzrsMaxItems

-> ListTrafficPolicyInstancesByHostedZoneResponse

Creates a value of ListTrafficPolicyInstancesByHostedZoneResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltpibhzrsTrafficPolicyInstanceTypeMarker - If IsTruncated is true, TrafficPolicyInstanceTypeMarker is the DNS type of the resource record sets that are associated with the first traffic policy instance in the next group of traffic policy instances.

ltpibhzrsTrafficPolicyInstanceNameMarker - If IsTruncated is true , TrafficPolicyInstanceNameMarker is the name of the first traffic policy instance in the next group of traffic policy instances.

ltpibhzrsResponseStatus - -- | The response status code.

ltpibhzrsTrafficPolicyInstances - A list that contains one TrafficPolicyInstance element for each traffic policy instance that matches the elements in the request.

ltpibhzrsIsTruncated - A flag that indicates whether there are more traffic policy instances to be listed. If the response was truncated, you can get the next group of traffic policy instances by submitting another ListTrafficPolicyInstancesByHostedZone request and specifying the values of HostedZoneIdMarker , TrafficPolicyInstanceNameMarker , and TrafficPolicyInstanceTypeMarker in the corresponding request parameters.

ltpibhzrsMaxItems - The value that you specified for the MaxItems parameter in the ListTrafficPolicyInstancesByHostedZone request that produced the current response.

data ListTrafficPolicyInstancesByHostedZoneResponse

A complex type that contains the response information for the request.

See: listTrafficPolicyInstancesByHostedZoneResponse smart constructor.

Response Lenses

ltpibhzrsTrafficPolicyInstanceTypeMarker :: Lens' ListTrafficPolicyInstancesByHostedZoneResponse (Maybe RecordType)

If IsTruncated is true, TrafficPolicyInstanceTypeMarker is the DNS type of the resource record sets that are associated with the first traffic policy instance in the next group of traffic policy instances.

ltpibhzrsTrafficPolicyInstanceNameMarker :: Lens' ListTrafficPolicyInstancesByHostedZoneResponse (Maybe Text)

If IsTruncated is true , TrafficPolicyInstanceNameMarker is the name of the first traffic policy instance in the next group of traffic policy instances.

ltpibhzrsResponseStatus :: Lens' ListTrafficPolicyInstancesByHostedZoneResponse Int

- | The response status code.

ltpibhzrsTrafficPolicyInstances :: Lens' ListTrafficPolicyInstancesByHostedZoneResponse [TrafficPolicyInstance]

A list that contains one TrafficPolicyInstance element for each traffic policy instance that matches the elements in the request.

ltpibhzrsIsTruncated :: Lens' ListTrafficPolicyInstancesByHostedZoneResponse Bool

A flag that indicates whether there are more traffic policy instances to be listed. If the response was truncated, you can get the next group of traffic policy instances by submitting another ListTrafficPolicyInstancesByHostedZone request and specifying the values of HostedZoneIdMarker , TrafficPolicyInstanceNameMarker , and TrafficPolicyInstanceTypeMarker in the corresponding request parameters.

ltpibhzrsMaxItems :: Lens' ListTrafficPolicyInstancesByHostedZoneResponse Text

The value that you specified for the MaxItems parameter in the ListTrafficPolicyInstancesByHostedZone request that produced the current response.