Network.AWS.Route53.ListTrafficPolicyInstancesByPolicy

Description

Gets information about the traffic policy instances that you created by using a specify traffic policy version.

Amazon Route 53 returns a maximum of 100 items in each response. If you have a lot of traffic policy instances, you can use the MaxItems parameter to list them in groups of up to 100.

Creating a Request

listTrafficPolicyInstancesByPolicy

Arguments

:: Text
ltpibpTrafficPolicyId

-> Natural
ltpibpTrafficPolicyVersion

-> ListTrafficPolicyInstancesByPolicy

Creates a value of ListTrafficPolicyInstancesByPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltpibpTrafficPolicyInstanceTypeMarker - If the value of IsTruncated in the previous response was true , you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstancesByPolicy request. For the value of trafficpolicyinstancetype , specify the value of TrafficPolicyInstanceTypeMarker from the previous response, which is the name of the first traffic policy instance that Amazon Route 53 will return if you submit another request. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpibpMaxItems - The maximum number of traffic policy instances to be included in the response body for this request. If you have more than MaxItems traffic policy instances, the value of the IsTruncated element in the response is true , and the values of HostedZoneIdMarker , TrafficPolicyInstanceNameMarker , and TrafficPolicyInstanceTypeMarker represent the first traffic policy instance that Amazon Route 53 will return if you submit another request.

ltpibpHostedZoneIdMarker - If the value of IsTruncated in the previous response was true , you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstancesByPolicy request. For the value of hostedzoneid , specify the value of HostedZoneIdMarker from the previous response, which is the hosted zone ID of the first traffic policy instance that Amazon Route 53 will return if you submit another request. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpibpTrafficPolicyInstanceNameMarker - If the value of IsTruncated in the previous response was true , you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstancesByPolicy request. For the value of trafficpolicyinstancename , specify the value of TrafficPolicyInstanceNameMarker from the previous response, which is the name of the first traffic policy instance that Amazon Route 53 will return if you submit another request. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpibpTrafficPolicyId - The ID of the traffic policy for which you want to list traffic policy instances.

ltpibpTrafficPolicyVersion - The version of the traffic policy for which you want to list traffic policy instances. The version must be associated with the traffic policy that is specified by TrafficPolicyId .

data ListTrafficPolicyInstancesByPolicy

A complex type that contains the information about the request to list your traffic policy instances.

See: listTrafficPolicyInstancesByPolicy smart constructor.

Request Lenses

ltpibpTrafficPolicyInstanceTypeMarker :: Lens' ListTrafficPolicyInstancesByPolicy (Maybe RecordType)

If the value of IsTruncated in the previous response was true , you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstancesByPolicy request. For the value of trafficpolicyinstancetype , specify the value of TrafficPolicyInstanceTypeMarker from the previous response, which is the name of the first traffic policy instance that Amazon Route 53 will return if you submit another request. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpibpMaxItems :: Lens' ListTrafficPolicyInstancesByPolicy (Maybe Text)

The maximum number of traffic policy instances to be included in the response body for this request. If you have more than MaxItems traffic policy instances, the value of the IsTruncated element in the response is true , and the values of HostedZoneIdMarker , TrafficPolicyInstanceNameMarker , and TrafficPolicyInstanceTypeMarker represent the first traffic policy instance that Amazon Route 53 will return if you submit another request.

ltpibpHostedZoneIdMarker :: Lens' ListTrafficPolicyInstancesByPolicy (Maybe ResourceId)

If the value of IsTruncated in the previous response was true , you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstancesByPolicy request. For the value of hostedzoneid , specify the value of HostedZoneIdMarker from the previous response, which is the hosted zone ID of the first traffic policy instance that Amazon Route 53 will return if you submit another request. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpibpTrafficPolicyInstanceNameMarker :: Lens' ListTrafficPolicyInstancesByPolicy (Maybe Text)

If the value of IsTruncated in the previous response was true , you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstancesByPolicy request. For the value of trafficpolicyinstancename , specify the value of TrafficPolicyInstanceNameMarker from the previous response, which is the name of the first traffic policy instance that Amazon Route 53 will return if you submit another request. If the value of IsTruncated in the previous response was false , there are no more traffic policy instances to get.

ltpibpTrafficPolicyId :: Lens' ListTrafficPolicyInstancesByPolicy Text

The ID of the traffic policy for which you want to list traffic policy instances.

ltpibpTrafficPolicyVersion :: Lens' ListTrafficPolicyInstancesByPolicy Natural

The version of the traffic policy for which you want to list traffic policy instances. The version must be associated with the traffic policy that is specified by TrafficPolicyId .

Destructuring the Response

listTrafficPolicyInstancesByPolicyResponse

Arguments

:: Int
ltpibprsResponseStatus

-> Bool
ltpibprsIsTruncated

-> Text
ltpibprsMaxItems

-> ListTrafficPolicyInstancesByPolicyResponse

Creates a value of ListTrafficPolicyInstancesByPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltpibprsTrafficPolicyInstanceTypeMarker - If IsTruncated is true , TrafficPolicyInstanceTypeMarker is the DNS type of the resource record sets that are associated with the first traffic policy instance in the next group of MaxItems traffic policy instances.

ltpibprsHostedZoneIdMarker - If IsTruncated is true , HostedZoneIdMarker is the ID of the hosted zone of the first traffic policy instance in the next group of traffic policy instances.

ltpibprsTrafficPolicyInstanceNameMarker - If IsTruncated is true , TrafficPolicyInstanceNameMarker is the name of the first traffic policy instance in the next group of MaxItems traffic policy instances.

ltpibprsResponseStatus - -- | The response status code.

ltpibprsTrafficPolicyInstances - A list that contains one TrafficPolicyInstance element for each traffic policy instance that matches the elements in the request.

ltpibprsIsTruncated - A flag that indicates whether there are more traffic policy instances to be listed. If the response was truncated, you can get the next group of traffic policy instances by calling ListTrafficPolicyInstancesByPolicy again and specifying the values of the HostedZoneIdMarker , TrafficPolicyInstanceNameMarker , and TrafficPolicyInstanceTypeMarker elements in the corresponding request parameters.

ltpibprsMaxItems - The value that you specified for the MaxItems parameter in the call to ListTrafficPolicyInstancesByPolicy that produced the current response.

data ListTrafficPolicyInstancesByPolicyResponse

A complex type that contains the response information for the request.

See: listTrafficPolicyInstancesByPolicyResponse smart constructor.

Response Lenses

ltpibprsTrafficPolicyInstanceTypeMarker :: Lens' ListTrafficPolicyInstancesByPolicyResponse (Maybe RecordType)

If IsTruncated is true , TrafficPolicyInstanceTypeMarker is the DNS type of the resource record sets that are associated with the first traffic policy instance in the next group of MaxItems traffic policy instances.

ltpibprsHostedZoneIdMarker :: Lens' ListTrafficPolicyInstancesByPolicyResponse (Maybe ResourceId)

If IsTruncated is true , HostedZoneIdMarker is the ID of the hosted zone of the first traffic policy instance in the next group of traffic policy instances.

ltpibprsTrafficPolicyInstanceNameMarker :: Lens' ListTrafficPolicyInstancesByPolicyResponse (Maybe Text)

If IsTruncated is true , TrafficPolicyInstanceNameMarker is the name of the first traffic policy instance in the next group of MaxItems traffic policy instances.

ltpibprsResponseStatus :: Lens' ListTrafficPolicyInstancesByPolicyResponse Int

- | The response status code.

ltpibprsTrafficPolicyInstances :: Lens' ListTrafficPolicyInstancesByPolicyResponse [TrafficPolicyInstance]

A list that contains one TrafficPolicyInstance element for each traffic policy instance that matches the elements in the request.

ltpibprsIsTruncated :: Lens' ListTrafficPolicyInstancesByPolicyResponse Bool

A flag that indicates whether there are more traffic policy instances to be listed. If the response was truncated, you can get the next group of traffic policy instances by calling ListTrafficPolicyInstancesByPolicy again and specifying the values of the HostedZoneIdMarker , TrafficPolicyInstanceNameMarker , and TrafficPolicyInstanceTypeMarker elements in the corresponding request parameters.

ltpibprsMaxItems :: Lens' ListTrafficPolicyInstancesByPolicyResponse Text

The value that you specified for the MaxItems parameter in the call to ListTrafficPolicyInstancesByPolicy that produced the current response.