Network.AWS.Route53.ListTrafficPolicyVersions

Description

Gets information about all of the versions for a specified traffic policy.

Traffic policy versions are listed in numerical order by VersionNumber .

Creating a Request

listTrafficPolicyVersions

Arguments

:: Text
ltpvId

-> ListTrafficPolicyVersions

Creates a value of ListTrafficPolicyVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltpvMaxItems - The maximum number of traffic policy versions that you want Amazon Route 53 to include in the response body for this request. If the specified traffic policy has more than MaxItems versions, the value of IsTruncated in the response is true , and the value of the TrafficPolicyVersionMarker element is the ID of the first version that Amazon Route 53 will return if you submit another request.

ltpvTrafficPolicyVersionMarker - For your first request to ListTrafficPolicyVersions , don't include the TrafficPolicyVersionMarker parameter. If you have more traffic policy versions than the value of MaxItems , ListTrafficPolicyVersions returns only the first group of MaxItems versions. To get more traffic policy versions, submit another ListTrafficPolicyVersions request. For the value of TrafficPolicyVersionMarker , specify the value of TrafficPolicyVersionMarker in the previous response.

ltpvId - Specify the value of Id of the traffic policy for which you want to list all versions.

data ListTrafficPolicyVersions

A complex type that contains the information about the request to list your traffic policies.

See: listTrafficPolicyVersions smart constructor.

Request Lenses

ltpvMaxItems :: Lens' ListTrafficPolicyVersions (Maybe Text)

The maximum number of traffic policy versions that you want Amazon Route 53 to include in the response body for this request. If the specified traffic policy has more than MaxItems versions, the value of IsTruncated in the response is true , and the value of the TrafficPolicyVersionMarker element is the ID of the first version that Amazon Route 53 will return if you submit another request.

ltpvTrafficPolicyVersionMarker :: Lens' ListTrafficPolicyVersions (Maybe Text)

For your first request to ListTrafficPolicyVersions , don't include the TrafficPolicyVersionMarker parameter. If you have more traffic policy versions than the value of MaxItems , ListTrafficPolicyVersions returns only the first group of MaxItems versions. To get more traffic policy versions, submit another ListTrafficPolicyVersions request. For the value of TrafficPolicyVersionMarker , specify the value of TrafficPolicyVersionMarker in the previous response.

ltpvId :: Lens' ListTrafficPolicyVersions Text

Specify the value of Id of the traffic policy for which you want to list all versions.

Destructuring the Response

listTrafficPolicyVersionsResponse

Arguments

:: Int
ltpvrsResponseStatus

-> Bool
ltpvrsIsTruncated

-> Text
ltpvrsTrafficPolicyVersionMarker

-> Text
ltpvrsMaxItems

-> ListTrafficPolicyVersionsResponse

Creates a value of ListTrafficPolicyVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltpvrsResponseStatus - -- | The response status code.

ltpvrsTrafficPolicies - A list that contains one TrafficPolicy element for each traffic policy version that is associated with the specified traffic policy.

ltpvrsIsTruncated - A flag that indicates whether there are more traffic policies to be listed. If the response was truncated, you can get the next group of traffic policies by submitting another ListTrafficPolicyVersions request and specifying the value of NextMarker in the marker parameter.

ltpvrsTrafficPolicyVersionMarker - If IsTruncated is true , the value of TrafficPolicyVersionMarker identifies the first traffic policy that Amazon Route 53 will return if you submit another request. Call ListTrafficPolicyVersions again and specify the value of TrafficPolicyVersionMarker in the TrafficPolicyVersionMarker request parameter. This element is present only if IsTruncated is true .

ltpvrsMaxItems - The value that you specified for the maxitems parameter in the ListTrafficPolicyVersions request that produced the current response.

data ListTrafficPolicyVersionsResponse

A complex type that contains the response information for the request.

See: listTrafficPolicyVersionsResponse smart constructor.

Response Lenses

ltpvrsResponseStatus :: Lens' ListTrafficPolicyVersionsResponse Int

- | The response status code.

ltpvrsTrafficPolicies :: Lens' ListTrafficPolicyVersionsResponse [TrafficPolicy]

A list that contains one TrafficPolicy element for each traffic policy version that is associated with the specified traffic policy.

ltpvrsIsTruncated :: Lens' ListTrafficPolicyVersionsResponse Bool

A flag that indicates whether there are more traffic policies to be listed. If the response was truncated, you can get the next group of traffic policies by submitting another ListTrafficPolicyVersions request and specifying the value of NextMarker in the marker parameter.

ltpvrsTrafficPolicyVersionMarker :: Lens' ListTrafficPolicyVersionsResponse Text

If IsTruncated is true , the value of TrafficPolicyVersionMarker identifies the first traffic policy that Amazon Route 53 will return if you submit another request. Call ListTrafficPolicyVersions again and specify the value of TrafficPolicyVersionMarker in the TrafficPolicyVersionMarker request parameter. This element is present only if IsTruncated is true .

ltpvrsMaxItems :: Lens' ListTrafficPolicyVersionsResponse Text

The value that you specified for the maxitems parameter in the ListTrafficPolicyVersions request that produced the current response.