Network.AWS.Route53.Monadic

Description

Amazon Route 53 is a scalable Domain Name System (DNS) web service. It provides secure and reliable routing to your infrastructure that uses Amazon Web Services (AWS) products, such as Amazon Elastic Compute Cloud (Amazon EC2), Elastic Load Balancing, or Amazon Simple Storage Service (Amazon S3). You can also use Amazon Route 53 to route users to your infrastructure outside of AWS.

This module is provided for convenience. It offers an alternative to the common idiom of supplying required fields to an operations's smart constructor, using the operation's lenses to modify additional fields, and then sending or paginating the request.

As an example: using Network.AWS.Route53 with the smart constructor and basic lens syntax, before explicitly calling send:

import Control.Monad.Trans.AWS
import Network.AWS.Route53

send $ (mkOperationName w x)
     & onLensField1 .~ y
     & onLensField2 .~ z

Versus using Network.AWS.Route53.Monadic with the State operator variants from Control.Lens.Setter such as .= to modify any additional request parameters before sending:

import Network.AWS.Route53.Monadic

operationName w x $ do
    onLensField1 .= y
    onLensField2 .= z

ChangeResourceRecordSets

Use this action to create or change your authoritative DNS information. To use this action, send a POST request to the 2013-04-01hostedzonehosted Zone ID/rrset resource. The request body must include an XML document with a ChangeResourceRecordSetsRequest element. Changes are a list of change items and are considered transactional. For more information on transactional changes, also known as change batches, see Creating, Changing, and Deleting Resource Record Sets Using the Route 53 API in the Amazon Route 53 Developer Guide. Due to the nature of transactional changes, you cannot delete the same resource record set more than once in a single change batch. If you attempt to delete the same change batch more than once, Route 53 returns an InvalidChangeBatch error. In response to a ChangeResourceRecordSets request, your DNS data is changed on all Route 53 DNS servers. Initially, the status of a change is PENDING. This means the change has not yet propagated to all the authoritative Route 53 DNS servers. When the change is propagated to all hosts, the change returns a status of INSYNC. Note the following limitations on a ChangeResourceRecordSets request: - A request cannot contain more than 100 Change elements. - A request cannot contain more than 1000 ResourceRecord elements. The sum of the number of characters (including spaces) in all Value elements in a request cannot exceed 32,000 characters.

See: Route53

changeResourceRecordSets

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
crrsHostedZoneId

-> ChangeBatch
crrsChangeBatch

-> State ChangeResourceRecordSets a

-> m ChangeResourceRecordSetsResponse

changeResourceRecordSetsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
crrsHostedZoneId

-> ChangeBatch
crrsChangeBatch

-> State ChangeResourceRecordSets a

-> m (Either ServiceEr ChangeResourceRecordSetsResponse)

ChangeTagsForResource

See: Route53

changeTagsForResource

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> TagResourceType
ctfrResourceType

-> Text
ctfrResourceId

-> State ChangeTagsForResource a

-> m ChangeTagsForResourceResponse

changeTagsForResourceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> TagResourceType
ctfrResourceType

-> Text
ctfrResourceId

-> State ChangeTagsForResource a

-> m (Either ServiceEr ChangeTagsForResourceResponse)

CreateHealthCheck

This action creates a new health check. To create a new health check, send a POST request to the 2013-04-01/healthcheck resource. The request body must include an XML document with a CreateHealthCheckRequest element. The response returns the CreateHealthCheckResponse element that contains metadata about the health check.

See: Route53

createHealthCheck

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
chcCallerReference

-> HealthCheckConfig
chcHealthCheckConfig

-> State CreateHealthCheck a

-> m CreateHealthCheckResponse

createHealthCheckCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
chcCallerReference

-> HealthCheckConfig
chcHealthCheckConfig

-> State CreateHealthCheck a

-> m (Either ServiceEr CreateHealthCheckResponse)

CreateHostedZone

This action creates a new hosted zone. To create a new hosted zone, send a POST request to the 2013-04-01/hostedzone resource. The request body must include an XML document with a CreateHostedZoneRequest element. The response returns the CreateHostedZoneResponse element that contains metadata about the hosted zone. Route 53 automatically creates a default SOA record and four NS records for the zone. The NS records in the hosted zone are the name servers you give your registrar to delegate your domain to. For more information about SOA and NS records, see NS and SOA Records that Route 53 Creates for a Hosted Zone in the Amazon Route 53 Developer Guide. When you create a zone, its initial status is PENDING. This means that it is not yet available on all DNS servers. The status of the zone changes to INSYNC when the NS and SOA records are available on all Route 53 DNS servers.

See: Route53

createHostedZone

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
chzName

-> Text
chzCallerReference

-> State CreateHostedZone a

-> m CreateHostedZoneResponse

createHostedZoneCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
chzName

-> Text
chzCallerReference

-> State CreateHostedZone a

-> m (Either ServiceEr CreateHostedZoneResponse)

DeleteHealthCheck

This action deletes a health check. To delete a health check, send a DELETE request to the 2013-04-01healthcheckhealth check ID resource. You can delete a health check only if there are no resource record sets associated with this health check. If resource record sets are associated with this health check, you must disassociate them before you can delete your health check. If you try to delete a health check that is associated with resource record sets, Route 53 will deny your request with a HealthCheckInUse error. For information about disassociating the records from your health check, see ChangeResourceRecordSets.

See: Route53

deleteHealthCheck

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dhcHealthCheckId

-> State DeleteHealthCheck a

-> m DeleteHealthCheckResponse

deleteHealthCheckCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dhcHealthCheckId

-> State DeleteHealthCheck a

-> m (Either ServiceEr DeleteHealthCheckResponse)

DeleteHostedZone

This action deletes a hosted zone. To delete a hosted zone, send a DELETE request to the 2013-04-01hostedzonehosted zone ID resource. For more information about deleting a hosted zone, see Deleting a Hosted Zone in the Amazon Route 53 Developer Guide. You can delete a hosted zone only if there are no resource record sets other than the default SOA record and NS resource record sets. If your hosted zone contains other resource record sets, you must delete them before you can delete your hosted zone. If you try to delete a hosted zone that contains other resource record sets, Route 53 will deny your request with a HostedZoneNotEmpty error. For information about deleting records from your hosted zone, see ChangeResourceRecordSets.

See: Route53

deleteHostedZone

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dhzId

-> State DeleteHostedZone a

-> m DeleteHostedZoneResponse

deleteHostedZoneCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dhzId

-> State DeleteHostedZone a

-> m (Either ServiceEr DeleteHostedZoneResponse)

GetChange

This action returns the current status of a change batch request. The status is one of the following values: - PENDING indicates that the changes in this request have not replicated to all Route 53 DNS servers. This is the initial status of all change batch requests. - INSYNC indicates that the changes have replicated to all Amazon Route 53 DNS servers.

See: Route53

getChange

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
gcId

-> State GetChange a

-> m GetChangeResponse

getChangeCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
gcId

-> State GetChange a

-> m (Either ServiceEr GetChangeResponse)

GetCheckerIpRanges

To retrieve a list of the IP ranges used by Amazon Route 53 health checkers to check the health of your resources, send a GET request to the 2013-04-01/checkeripranges resource. You can use these IP addresses to configure router and firewall rules to allow health checkers to check the health of your resources.

See: Route53

getCheckerIpRanges :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State GetCheckerIpRanges a -> m GetCheckerIpRangesResponse

getCheckerIpRangesCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State GetCheckerIpRanges a -> m (Either ServiceEr GetCheckerIpRangesResponse)

GetGeoLocation

See: Route53

getGeoLocation :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State GetGeoLocation a -> m GetGeoLocationResponse

getGeoLocationCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State GetGeoLocation a -> m (Either ServiceEr GetGeoLocationResponse)

GetHealthCheck

To retrieve the health check, send a GET request to the 2013-04-01healthcheckhealth check ID resource.

See: Route53

getHealthCheck

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ghcHealthCheckId

-> State GetHealthCheck a

-> m GetHealthCheckResponse

getHealthCheckCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ghcHealthCheckId

-> State GetHealthCheck a

-> m (Either ServiceEr GetHealthCheckResponse)

GetHealthCheckCount

To retrieve a count of all your health checks, send a GET request to the 2013-04-01/healthcheckcount resource.

See: Route53

getHealthCheckCount :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State GetHealthCheckCount a -> m GetHealthCheckCountResponse

getHealthCheckCountCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State GetHealthCheckCount a -> m (Either ServiceEr GetHealthCheckCountResponse)

GetHostedZone

To retrieve the delegation set for a hosted zone, send a GET request to the 2013-04-01hostedzonehosted zone ID resource. The delegation set is the four Route 53 name servers that were assigned to the hosted zone when you created it.

See: Route53

getHostedZone

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ghzId

-> State GetHostedZone a

-> m GetHostedZoneResponse

getHostedZoneCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ghzId

-> State GetHostedZone a

-> m (Either ServiceEr GetHostedZoneResponse)

ListGeoLocations

See: Route53

listGeoLocations :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State ListGeoLocations a -> m ListGeoLocationsResponse

listGeoLocationsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State ListGeoLocations a -> m (Either ServiceEr ListGeoLocationsResponse)

ListHealthChecks

To retrieve a list of your health checks, send a GET request to the 2013-04-01/healthcheck resource. The response to this request includes a HealthChecks element with zero, one, or multiple HealthCheck child elements. By default, the list of health checks is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. You can use the Marker parameter to control the health check that the list begins with. Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Amazon Route 53 returns only the first 100.

See: Route53

listHealthChecks :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m)) => State ListHealthChecks a -> ResumableSource m ListHealthChecksResponse

listHealthChecksCatch :: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m)) => State ListHealthChecks a -> ResumableSource m (Either ServiceEr ListHealthChecksResponse)

ListHostedZones

To retrieve a list of your hosted zones, send a GET request to the 2013-04-01/hostedzone resource. The response to this request includes a HostedZones element with zero, one, or multiple HostedZone child elements. By default, the list of hosted zones is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. You can use the Marker parameter to control the hosted zone that the list begins with. Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Amazon Route 53 returns only the first 100.

See: Route53

listHostedZones :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m)) => State ListHostedZones a -> ResumableSource m ListHostedZonesResponse

listHostedZonesCatch :: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m)) => State ListHostedZones a -> ResumableSource m (Either ServiceEr ListHostedZonesResponse)

ListResourceRecordSets

Imagine all the resource record sets in a zone listed out in front of you. Imagine them sorted lexicographically first by DNS name (with the labels reversed, like "com.amazon.www" for example), and secondarily, lexicographically by record type. This operation retrieves at most MaxItems resource record sets from this list, in order, starting at a position specified by the Name and Type arguments: If both Name and Type are omitted, this means start the results at the first RRSET in the HostedZone. If Name is specified but Type is omitted, this means start the results at the first RRSET in the list whose name is greater than or equal to Name. If both Name and Type are specified, this means start the results at the first RRSET in the list whose name is greater than or equal to Name and whose type is greater than or equal to Type. It is an error to specify the Type but not the Name. Use ListResourceRecordSets to retrieve a single known record set by specifying the record set's name and type, and setting MaxItems = 1 To retrieve all the records in a HostedZone, first pause any processes making calls to ChangeResourceRecordSets. Initially call ListResourceRecordSets without a Name and Type to get the first page of record sets. For subsequent calls, set Name and Type to the NextName and NextType values returned by the previous response. In the presence of concurrent ChangeResourceRecordSets calls, there is no consistency of results across calls to ListResourceRecordSets. The only way to get a consistent multi-page snapshot of all RRSETs in a zone is to stop making changes while pagination is in progress. However, the results from ListResourceRecordSets are consistent within a page. If MakeChange calls are taking place concurrently, the result of each one will either be completely visible in your results or not at all. You will not see partial changes, or changes that do not ultimately succeed. (This follows from the fact that MakeChange is atomic) The results from ListResourceRecordSets are strongly consistent with ChangeResourceRecordSets. To be precise, if a single process makes a call to ChangeResourceRecordSets and receives a successful response, the effects of that change will be visible in a subsequent call to ListResourceRecordSets by that process.

See: Route53

listResourceRecordSets

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m))

=> Text
lrrsHostedZoneId

-> State ListResourceRecordSets a

-> ResumableSource m ListResourceRecordSetsResponse

listResourceRecordSetsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m))

=> Text
lrrsHostedZoneId

-> State ListResourceRecordSets a

-> ResumableSource m (Either ServiceEr ListResourceRecordSetsResponse)

ListTagsForResource

See: Route53

listTagsForResource

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> TagResourceType
ltfrResourceType

-> Text
ltfrResourceId

-> State ListTagsForResource a

-> m ListTagsForResourceResponse

listTagsForResourceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> TagResourceType
ltfrResourceType

-> Text
ltfrResourceId

-> State ListTagsForResource a

-> m (Either ServiceEr ListTagsForResourceResponse)

ListTagsForResources

See: Route53

listTagsForResources

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> TagResourceType
ltfr1ResourceType

-> List1 Text
ltfr1ResourceIds

-> State ListTagsForResources a

-> m ListTagsForResourcesResponse

listTagsForResourcesCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> TagResourceType
ltfr1ResourceType

-> List1 Text
ltfr1ResourceIds

-> State ListTagsForResources a

-> m (Either ServiceEr ListTagsForResourcesResponse)

UpdateHealthCheck

This action updates an existing health check. To update a health check, send a POST request to the 2013-05-27healthcheckhealth check ID resource. The request body must include an XML document with an UpdateHealthCheckRequest element. The response returns an UpdateHealthCheckResponse element, which contains metadata about the health check.

See: Route53

updateHealthCheck

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
uhcHealthCheckId

-> State UpdateHealthCheck a

-> m UpdateHealthCheckResponse

updateHealthCheckCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
uhcHealthCheckId

-> State UpdateHealthCheck a

-> m (Either ServiceEr UpdateHealthCheckResponse)

Re-exported

module Network.AWS.Route53

(.=) :: MonadState s m => ASetter s s a b -> b -> m ()

(?=) :: MonadState s m => ASetter s s a (Maybe b) -> b -> m ()

(<>=) :: (MonadState s m, Monoid a) => ASetter' s a -> a -> m ()

(%=) :: (Profunctor p, MonadState s m) => Setting p s s a b -> p a b -> m ()