Network.AWS.Route53.TestDNSAnswer

Description

Gets the value that Amazon Route 53 returns in response to a DNS request for a specified record name and type. You can optionally specify the IP address of a DNS resolver, an EDNS0 client subnet IP address, and a subnet mask.

Creating a Request

testDNSAnswer

Arguments

:: ResourceId
tdaHostedZoneId

-> Text
tdaRecordName

-> RecordType
tdaRecordType

-> TestDNSAnswer

Creates a value of TestDNSAnswer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tdaResolverIP - If you want to simulate a request from a specific DNS resolver, specify the IP address for that resolver. If you omit this value, TestDnsAnswer uses the IP address of a DNS resolver in the AWS US East (N. Virginia) Region (us-east-1 ).

tdaEDNS0ClientSubnetIP - If the resolver that you specified for resolverip supports EDNS0, specify the IPv4 or IPv6 address of a client in the applicable location, for example, 192.0.2.44 or 2001:db8:85a3::8a2e:370:7334 .

tdaEDNS0ClientSubnetMask - If you specify an IP address for edns0clientsubnetip , you can optionally specify the number of bits of the IP address that you want the checking tool to include in the DNS query. For example, if you specify 192.0.2.44 for edns0clientsubnetip and 24 for edns0clientsubnetmask , the checking tool will simulate a request from 192.0.2.0/24. The default value is 24 bits for IPv4 addresses and 64 bits for IPv6 addresses.

tdaHostedZoneId - The ID of the hosted zone that you want Amazon Route 53 to simulate a query for.

tdaRecordName - The name of the resource record set that you want Amazon Route 53 to simulate a query for.

tdaRecordType - The type of the resource record set.

data TestDNSAnswer

Gets the value that Amazon Route 53 returns in response to a DNS request for a specified record name and type. You can optionally specify the IP address of a DNS resolver, an EDNS0 client subnet IP address, and a subnet mask.

See: testDNSAnswer smart constructor.

Request Lenses

tdaResolverIP :: Lens' TestDNSAnswer (Maybe Text)

If you want to simulate a request from a specific DNS resolver, specify the IP address for that resolver. If you omit this value, TestDnsAnswer uses the IP address of a DNS resolver in the AWS US East (N. Virginia) Region (us-east-1 ).

tdaEDNS0ClientSubnetIP :: Lens' TestDNSAnswer (Maybe Text)

If the resolver that you specified for resolverip supports EDNS0, specify the IPv4 or IPv6 address of a client in the applicable location, for example, 192.0.2.44 or 2001:db8:85a3::8a2e:370:7334 .

tdaEDNS0ClientSubnetMask :: Lens' TestDNSAnswer (Maybe Text)

If you specify an IP address for edns0clientsubnetip , you can optionally specify the number of bits of the IP address that you want the checking tool to include in the DNS query. For example, if you specify 192.0.2.44 for edns0clientsubnetip and 24 for edns0clientsubnetmask , the checking tool will simulate a request from 192.0.2.0/24. The default value is 24 bits for IPv4 addresses and 64 bits for IPv6 addresses.

tdaHostedZoneId :: Lens' TestDNSAnswer ResourceId

The ID of the hosted zone that you want Amazon Route 53 to simulate a query for.

tdaRecordName :: Lens' TestDNSAnswer Text

The name of the resource record set that you want Amazon Route 53 to simulate a query for.

tdaRecordType :: Lens' TestDNSAnswer RecordType

The type of the resource record set.

Destructuring the Response

testDNSAnswerResponse

Arguments

:: Int
tdarsResponseStatus

-> Text
tdarsNameserver

-> Text
tdarsRecordName

-> RecordType
tdarsRecordType

-> Text
tdarsResponseCode

-> Text
tdarsProtocol

-> TestDNSAnswerResponse

Creates a value of TestDNSAnswerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tdarsResponseStatus - -- | The response status code.

tdarsNameserver - The Amazon Route 53 name server used to respond to the request.

tdarsRecordName - The name of the resource record set that you submitted a request for.

tdarsRecordType - The type of the resource record set that you submitted a request for.

tdarsRecordData - A list that contains values that Amazon Route 53 returned for this resource record set.

tdarsResponseCode - A code that indicates whether the request is valid or not. The most common response code is NOERROR , meaning that the request is valid. If the response is not valid, Amazon Route 53 returns a response code that describes the error. For a list of possible response codes, see DNS RCODES on the IANA website.

tdarsProtocol - The protocol that Amazon Route 53 used to respond to the request, either UDP or TCP .

data TestDNSAnswerResponse

A complex type that contains the response to a TestDNSAnswer request.

See: testDNSAnswerResponse smart constructor.

Response Lenses

tdarsResponseStatus :: Lens' TestDNSAnswerResponse Int

- | The response status code.

tdarsNameserver :: Lens' TestDNSAnswerResponse Text

The Amazon Route 53 name server used to respond to the request.

tdarsRecordName :: Lens' TestDNSAnswerResponse Text

The name of the resource record set that you submitted a request for.

tdarsRecordType :: Lens' TestDNSAnswerResponse RecordType

The type of the resource record set that you submitted a request for.

tdarsRecordData :: Lens' TestDNSAnswerResponse [Text]

A list that contains values that Amazon Route 53 returned for this resource record set.

tdarsResponseCode :: Lens' TestDNSAnswerResponse Text

A code that indicates whether the request is valid or not. The most common response code is NOERROR , meaning that the request is valid. If the response is not valid, Amazon Route 53 returns a response code that describes the error. For a list of possible response codes, see DNS RCODES on the IANA website.

tdarsProtocol :: Lens' TestDNSAnswerResponse Text

The protocol that Amazon Route 53 used to respond to the request, either UDP or TCP .