Network.AWS.Route53.Types

Service

data Route53

Version 2013-04-01 of the Amazon Route 53 service.

Error

data RESTError :: *

XML

ns :: Text

AliasTarget

data AliasTarget

aliasTarget

Arguments

:: Text
atHostedZoneId

-> Text
atDNSName

-> Bool
atEvaluateTargetHealth

-> AliasTarget

AliasTarget constructor.

The fields accessible through corresponding lenses are:

atDNSName :: Text

atEvaluateTargetHealth :: Bool

atHostedZoneId :: Text

atDNSName :: Lens' AliasTarget Text

Alias resource record sets only: The external DNS name associated with the AWS Resource.

For more information and an example, see Creating Alias Resource Record Sets in the Amazon Route 53 Developer Guide

.

atEvaluateTargetHealth :: Lens' AliasTarget Bool

Alias resource record sets only: A boolean value that indicates whether this Resource Record Set should respect the health status of any health checks associated with the ALIAS target record which it is linked to.

For more information and an example, see Creating Alias Resource Record Sets in the Amazon Route 53 Developer Guide

.

atHostedZoneId :: Lens' AliasTarget Text

Alias resource record sets only: The value of the hosted zone ID for the AWS resource.

For more information and an example, see Creating Alias Resource Record Sets in the Amazon Route 53 Developer Guide

.

ResourceRecord

data ResourceRecord

resourceRecord

Arguments

:: Text
rrValue

-> ResourceRecord

ResourceRecord constructor.

The fields accessible through corresponding lenses are:

rrValue :: Text

rrValue :: Lens' ResourceRecord Text

The value of the Value element for the current resource record set.

Tag

data Tag

tag :: Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Maybe Text

tagValue :: Maybe Text

tagKey :: Lens' Tag (Maybe Text)

The key for a Tag.

tagValue :: Lens' Tag (Maybe Text)

The value for a Tag.

GeoLocationDetails

data GeoLocationDetails

geoLocationDetails :: GeoLocationDetails

GeoLocationDetails constructor.

The fields accessible through corresponding lenses are:

gldContinentCode :: Maybe Text

gldContinentName :: Maybe Text

gldCountryCode :: Maybe Text

gldCountryName :: Maybe Text

gldSubdivisionCode :: Maybe Text

gldSubdivisionName :: Maybe Text

gldContinentCode :: Lens' GeoLocationDetails (Maybe Text)

The code for a continent geo location. Note: only continent locations have a continent code.

gldContinentName :: Lens' GeoLocationDetails (Maybe Text)

The name of the continent. This element is only present if ContinentCode is also present.

gldCountryCode :: Lens' GeoLocationDetails (Maybe Text)

The code for a country geo location. The default location uses * for the country code and will match all locations that are not matched by a geo location.

The default geo location uses a * for the country code. All other country codes follow the ISO 3166 two-character code.

gldCountryName :: Lens' GeoLocationDetails (Maybe Text)

The name of the country. This element is only present if CountryCode is also present.

gldSubdivisionCode :: Lens' GeoLocationDetails (Maybe Text)

The code for a country's subdivision (e.g., a province of Canada). A subdivision code is only valid with the appropriate country code.

gldSubdivisionName :: Lens' GeoLocationDetails (Maybe Text)

The name of the subdivision. This element is only present if SubdivisionCode is also present.

HealthCheck

data HealthCheck

healthCheck

Arguments

:: Text
hcId

-> Text
hcCallerReference

-> HealthCheckConfig
hcHealthCheckConfig

-> Natural
hcHealthCheckVersion

-> HealthCheck

HealthCheck constructor.

The fields accessible through corresponding lenses are:

hcCallerReference :: Text

hcHealthCheckConfig :: HealthCheckConfig

hcHealthCheckVersion :: Natural

hcId :: Text

hcCallerReference :: Lens' HealthCheck Text

A unique string that identifies the request to create the health check.

hcHealthCheckConfig :: Lens' HealthCheck HealthCheckConfig

A complex type that contains the health check configuration.

hcHealthCheckVersion :: Lens' HealthCheck Natural

The version of the health check. You can optionally pass this value in a call to UpdateHealthCheck to prevent overwriting another change to the health check.

hcId :: Lens' HealthCheck Text

The ID of the specified health check.

VPCRegion

data VPCRegion

Constructors

ApNortheast1
ap-northeast-1

ApSoutheast1
ap-southeast-1

ApSoutheast2
ap-southeast-2

CnNorth1
cn-north-1

EuCentral1
eu-central-1

EuWest1
eu-west-1

SaEast1
sa-east-1

UsEast1
us-east-1

UsWest1
us-west-1

UsWest2
us-west-2

RecordType

data RecordType

Constructors

A
A

AAAA
AAAA

CNAME
CNAME

MX
MX

NS
NS

PTR
PTR

SOA
SOA

SPF
SPF

SRV
SRV

TXT
TXT

ChangeAction

data ChangeAction

Constructors

Create
CREATE

Delete'
DELETE

Upsert
UPSERT

TagResourceType

data TagResourceType

Constructors

Healthcheck
healthcheck

Hostedzone
hostedzone

HealthCheckConfig

data HealthCheckConfig

healthCheckConfig

Arguments

:: HealthCheckType
hccType

-> HealthCheckConfig

HealthCheckConfig constructor.

The fields accessible through corresponding lenses are:

hccFailureThreshold :: Maybe Natural

hccFullyQualifiedDomainName :: Maybe Text

hccIPAddress :: Maybe Text

hccPort :: Maybe Natural

hccRequestInterval :: Maybe Natural

hccResourcePath :: Maybe Text

hccSearchString :: Maybe Text

hccType :: HealthCheckType

hccFailureThreshold :: Lens' HealthCheckConfig (Maybe Natural)

The number of consecutive health checks that an endpoint must pass or fail for Route 53 to change the current status of the endpoint from unhealthy to healthy or vice versa.

Valid values are integers between 1 and 10. For more information, see "How Amazon Route 53 Determines Whether an Endpoint Is Healthy" in the Amazon Route 53 Developer Guide.

hccFullyQualifiedDomainName :: Lens' HealthCheckConfig (Maybe Text)

Fully qualified domain name of the instance to be health checked.

hccIPAddress :: Lens' HealthCheckConfig (Maybe Text)

IP Address of the instance being checked.

hccPort :: Lens' HealthCheckConfig (Maybe Natural)

Port on which connection will be opened to the instance to health check. For HTTP and HTTP_STR_MATCH this defaults to 80 if the port is not specified. For HTTPS and HTTPS_STR_MATCH this defaults to 443 if the port is not specified.

hccRequestInterval :: Lens' HealthCheckConfig (Maybe Natural)

The number of seconds between the time that Route 53 gets a response from your endpoint and the time that it sends the next health-check request.

Each Route 53 health checker makes requests at this interval. Valid values are 10 and 30. The default value is 30.

hccResourcePath :: Lens' HealthCheckConfig (Maybe Text)

Path to ping on the instance to check the health. Required for HTTP, HTTPS, HTTP_STR_MATCH, and HTTPS_STR_MATCH health checks, HTTP request is issued to the instance on the given port and path.

hccSearchString :: Lens' HealthCheckConfig (Maybe Text)

A string to search for in the body of a health check response. Required for HTTP_STR_MATCH and HTTPS_STR_MATCH health checks.

hccType :: Lens' HealthCheckConfig HealthCheckType

The type of health check to be performed. Currently supported types are TCP, HTTP, HTTPS, HTTP_STR_MATCH, and HTTPS_STR_MATCH.

Change

data Change

change

Arguments

:: ChangeAction
cAction

-> ResourceRecordSet
cResourceRecordSet

-> Change

Change constructor.

The fields accessible through corresponding lenses are:

cAction :: ChangeAction

cResourceRecordSet :: ResourceRecordSet

cAction :: Lens' Change ChangeAction

The action to perform.

Valid values: CREATE | DELETE | UPSERT

cResourceRecordSet :: Lens' Change ResourceRecordSet

Information about the resource record set to create or delete.

Failover

data Failover

Constructors

Primary
PRIMARY

Secondary
SECONDARY

HostedZone

data HostedZone

hostedZone

Arguments

:: Text
hzId

-> Text
hzName

-> Text
hzCallerReference

-> HostedZone

HostedZone constructor.

The fields accessible through corresponding lenses are:

hzCallerReference :: Text

hzConfig :: Maybe HostedZoneConfig

hzId :: Text

hzName :: Text

hzResourceRecordSetCount :: Maybe Integer

hzCallerReference :: Lens' HostedZone Text

A unique string that identifies the request to create the hosted zone.

hzConfig :: Lens' HostedZone (Maybe HostedZoneConfig)

A complex type that contains the Comment element.

hzId :: Lens' HostedZone Text

The ID of the specified hosted zone.

hzName :: Lens' HostedZone Text

The name of the domain. This must be a fully-specified domain, for example, www.example.com. The trailing dot is optional; Route 53 assumes that the domain name is fully qualified. This means that Route 53 treats www.example.com (without a trailing dot) and www.example.com. (with a trailing dot) as identical.

This is the name you have registered with your DNS registrar. You should ask your registrar to change the authoritative name servers for your domain to the set of NameServers elements returned in DelegationSet.

hzResourceRecordSetCount :: Lens' HostedZone (Maybe Integer)

Total number of resource record sets in the hosted zone.

ResourceTagSet

data ResourceTagSet

resourceTagSet

Arguments

:: NonEmpty Tag
rtsTags

-> ResourceTagSet

ResourceTagSet constructor.

The fields accessible through corresponding lenses are:

rtsResourceId :: Maybe Text

rtsResourceType :: Maybe TagResourceType

rtsTags :: NonEmpty Tag

rtsResourceId :: Lens' ResourceTagSet (Maybe Text)

The ID for the specified resource.

rtsResourceType :: Lens' ResourceTagSet (Maybe TagResourceType)

The type of the resource.

The resource type for health checks is healthcheck.

The resource type for hosted zones is hostedzone.

rtsTags :: Lens' ResourceTagSet (NonEmpty Tag)

The tags associated with the specified resource.

ChangeStatus

data ChangeStatus

Constructors

Insync
INSYNC

Pending
PENDING

ChangeBatch

data ChangeBatch

changeBatch

Arguments

:: NonEmpty Change
cbChanges

-> ChangeBatch

ChangeBatch constructor.

The fields accessible through corresponding lenses are:

cbChanges :: NonEmpty Change

cbComment :: Maybe Text

cbChanges :: Lens' ChangeBatch (NonEmpty Change)

A complex type that contains one Change element for each resource record set that you want to create or delete.

cbComment :: Lens' ChangeBatch (Maybe Text)

Optional: Any comments you want to include about a change batch request.

StatusReport

data StatusReport

statusReport :: StatusReport

StatusReport constructor.

The fields accessible through corresponding lenses are:

srCheckedTime :: Maybe UTCTime

srStatus :: Maybe Text

srCheckedTime :: Lens' StatusReport (Maybe UTCTime)

The date and time the health check status was observed, in the format 'YYYY-MM-DDThh:mm:ssZ', as specified in the ISO 8601 standard (for example, 2009-11-19T19:37:58Z). The Z after the time indicates that the time is listed in Coordinated Universal Time (UTC), which is synonymous with Greenwich Mean Time in this context.

srStatus :: Lens' StatusReport (Maybe Text)

The observed health check status.

HealthCheckType

data HealthCheckType

Constructors

Http
HTTP

HttpStrMatch
HTTP_STR_MATCH

Https
HTTPS

HttpsStrMatch
HTTPS_STR_MATCH

Tcp
TCP

VPC

data VPC

vpc :: VPC

VPC constructor.

The fields accessible through corresponding lenses are:

vpcVPCId :: Maybe Text

vpcVPCRegion :: Maybe VPCRegion

vpcVPCId :: Lens' VPC (Maybe Text)

vpcVPCRegion :: Lens' VPC (Maybe VPCRegion)

HostedZoneConfig

data HostedZoneConfig

hostedZoneConfig :: HostedZoneConfig

HostedZoneConfig constructor.

The fields accessible through corresponding lenses are:

hzcComment :: Maybe Text

hzcPrivateZone :: Maybe Bool

hzcComment :: Lens' HostedZoneConfig (Maybe Text)

An optional comment about your hosted zone. If you don't want to specify a comment, you can omit the HostedZoneConfig and Comment elements from the XML document.

hzcPrivateZone :: Lens' HostedZoneConfig (Maybe Bool)

A value that indicates whether this is a private hosted zone. The value is returned in the response; do not specify it in the request.

ResourceRecordSet

data ResourceRecordSet

resourceRecordSet

Arguments

:: Text
rrsName

-> RecordType
rrsType

-> NonEmpty ResourceRecord
rrsResourceRecords

-> ResourceRecordSet

ResourceRecordSet constructor.

The fields accessible through corresponding lenses are:

rrsAliasTarget :: Maybe AliasTarget

rrsFailover :: Maybe Failover

rrsGeoLocation :: Maybe GeoLocation

rrsHealthCheckId :: Maybe Text

rrsName :: Text

rrsRegion :: Maybe Region

rrsResourceRecords :: NonEmpty ResourceRecord

rrsSetIdentifier :: Maybe Text

rrsTTL :: Maybe Natural

rrsType :: RecordType

rrsWeight :: Maybe Natural

rrsAliasTarget :: Lens' ResourceRecordSet (Maybe AliasTarget)

Alias resource record sets only: Information about the AWS resource to which you are redirecting traffic.

rrsFailover :: Lens' ResourceRecordSet (Maybe Failover)

Failover resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that indicates whether the current resource record set is a primary or secondary resource record set. A failover set may contain at most one resource record set marked as primary and one resource record set marked as secondary. A resource record set marked as primary will be returned if any of the following are true: (1) an associated health check is passing, (2) if the resource record set is an alias with the evaluate target health and at least one target resource record set is healthy, (3) both the primary and secondary resource record set are failing health checks or (4) there is no secondary resource record set. A secondary resource record set will be returned if: (1) the primary is failing a health check and either the secondary is passing a health check or has no associated health check, or (2) there is no primary resource record set.

Valid values: PRIMARY | SECONDARY

rrsGeoLocation :: Lens' ResourceRecordSet (Maybe GeoLocation)

Geo location resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that specifies the geo location for the current resource record set.

rrsHealthCheckId :: Lens' ResourceRecordSet (Maybe Text)

Health Check resource record sets only, not required for alias resourcerecord sets: An identifier that is used to identify health check associated with the resource record set.

rrsName :: Lens' ResourceRecordSet Text

The domain name of the current resource record set.

rrsRegion :: Lens' ResourceRecordSet (Maybe Region)

Latency-based resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that specifies the AWS region for the current resource record set.

rrsResourceRecords :: Lens' ResourceRecordSet (NonEmpty ResourceRecord)

A complex type that contains the resource records for the current resource record set.

rrsSetIdentifier :: Lens' ResourceRecordSet (Maybe Text)

Weighted, Latency, Geo, and Failover resource record sets only: An identifier that differentiates among multiple resource record sets that have the same combination of DNS name and type.

rrsTTL :: Lens' ResourceRecordSet (Maybe Natural)

The cache time to live for the current resource record set.

rrsType :: Lens' ResourceRecordSet RecordType

The type of the current resource record set.

rrsWeight :: Lens' ResourceRecordSet (Maybe Natural)

Weighted resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that determines what portion of traffic for the current resource record set is routed to the associated location.

DelegationSet

data DelegationSet

delegationSet

Arguments

:: NonEmpty Text
dsNameServers

-> DelegationSet

DelegationSet constructor.

The fields accessible through corresponding lenses are:

dsCallerReference :: Maybe Text

dsId :: Maybe Text

dsNameServers :: NonEmpty Text

dsCallerReference :: Lens' DelegationSet (Maybe Text)

dsId :: Lens' DelegationSet (Maybe Text)

dsNameServers :: Lens' DelegationSet (NonEmpty Text)

A complex type that contains the authoritative name servers for the hosted zone. Use the method provided by your domain registrar to add an NS record to your domain for each NameServer that is assigned to your hosted zone.

ChangeInfo

data ChangeInfo

changeInfo

Arguments

:: Text
ciId

-> ChangeStatus
ciStatus

-> UTCTime
ciSubmittedAt

-> ChangeInfo

ChangeInfo constructor.

The fields accessible through corresponding lenses are:

ciComment :: Maybe Text

ciId :: Text

ciStatus :: ChangeStatus

ciSubmittedAt :: UTCTime

ciComment :: Lens' ChangeInfo (Maybe Text)

A complex type that describes change information about changes made to your hosted zone.

This element contains an ID that you use when performing a GetChange action to get detailed information about the change.

ciId :: Lens' ChangeInfo Text

The ID of the request. Use this ID to track when the change has completed across all Amazon Route 53 DNS servers.

ciStatus :: Lens' ChangeInfo ChangeStatus

The current state of the request. PENDING indicates that this request has not yet been applied to all Amazon Route 53 DNS servers.

Valid Values: PENDING | INSYNC

ciSubmittedAt :: Lens' ChangeInfo UTCTime

The date and time the change was submitted, in the format 'YYYY-MM-DDThh:mm:ssZ' , as specified in the ISO 8601 standard (for example, 2009-11-19T19:37:58Z). The Z after the time indicates that the time is listed in Coordinated Universal Time (UTC), which is synonymous with Greenwich Mean Time in this context.

GeoLocation

data GeoLocation

geoLocation :: GeoLocation

GeoLocation constructor.

The fields accessible through corresponding lenses are:

glContinentCode :: Maybe Text

glCountryCode :: Maybe Text

glSubdivisionCode :: Maybe Text

glContinentCode :: Lens' GeoLocation (Maybe Text)

The code for a continent geo location. Note: only continent locations have a continent code.

Valid values: AF | AN | AS | EU | OC | NA | SA

Constraint: Specifying ContinentCode with either CountryCode or SubdivisionCode returns an InvalidInput error.

glCountryCode :: Lens' GeoLocation (Maybe Text)

The code for a country geo location. The default location uses * for the country code and will match all locations that are not matched by a geo location.

The default geo location uses a * for the country code. All other country codes follow the ISO 3166 two-character code.

glSubdivisionCode :: Lens' GeoLocation (Maybe Text)

The code for a country's subdivision (e.g., a province of Canada). A subdivision code is only valid with the appropriate country code.

Constraint: Specifying SubdivisionCode without CountryCode returns an InvalidInput error.

HealthCheckObservation

data HealthCheckObservation

healthCheckObservation :: HealthCheckObservation

HealthCheckObservation constructor.

The fields accessible through corresponding lenses are:

hcoIPAddress :: Maybe Text

hcoStatusReport :: Maybe StatusReport

hcoIPAddress :: Lens' HealthCheckObservation (Maybe Text)

The IP address of the Route 53 health checker that performed the health check.

hcoStatusReport :: Lens' HealthCheckObservation (Maybe StatusReport)

A complex type that contains information about the health check status for the current observation.

Common

data Region :: *

Constructors

Ireland

Frankfurt

Tokyo

Singapore

Sydney

Beijing

NorthVirginia

NorthCalifornia

Oregon

GovCloud

GovCloudFIPS

SaoPaulo