Network.AWS.Route53.UpdateHealthCheck

Description

This action updates an existing health check.

To update a health check, send a POST request to the '2013-04-01healthcheckhealth check ID resource. The request body must include an XML document with an UpdateHealthCheckRequest element. The response returns an UpdateHealthCheckResponse element, which contains metadata about the health check.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_UpdateHealthCheck.html

Request

data UpdateHealthCheck

Request constructor

updateHealthCheck

Arguments

:: Text
uhcHealthCheckId

-> UpdateHealthCheck

UpdateHealthCheck constructor.

The fields accessible through corresponding lenses are:

uhcFailureThreshold :: Maybe Natural

uhcFullyQualifiedDomainName :: Maybe Text

uhcHealthCheckId :: Text

uhcHealthCheckVersion :: Maybe Natural

uhcIPAddress :: Maybe Text

uhcPort :: Maybe Natural

uhcResourcePath :: Maybe Text

uhcSearchString :: Maybe Text

Request lenses

uhcFailureThreshold :: Lens' UpdateHealthCheck (Maybe Natural)

The number of consecutive health checks that an endpoint must pass or fail for Route 53 to change the current status of the endpoint from unhealthy to healthy or vice versa.

Valid values are integers between 1 and 10. For more information, see "How Amazon Route 53 Determines Whether an Endpoint Is Healthy" in the Amazon Route 53 Developer Guide.

Specify this value only if you want to change it.

uhcFullyQualifiedDomainName :: Lens' UpdateHealthCheck (Maybe Text)

Fully qualified domain name of the instance to be health checked.

Specify this value only if you want to change it.

uhcHealthCheckId :: Lens' UpdateHealthCheck Text

The ID of the health check to update.

uhcHealthCheckVersion :: Lens' UpdateHealthCheck (Maybe Natural)

Optional. When you specify a health check version, Route 53 compares this value with the current value in the health check, which prevents you from updating the health check when the versions don't match. Using HealthCheckVersion lets you prevent overwriting another change to the health check.

uhcIPAddress :: Lens' UpdateHealthCheck (Maybe Text)

The IP address of the resource that you want to check.

Specify this value only if you want to change it.

uhcPort :: Lens' UpdateHealthCheck (Maybe Natural)

The port on which you want Route 53 to open a connection to perform health checks.

Specify this value only if you want to change it.

uhcResourcePath :: Lens' UpdateHealthCheck (Maybe Text)

The path that you want Amazon Route 53 to request when performing health checks. The path can be any value for which your endpoint will return an HTTP status code of 2xx or 3xx when the endpoint is healthy, for example the file docsroute53-health-check.html.

Specify this value only if you want to change it.

uhcSearchString :: Lens' UpdateHealthCheck (Maybe Text)

If the value of Type is HTTP_STR_MATCH or HTTP_STR_MATCH, the string that you want Route 53 to search for in the response body from the specified resource. If the string appears in the response body, Route 53 considers the resource healthy.

Specify this value only if you want to change it.

Response

data UpdateHealthCheckResponse

Response constructor

updateHealthCheckResponse

Arguments

:: HealthCheck
uhcrHealthCheck

-> UpdateHealthCheckResponse

UpdateHealthCheckResponse constructor.

The fields accessible through corresponding lenses are:

uhcrHealthCheck :: HealthCheck

Response lenses

uhcrHealthCheck :: Lens' UpdateHealthCheckResponse HealthCheck