Network.AWS.Route53.UpdateHostedZoneComment

Description

To update the hosted zone comment, send a POST request to the '2013-04-01hostedzonehosted zone ID resource. The request body must include an XML document with a UpdateHostedZoneCommentRequest element. The response to this request includes the modified HostedZone element.

The comment can have a maximum length of 256 characters.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_UpdateHostedZoneComment.html

Request

data UpdateHostedZoneComment

Request constructor

updateHostedZoneComment

Arguments

:: Text
uhzcId

-> UpdateHostedZoneComment

UpdateHostedZoneComment constructor.

The fields accessible through corresponding lenses are:

uhzcComment :: Maybe Text

uhzcId :: Text

Request lenses

uhzcComment :: Lens' UpdateHostedZoneComment (Maybe Text)

A comment about your hosted zone.

uhzcId :: Lens' UpdateHostedZoneComment Text

The ID of the hosted zone you want to update.

Response

data UpdateHostedZoneCommentResponse

Response constructor

updateHostedZoneCommentResponse

Arguments

:: HostedZone
uhzcrHostedZone

-> UpdateHostedZoneCommentResponse

UpdateHostedZoneCommentResponse constructor.

The fields accessible through corresponding lenses are:

uhzcrHostedZone :: HostedZone

Response lenses

uhzcrHostedZone :: Lens' UpdateHostedZoneCommentResponse HostedZone