Network.AWS.Route53.V2013_04_01

Description

Amazon Route 53 is a scalable Domain Name System (DNS) web service. It provides secure and reliable routing to your infrastructure that uses Amazon Web Services (AWS) products, such as Amazon Elastic Compute Cloud (Amazon EC2), Elastic Load Balancing, or Amazon Simple Storage Service (Amazon S3). You can also use Amazon Route 53 to route users to your infrastructure outside of AWS.

Documentation

data ChangeResourceRecordSets

Constructors

ChangeResourceRecordSets

Fields

_crrsrChangeBatch :: ChangeBatch

A complex type that contains an optional comment and the Changes element.

_crrsrHostedZoneId :: Text

The ID of the hosted zone that contains the resource record sets that you want to change.

crrsrHostedZoneId :: Lens' ChangeResourceRecordSets Text

crrsrChangeBatch :: Lens' ChangeResourceRecordSets ChangeBatch

data ChangeResourceRecordSetsResponse

Constructors

ChangeResourceRecordSetsResponse

Fields

_crrssChangeInfo :: ChangeInfo

A complex type that contains information about changes made to your hosted zone. This element contains an ID that you use when performing a GetChange action to get detailed information about the change.

crrssChangeInfo :: Iso' ChangeResourceRecordSetsResponse ChangeInfo

changeTagsForResource

Arguments

:: Text
_ctfrrResourceId

-> TagResourceType
_ctfrrResourceType

-> ChangeTagsForResource

Minimum specification for a ChangeTagsForResource request.

data ChangeTagsForResource

Constructors

ChangeTagsForResource

Fields

_ctfrrResourceId :: Text

The ID of the resource for which you want to add, change, or delete tags.

_ctfrrResourceType :: TagResourceType

The type of the resource. The resource type for health checks is healthcheck.

_ctfrrRemoveTagKeys :: Maybe [Text]

A list of Tag keys that you want to remove from the specified resource.

_ctfrrAddTags :: Maybe [Tag]

A complex type that contains a list of Tag elements. Each Tag element identifies a tag that you want to add or update for the specified resource.

ctfrrResourceType :: Lens' ChangeTagsForResource TagResourceType

ctfrrResourceId :: Lens' ChangeTagsForResource Text

ctfrrRemoveTagKeys :: Lens' ChangeTagsForResource (Maybe [Text])

ctfrrAddTags :: Lens' ChangeTagsForResource (Maybe [Tag])

data ChangeTagsForResourceResponse

Constructors

ChangeTagsForResourceResponse

data CreateHealthCheck

Constructors

CreateHealthCheck

Fields

_chcrHealthCheckConfig :: HealthCheckConfig

A complex type that contains health check configuration.

_chcrCallerReference :: Text

A unique string that identifies the request and that allows failed CreateHealthCheck requests to be retried without the risk of executing the operation twice. You must use a unique CallerReference string every time you create a health check. CallerReference can be any unique string; you might choose to use a string that identifies your project. Valid characters are any Unicode code points that are legal in an XML 1.0 document. The UTF-8 encoding of the value must be less than 128 bytes.

chcrHealthCheckConfig :: Lens' CreateHealthCheck HealthCheckConfig

chcrCallerReference :: Lens' CreateHealthCheck Text

data CreateHealthCheckResponse

Constructors

CreateHealthCheckResponse

Fields

_chcsHealthCheck :: HealthCheck

A complex type that contains identifying information about the health check.

_chcsLocation :: Text

The unique URL representing the new health check.

chcsLocation :: Lens' CreateHealthCheckResponse Text

chcsHealthCheck :: Lens' CreateHealthCheckResponse HealthCheck

createHostedZone

Arguments

:: Text
_chzrName

-> Text
_chzrCallerReference

-> CreateHostedZone

Minimum specification for a CreateHostedZone request.

data CreateHostedZone

Constructors

CreateHostedZone

Fields

_chzrName :: Text

The name of the domain. This must be a fully-specified domain, for example, www.example.com. The trailing dot is optional; Route 53 assumes that the domain name is fully qualified. This means that Route 53 treats www.example.com (without a trailing dot) and www.example.com. (with a trailing dot) as identical. This is the name you have registered with your DNS registrar. You should ask your registrar to change the authoritative name servers for your domain to the set of NameServers elements returned in DelegationSet.

_chzrCallerReference :: Text

A unique string that identifies the request and that allows failed CreateHostedZone requests to be retried without the risk of executing the operation twice. You must use a unique CallerReference string every time you create a hosted zone. CallerReference can be any unique string; you might choose to use a string that identifies your project, such as DNSMigration_01. Valid characters are any Unicode code points that are legal in an XML 1.0 document. The UTF-8 encoding of the value must be less than 128 bytes.

_chzrHostedZoneConfig :: Maybe HostedZoneConfig

A complex type that contains an optional comment about your hosted zone.

chzrName :: Lens' CreateHostedZone Text

chzrHostedZoneConfig :: Lens' CreateHostedZone (Maybe HostedZoneConfig)

chzrCallerReference :: Lens' CreateHostedZone Text

data CreateHostedZoneResponse

Constructors

CreateHostedZoneResponse

Fields

_chzsChangeInfo :: ChangeInfo

A complex type that contains information about the request to create a hosted zone. This includes an ID that you use when you call the GetChange action to get the current status of the change request.

_chzsDelegationSet :: DelegationSet

A complex type that contains name server information.

_chzsHostedZone :: HostedZone

A complex type that contains identifying information about the hosted zone.

_chzsLocation :: Text

The unique URL representing the new hosted zone.

chzsLocation :: Lens' CreateHostedZoneResponse Text

chzsHostedZone :: Lens' CreateHostedZoneResponse HostedZone

chzsDelegationSet :: Lens' CreateHostedZoneResponse DelegationSet

chzsChangeInfo :: Lens' CreateHostedZoneResponse ChangeInfo

data DeleteHealthCheck

Constructors

DeleteHealthCheck

Fields

_dhcrHealthCheckId :: Text

The ID of the health check to delete.

dhcrHealthCheckId :: Iso' DeleteHealthCheck Text

data DeleteHealthCheckResponse

Constructors

DeleteHealthCheckResponse

data DeleteHostedZone

Constructors

DeleteHostedZone

Fields

_dhzrId :: Text

The ID of the request. Include this ID in a call to GetChange to track when the change has propagated to all Route 53 DNS servers.

dhzrId :: Iso' DeleteHostedZone Text

data DeleteHostedZoneResponse

Constructors

DeleteHostedZoneResponse

Fields

_dhzsChangeInfo :: ChangeInfo

A complex type that contains the ID, the status, and the date and time of your delete request.

dhzsChangeInfo :: Iso' DeleteHostedZoneResponse ChangeInfo

data GetChange

Constructors

GetChange

Fields

_gcrId :: Text

The ID of the change batch request. The value that you specify here is the value that ChangeResourceRecordSets returned in the Id element when you submitted the request.

gcrId :: Iso' GetChange Text

data GetChangeResponse

Constructors

GetChangeResponse

Fields

_gcsChangeInfo :: ChangeInfo

A complex type that contains information about the specified change batch, including the change batch ID, the status of the change, and the date and time of the request.

gcsChangeInfo :: Iso' GetChangeResponse ChangeInfo

data GetCheckerIpRanges

Constructors

GetCheckerIpRanges

data GetCheckerIpRangesResponse

Constructors

GetCheckerIpRangesResponse

Fields

_gcirsCheckerIpRanges :: [Text]

A complex type that contains sorted list of IP ranges in CIDR format for Amazon Route 53 health checkers.

gcirsCheckerIpRanges :: Iso' GetCheckerIpRangesResponse [Text]

getGeoLocation :: GetGeoLocation

Minimum specification for a GetGeoLocation request.

data GetGeoLocation

Constructors

GetGeoLocation

Fields

_gglrContinentCode :: Maybe Text

The code for a continent geo location. Note: only continent locations have a continent code. Valid values: AF | AN | AS | EU | OC | NA | SA Constraint: Specifying ContinentCode with either CountryCode or SubdivisionCode returns an InvalidInput error.

_gglrCountryCode :: Maybe Text

The code for a country geo location. The default location uses * for the country code and will match all locations that are not matched by a geo location. The default geo location uses a * for the country code. All other country codes follow the ISO 3166 two-character code.

_gglrSubdivisionCode :: Maybe Text

The code for a country's subdivision (e.g., a province of Canada). A subdivision code is only valid with the appropriate country code. Constraint: Specifying SubdivisionCode without CountryCode returns an InvalidInput error.

gglrSubdivisionCode :: Lens' GetGeoLocation (Maybe Text)

gglrCountryCode :: Lens' GetGeoLocation (Maybe Text)

gglrContinentCode :: Lens' GetGeoLocation (Maybe Text)

data GetGeoLocationResponse

Constructors

GetGeoLocationResponse

Fields

_gglsGeoLocationDetails :: GeoLocationDetails

A complex type that contains the information about the specified geo location.

gglsGeoLocationDetails :: Iso' GetGeoLocationResponse GeoLocationDetails

data GetHealthCheck

Constructors

GetHealthCheck

Fields

_ghcrHealthCheckId :: Text

The ID of the health check to retrieve.

ghcrHealthCheckId :: Iso' GetHealthCheck Text

data GetHealthCheckResponse

Constructors

GetHealthCheckResponse

Fields

_ghcsHealthCheck :: HealthCheck

A complex type that contains the information about the specified health check.

ghcsHealthCheck :: Iso' GetHealthCheckResponse HealthCheck

data GetHealthCheckCount

Constructors

GetHealthCheckCount

data GetHealthCheckCountResponse

Constructors

GetHealthCheckCountResponse

Fields

_ghccsHealthCheckCount :: Integer

The number of health checks associated with the current AWS account.

ghccsHealthCheckCount :: Iso' GetHealthCheckCountResponse Integer

data GetHostedZone

Constructors

GetHostedZone

Fields

_ghzrId :: Text

The ID of the hosted zone for which you want to get a list of the name servers in the delegation set.

ghzrId :: Iso' GetHostedZone Text

data GetHostedZoneResponse

Constructors

GetHostedZoneResponse

Fields

_ghzsDelegationSet :: DelegationSet

A complex type that contains information about the name servers for the specified hosted zone.

_ghzsHostedZone :: HostedZone

A complex type that contains the information about the specified hosted zone.

ghzsHostedZone :: Lens' GetHostedZoneResponse HostedZone

ghzsDelegationSet :: Lens' GetHostedZoneResponse DelegationSet

listGeoLocations :: ListGeoLocations

Minimum specification for a ListGeoLocations request.

data ListGeoLocations

Constructors

ListGeoLocations

Fields

_lglrStartContinentCode :: Maybe Text

The first continent code in the lexicographic ordering of geo locations that you want the ListGeoLocations request to list. For non-continent geo locations, this should be null. Valid values: AF | AN | AS | EU | OC | NA | SA Constraint: Specifying ContinentCode with either CountryCode or SubdivisionCode returns an InvalidInput error.

_lglrStartCountryCode :: Maybe Text

The first country code in the lexicographic ordering of geo locations that you want the ListGeoLocations request to list. The default geo location uses a * for the country code. All other country codes follow the ISO 3166 two-character code.

_lglrStartSubdivisionCode :: Maybe Text

The first subdivision code in the lexicographic ordering of geo locations that you want the ListGeoLocations request to list. Constraint: Specifying SubdivisionCode without CountryCode returns an InvalidInput error.

_lglrMaxItems :: Maybe Text

The maximum number of geo locations you want in the response body.

lglrStartSubdivisionCode :: Lens' ListGeoLocations (Maybe Text)

lglrStartCountryCode :: Lens' ListGeoLocations (Maybe Text)

lglrStartContinentCode :: Lens' ListGeoLocations (Maybe Text)

lglrMaxItems :: Lens' ListGeoLocations (Maybe Text)

data ListGeoLocationsResponse

Constructors

ListGeoLocationsResponse

Fields

_lglsGeoLocationDetailsList :: [GeoLocationDetails]

A complex type that contains information about the geo locations that are returned by the request.

_lglsMaxItems :: Text

The maximum number of records you requested. The maximum value of MaxItems is 100.

_lglsIsTruncated :: Bool

A flag that indicates whether there are more geo locations to be listed. If your results were truncated, you can make a follow-up request for the next page of results by using the values included in the ListGeoLocationsResponse$NextContinentCode, ListGeoLocationsResponse$NextCountryCode and ListGeoLocationsResponse$NextSubdivisionCode elements. Valid Values: true | false.

_lglsNextContinentCode :: Maybe Text

If the results were truncated, the continent code of the next geo location in the list. This element is present only if ListGeoLocationsResponse$IsTruncated is true and the next geo location to list is a continent location.

_lglsNextCountryCode :: Maybe Text

If the results were truncated, the country code of the next geo location in the list. This element is present only if ListGeoLocationsResponse$IsTruncated is true and the next geo location to list is not a continent location.

_lglsNextSubdivisionCode :: Maybe Text

If the results were truncated, the subdivision code of the next geo location in the list. This element is present only if ListGeoLocationsResponse$IsTruncated is true and the next geo location has a subdivision.

lglsNextSubdivisionCode :: Lens' ListGeoLocationsResponse (Maybe Text)

lglsNextCountryCode :: Lens' ListGeoLocationsResponse (Maybe Text)

lglsNextContinentCode :: Lens' ListGeoLocationsResponse (Maybe Text)

lglsMaxItems :: Lens' ListGeoLocationsResponse Text

lglsIsTruncated :: Lens' ListGeoLocationsResponse Bool

lglsGeoLocationDetailsList :: Lens' ListGeoLocationsResponse [GeoLocationDetails]

listHealthChecks :: ListHealthChecks

Minimum specification for a ListHealthChecks request.

data ListHealthChecks

Constructors

ListHealthChecks

Fields

_lhcrMarker :: Maybe Text

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

_lhcrMaxItems :: Maybe Text

Specify the maximum number of health checks to return per page of results.

lhcrMaxItems :: Lens' ListHealthChecks (Maybe Text)

lhcrMarker :: Lens' ListHealthChecks (Maybe Text)

data ListHealthChecksResponse

Constructors

ListHealthChecksResponse

Fields

_lhcsHealthChecks :: [HealthCheck]

A complex type that contains information about the health checks associated with the current AWS account.

_lhcsMarker :: Text

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

_lhcsMaxItems :: Text

The maximum number of health checks to be included in the response body. If the number of health checks associated with this AWS account exceeds MaxItems, the value of ListHealthChecksResponse$IsTruncated in the response is true. Call ListHealthChecks again and specify the value of ListHealthChecksResponse$NextMarker in the ListHostedZonesRequest$Marker element to get the next page of results.

_lhcsIsTruncated :: Bool

A flag indicating whether there are more health checks to be listed. If your results were truncated, you can make a follow-up request for the next page of results by using the Marker element. Valid Values: true | false.

_lhcsNextMarker :: Maybe Text

Indicates where to continue listing health checks. If ListHealthChecksResponse$IsTruncated is true, make another request to ListHealthChecks and include the value of the NextMarker element in the Marker element to get the next page of results.

lhcsNextMarker :: Lens' ListHealthChecksResponse (Maybe Text)

lhcsMaxItems :: Lens' ListHealthChecksResponse Text

lhcsMarker :: Lens' ListHealthChecksResponse Text

lhcsIsTruncated :: Lens' ListHealthChecksResponse Bool

lhcsHealthChecks :: Lens' ListHealthChecksResponse [HealthCheck]

listHostedZones :: ListHostedZones

Minimum specification for a ListHostedZones request.

data ListHostedZones

Constructors

ListHostedZones

Fields

_lhzrMarker :: Maybe Text

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

_lhzrMaxItems :: Maybe Text

Specify the maximum number of hosted zones to return per page of results.

lhzrMaxItems :: Lens' ListHostedZones (Maybe Text)

lhzrMarker :: Lens' ListHostedZones (Maybe Text)

data ListHostedZonesResponse

Constructors

ListHostedZonesResponse

Fields

_lhzsHostedZones :: [HostedZone]

A complex type that contains information about the hosted zones associated with the current AWS account.

_lhzsMarker :: Text

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

_lhzsMaxItems :: Text

The maximum number of hosted zones to be included in the response body. If the number of hosted zones associated with this AWS account exceeds MaxItems, the value of ListHostedZonesResponse$IsTruncated in the response is true. Call ListHostedZones again and specify the value of ListHostedZonesResponse$NextMarker in the ListHostedZonesRequest$Marker element to get the next page of results.

_lhzsIsTruncated :: Bool

A flag indicating whether there are more hosted zones to be listed. If your results were truncated, you can make a follow-up request for the next page of results by using the Marker element. Valid Values: true | false.

_lhzsNextMarker :: Maybe Text

Indicates where to continue listing hosted zones. If ListHostedZonesResponse$IsTruncated is true, make another request to ListHostedZones and include the value of the NextMarker element in the Marker element to get the next page of results.

lhzsNextMarker :: Lens' ListHostedZonesResponse (Maybe Text)

lhzsMaxItems :: Lens' ListHostedZonesResponse Text

lhzsMarker :: Lens' ListHostedZonesResponse Text

lhzsIsTruncated :: Lens' ListHostedZonesResponse Bool

lhzsHostedZones :: Lens' ListHostedZonesResponse [HostedZone]

listResourceRecordSets

Arguments

:: Text
_lrrsrHostedZoneId

-> ListResourceRecordSets

Minimum specification for a ListResourceRecordSets request.

data ListResourceRecordSets

Constructors

ListResourceRecordSets

Fields

_lrrsrHostedZoneId :: Text

The ID of the hosted zone that contains the resource record sets that you want to get.

_lrrsrStartRecordName :: Maybe Text

The first name in the lexicographic ordering of domain names that you want the ListResourceRecordSets request to list.

_lrrsrMaxItems :: Maybe Text

The maximum number of records you want in the response body.

_lrrsrStartRecordType :: Maybe RecordType

The DNS type at which to begin the listing of resource record sets. Valid values: A | AAAA | CNAME | MX | NS | PTR | SOA | SPF | SRV | TXT Values for Weighted Resource Record Sets: A | AAAA | CNAME | TXT Values for Regional Resource Record Sets: A | AAAA | CNAME | TXT Values for Alias Resource Record Sets: A | AAAA Constraint: Specifying type without specifying name returns an InvalidInput error.

_lrrsrStartRecordIdentifier :: Maybe Text

Weighted resource record sets only: If results were truncated for a given DNS name and type, specify the value of ListResourceRecordSetsResponse$NextRecordIdentifier from the previous response to get the next resource record set that has the current DNS name and type.

lrrsrStartRecordType :: Lens' ListResourceRecordSets (Maybe RecordType)

lrrsrStartRecordName :: Lens' ListResourceRecordSets (Maybe Text)

lrrsrStartRecordIdentifier :: Lens' ListResourceRecordSets (Maybe Text)

lrrsrMaxItems :: Lens' ListResourceRecordSets (Maybe Text)

lrrsrHostedZoneId :: Lens' ListResourceRecordSets Text

data ListResourceRecordSetsResponse

Constructors

ListResourceRecordSetsResponse

Fields

_lrrssMaxItems :: Text

The maximum number of records you requested. The maximum value of MaxItems is 100.

_lrrssIsTruncated :: Bool

A flag that indicates whether there are more resource record sets to be listed. If your results were truncated, you can make a follow-up request for the next page of results by using the ListResourceRecordSetsResponse$NextRecordName element. Valid Values: true | false.

_lrrssResourceRecordSets :: [ResourceRecordSet]

A complex type that contains information about the resource record sets that are returned by the request.

_lrrssNextRecordName :: Maybe Text

If the results were truncated, the name of the next record in the list. This element is present only if ListResourceRecordSetsResponse$IsTruncated is true.

_lrrssNextRecordType :: Maybe RecordType

If the results were truncated, the type of the next record in the list. This element is present only if ListResourceRecordSetsResponse$IsTruncated is true.

_lrrssNextRecordIdentifier :: Maybe Text

Weighted resource record sets only: If results were truncated for a given DNS name and type, the value of SetIdentifier for the next resource record set that has the current DNS name and type.

lrrssResourceRecordSets :: Lens' ListResourceRecordSetsResponse [ResourceRecordSet]

lrrssNextRecordType :: Lens' ListResourceRecordSetsResponse (Maybe RecordType)

lrrssNextRecordName :: Lens' ListResourceRecordSetsResponse (Maybe Text)

lrrssNextRecordIdentifier :: Lens' ListResourceRecordSetsResponse (Maybe Text)

lrrssMaxItems :: Lens' ListResourceRecordSetsResponse Text

lrrssIsTruncated :: Lens' ListResourceRecordSetsResponse Bool

data ListTagsForResource

Constructors

ListTagsForResource

Fields

_ltfrrResourceId :: Text

The ID of the resource for which you want to retrieve tags.

_ltfrrResourceType :: TagResourceType

The type of the resource. The resource type for health checks is healthcheck.

ltfrrResourceType :: Lens' ListTagsForResource TagResourceType

ltfrrResourceId :: Lens' ListTagsForResource Text

data ListTagsForResourceResponse

Constructors

ListTagsForResourceResponse

Fields

_ltfrsResourceTagSet :: ResourceTagSet

A ResourceTagSet containing tags associated with the specified resource.

ltfrsResourceTagSet :: Iso' ListTagsForResourceResponse ResourceTagSet

data ListTagsForResources

Constructors

ListTagsForResources

Fields

_ltfrtResourceIds :: [Text]

A complex type that contains the ResourceId element for each resource for which you want to get a list of tags.

_ltfrtResourceType :: TagResourceType

The type of the resources. The resource type for health checks is healthcheck.

ltfrtResourceType :: Lens' ListTagsForResources TagResourceType

ltfrtResourceIds :: Lens' ListTagsForResources [Text]

data ListTagsForResourcesResponse

Constructors

ListTagsForResourcesResponse

Fields

_ltfruResourceTagSets :: [ResourceTagSet]

A list of ResourceTagSets containing tags associated with the specified resources.

ltfruResourceTagSets :: Iso' ListTagsForResourcesResponse [ResourceTagSet]

data Route53

Supported version (2013-04-01) of the Amazon Route 53 service.

xmlOptions :: Tagged a XMLOptions

data ChangeAction

The action to perform. Valid values: CREATE | DELETE | UPSERT.

Constructors

ChangeActionCreate
CREATE

ChangeActionDelete
DELETE

ChangeActionUpsert
UPSERT

data ChangeStatus

The current state of the request. PENDING indicates that this request has not yet been applied to all Amazon Route 53 DNS servers. Valid Values: PENDING | INSYNC.

Constructors

ChangeStatusInsync
INSYNC

ChangeStatusPending
PENDING

data HealthCheckType

The type of health check to be performed. Currently supported types are TCP, HTTP, HTTPS, HTTP_STR_MATCH, and HTTPS_STR_MATCH.

Constructors

HealthCheckTypeHttp
HTTP

HealthCheckTypeHttpStrMatch
HTTP_STR_MATCH

HealthCheckTypeHttps
HTTPS

HealthCheckTypeHttpsStrMatch
HTTPS_STR_MATCH

HealthCheckTypeTcp
TCP

data ResourceRecordSetFailover

Failover resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that indicates whether the current resource record set is a primary or secondary resource record set. A failover set may contain at most one resource record set marked as primary and one resource record set marked as secondary. A resource record set marked as primary will be returned if any of the following are true: (1) an associated health check is passing, (2) if the resource record set is an alias with the evaluate target health and at least one target resource record set is healthy, (3) both the primary and secondary resource record set are failing health checks or (4) there is no secondary resource record set. A secondary resource record set will be returned if: (1) the primary is failing a health check and either the secondary is passing a health check or has no associated health check, or (2) there is no primary resource record set. Valid values: PRIMARY | SECONDARY.

Constructors

ResourceRecordSetFailoverPrimary
PRIMARY

ResourceRecordSetFailoverSecondary
SECONDARY

data TagResourceType

The type of the resource. The resource type for health checks is healthcheck.

Constructors

TagResourceTypeHealthcheck
healthcheck

newtype DelegationSet

A complex type that contains name server information.

Constructors

DelegationSet

Fields

_dsNameServers :: [Text]

A complex type that contains the authoritative name servers for the hosted zone. Use the method provided by your domain registrar to add an NS record to your domain for each NameServer that is assigned to your hosted zone.

newtype HostedZoneConfig

A complex type that contains an optional comment about your hosted zone.

Constructors

HostedZoneConfig

Fields

_hzcComment :: Maybe Text

An optional comment about your hosted zone. If you don't want to specify a comment, you can omit the HostedZoneConfig and Comment elements from the XML document.

newtype ResourceRecord

A complex type that contains the value of the Value element for the current resource record set.

Constructors

ResourceRecord

Fields

_rsValue :: Text

The value of the Value element for the current resource record set.

data AliasTarget

Alias resource record sets only: Information about the AWS resource to which you are redirecting traffic.

Constructors

AliasTarget

Fields

_atHostedZoneId :: Text

Alias resource record sets only: The value of the hosted zone ID for the AWS resource. For more information and an example, see Creating Alias Resource Record Sets in the Amazon Route 53 Developer Guide.

_atEvaluateTargetHealth :: Bool

Alias resource record sets only: A boolean value that indicates whether this Resource Record Set should respect the health status of any health checks associated with the ALIAS target record which it is linked to. For more information and an example, see Creating Alias Resource Record Sets in the Amazon Route 53 Developer Guide.

_atDNSName :: Text

Alias resource record sets only: The external DNS name associated with the AWS Resource. For more information and an example, see Creating Alias Resource Record Sets in the Amazon Route 53 Developer Guide.

data Change

A complex type that contains the information for each change in a change batch request.

Constructors

Change

Fields

_dAction :: ChangeAction

The action to perform. Valid values: CREATE | DELETE | UPSERT.

_dResourceRecordSet :: ResourceRecordSet

Information about the resource record set to create or delete.

data ChangeBatch

A complex type that contains an optional comment and the Changes element.

Constructors

ChangeBatch

Fields

_cbChanges :: [Change]

A complex type that contains one Change element for each resource record set that you want to create or delete.

_cbComment :: Maybe Text

Optional: Any comments you want to include about a change batch request.

data ChangeInfo

A complex type that contains information about the specified change batch, including the change batch ID, the status of the change, and the date and time of the request.

Constructors

ChangeInfo

Fields

_ciStatus :: ChangeStatus

The current state of the request. PENDING indicates that this request has not yet been applied to all Amazon Route 53 DNS servers. Valid Values: PENDING | INSYNC.

_ciSubmittedAt :: ISO8601

The date and time the change was submitted, in the format YYYY-MM-DDThh:mm:ssZ, as specified in the ISO 8601 standard (for example, 2009-11-19T19:37:58Z). The Z after the time indicates that the time is listed in Coordinated Universal Time (UTC), which is synonymous with Greenwich Mean Time in this context.

_ciId :: Text

The ID of the request. Use this ID to track when the change has completed across all Amazon Route 53 DNS servers.

_ciComment :: Maybe Text

A complex type that describes change information about changes made to your hosted zone. This element contains an ID that you use when performing a GetChange action to get detailed information about the change.

data GeoLocation

Geo location resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that specifies the geo location for the current resource record set.

Constructors

GeoLocation

Fields

_glSubdivisionCode :: Maybe Text

The code for a country's subdivision (e.g., a province of Canada). A subdivision code is only valid with the appropriate country code. Constraint: Specifying SubdivisionCode without CountryCode returns an InvalidInput error.

_glCountryCode :: Maybe Text

The code for a country geo location. The default location uses * for the country code and will match all locations that are not matched by a geo location. The default geo location uses a * for the country code. All other country codes follow the ISO 3166 two-character code.

_glContinentCode :: Maybe Text

The code for a continent geo location. Note: only continent locations have a continent code. Valid values: AF | AN | AS | EU | OC | NA | SA Constraint: Specifying ContinentCode with either CountryCode or SubdivisionCode returns an InvalidInput error.

data GeoLocationDetails

A complex type that contains information about a GeoLocation.

Constructors

GeoLocationDetails

Fields

_gldSubdivisionName :: Maybe Text

The name of the subdivision. This element is only present if SubdivisionCode is also present.

_gldSubdivisionCode :: Maybe Text

The code for a country's subdivision (e.g., a province of Canada). A subdivision code is only valid with the appropriate country code.

_gldCountryName :: Maybe Text

The name of the country. This element is only present if CountryCode is also present.

_gldCountryCode :: Maybe Text

The code for a country geo location. The default location uses * for the country code and will match all locations that are not matched by a geo location. The default geo location uses a * for the country code. All other country codes follow the ISO 3166 two-character code.

_gldContinentCode :: Maybe Text

The code for a continent geo location. Note: only continent locations have a continent code.

_gldContinentName :: Maybe Text

The name of the continent. This element is only present if ContinentCode is also present.

data HealthCheck

A complex type that contains identifying information about the health check.

Constructors

HealthCheck

Fields

_hcHealthCheckConfig :: HealthCheckConfig

A complex type that contains the health check configuration.

_hcId :: Text

The ID of the specified health check.

_hcHealthCheckVersion :: Integer

The version of the health check. You can optionally pass this value in a call to UpdateHealthCheck to prevent overwriting another change to the health check.

_hcCallerReference :: Text

A unique string that identifies the request to create the health check.

data HealthCheckConfig

A complex type that contains the health check configuration.

Constructors

HealthCheckConfig

Fields

_hccFailureThreshold :: Maybe Integer

The number of consecutive health checks that an endpoint must pass or fail for Route 53 to change the current status of the endpoint from unhealthy to healthy or vice versa. Valid values are integers between 1 and 10. For more information, see "How Amazon Route 53 Determines Whether an Endpoint Is Healthy" in the Amazon Route 53 Developer Guide.

_hccIPAddress :: Maybe Text

IP Address of the instance being checked.

_hccSearchString :: Maybe Text

A string to search for in the body of a health check response. Required for HTTP_STR_MATCH and HTTPS_STR_MATCH health checks.

_hccResourcePath :: Maybe Text

Path to ping on the instance to check the health. Required for HTTP, HTTPS, HTTP_STR_MATCH, and HTTPS_STR_MATCH health checks, HTTP request is issued to the instance on the given port and path.

_hccType :: HealthCheckType

The type of health check to be performed. Currently supported types are TCP, HTTP, HTTPS, HTTP_STR_MATCH, and HTTPS_STR_MATCH.

_hccFullyQualifiedDomainName :: Maybe Text

Fully qualified domain name of the instance to be health checked.

_hccRequestInterval :: Maybe Integer

The number of seconds between the time that Route 53 gets a response from your endpoint and the time that it sends the next health-check request. Each Route 53 health checker makes requests at this interval. Valid values are 10 and 30. The default value is 30.

_hccPort :: Maybe Integer

Port on which connection will be opened to the instance to health check. For HTTP and HTTP_STR_MATCH this defaults to 80 if the port is not specified. For HTTPS and HTTPS_STR_MATCH this defaults to 443 if the port is not specified.

data HostedZone

A complex type that contains identifying information about the hosted zone.

Constructors

HostedZone

Fields

_hzConfig :: Maybe HostedZoneConfig

A complex type that contains the Comment element.

_hzName :: Text

The name of the domain. This must be a fully-specified domain, for example, www.example.com. The trailing dot is optional; Route 53 assumes that the domain name is fully qualified. This means that Route 53 treats www.example.com (without a trailing dot) and www.example.com. (with a trailing dot) as identical. This is the name you have registered with your DNS registrar. You should ask your registrar to change the authoritative name servers for your domain to the set of NameServers elements returned in DelegationSet.

_hzId :: Text

The ID of the specified hosted zone.

_hzResourceRecordSetCount :: Maybe Integer

Total number of resource record sets in the hosted zone.

_hzCallerReference :: Text

A unique string that identifies the request to create the hosted zone.

data ResourceRecordSet

Information about the resource record set to create or delete.

Constructors

ResourceRecordSet

Fields

_rrsTTL :: Maybe Integer

The cache time to live for the current resource record set.

_rrsResourceRecords :: Maybe [ResourceRecord]

A complex type that contains the resource records for the current resource record set.

_rrsAliasTarget :: Maybe AliasTarget

Alias resource record sets only: Information about the AWS resource to which you are redirecting traffic.

_rrsWeight :: Maybe Integer

Weighted resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that determines what portion of traffic for the current resource record set is routed to the associated location.

_rrsSetIdentifier :: Maybe Text

Weighted, Latency, Geo, and Failover resource record sets only: An identifier that differentiates among multiple resource record sets that have the same combination of DNS name and type.

_rrsFailover :: Maybe ResourceRecordSetFailover

Failover resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that indicates whether the current resource record set is a primary or secondary resource record set. A failover set may contain at most one resource record set marked as primary and one resource record set marked as secondary. A resource record set marked as primary will be returned if any of the following are true: (1) an associated health check is passing, (2) if the resource record set is an alias with the evaluate target health and at least one target resource record set is healthy, (3) both the primary and secondary resource record set are failing health checks or (4) there is no secondary resource record set. A secondary resource record set will be returned if: (1) the primary is failing a health check and either the secondary is passing a health check or has no associated health check, or (2) there is no primary resource record set. Valid values: PRIMARY | SECONDARY.

_rrsName :: Text

The domain name of the current resource record set.

_rrsHealthCheckId :: Maybe Text

Health Check resource record sets only, not required for alias resource record sets: An identifier that is used to identify health check associated with the resource record set.

_rrsRegion :: Maybe Region

Latency-based resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that specifies the AWS region for the current resource record set.

_rrsType :: RecordType

The type of the current resource record set.

_rrsGeoLocation :: Maybe GeoLocation

Geo location resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that specifies the geo location for the current resource record set.

data ResourceTagSet

A ResourceTagSet containing tags associated with the specified resource.

Constructors

ResourceTagSet

Fields

_rtsResourceId :: Maybe Text

The ID for the specified resource.

_rtsResourceType :: Maybe TagResourceType

The type of the resource. The resource type for health checks is healthcheck.

_rtsTags :: Maybe [Tag]

The tags associated with the specified resource.

data Tag

A single tag containing a key and value.

Constructors

Tag

Fields

_tValue :: Maybe Text

The value for a Tag.

_tKey :: Maybe Text

The key for a Tag.

dsNameServers :: Iso' DelegationSet [Text]

hzcComment :: Iso' HostedZoneConfig (Maybe Text)

rsValue :: Iso' ResourceRecord Text

atHostedZoneId :: Lens' AliasTarget Text

atEvaluateTargetHealth :: Lens' AliasTarget Bool

atDNSName :: Lens' AliasTarget Text

dResourceRecordSet :: Lens' Change ResourceRecordSet

dAction :: Lens' Change ChangeAction

cbComment :: Lens' ChangeBatch (Maybe Text)

cbChanges :: Lens' ChangeBatch [Change]

ciSubmittedAt :: Lens' ChangeInfo ISO8601

ciStatus :: Lens' ChangeInfo ChangeStatus

ciId :: Lens' ChangeInfo Text

ciComment :: Lens' ChangeInfo (Maybe Text)

glSubdivisionCode :: Lens' GeoLocation (Maybe Text)

glCountryCode :: Lens' GeoLocation (Maybe Text)

glContinentCode :: Lens' GeoLocation (Maybe Text)

gldSubdivisionName :: Lens' GeoLocationDetails (Maybe Text)

gldSubdivisionCode :: Lens' GeoLocationDetails (Maybe Text)

gldCountryName :: Lens' GeoLocationDetails (Maybe Text)

gldCountryCode :: Lens' GeoLocationDetails (Maybe Text)

gldContinentName :: Lens' GeoLocationDetails (Maybe Text)

gldContinentCode :: Lens' GeoLocationDetails (Maybe Text)

hcId :: Lens' HealthCheck Text

hcHealthCheckVersion :: Lens' HealthCheck Integer

hcHealthCheckConfig :: Lens' HealthCheck HealthCheckConfig

hcCallerReference :: Lens' HealthCheck Text

hccType :: Lens' HealthCheckConfig HealthCheckType

hccSearchString :: Lens' HealthCheckConfig (Maybe Text)

hccResourcePath :: Lens' HealthCheckConfig (Maybe Text)

hccRequestInterval :: Lens' HealthCheckConfig (Maybe Integer)

hccPort :: Lens' HealthCheckConfig (Maybe Integer)

hccIPAddress :: Lens' HealthCheckConfig (Maybe Text)

hccFullyQualifiedDomainName :: Lens' HealthCheckConfig (Maybe Text)

hccFailureThreshold :: Lens' HealthCheckConfig (Maybe Integer)

hzResourceRecordSetCount :: Lens' HostedZone (Maybe Integer)

hzName :: Lens' HostedZone Text

hzId :: Lens' HostedZone Text

hzConfig :: Lens' HostedZone (Maybe HostedZoneConfig)

hzCallerReference :: Lens' HostedZone Text

rrsWeight :: Lens' ResourceRecordSet (Maybe Integer)

rrsType :: Lens' ResourceRecordSet RecordType

rrsTTL :: Lens' ResourceRecordSet (Maybe Integer)

rrsSetIdentifier :: Lens' ResourceRecordSet (Maybe Text)

rrsResourceRecords :: Lens' ResourceRecordSet (Maybe [ResourceRecord])

rrsRegion :: Lens' ResourceRecordSet (Maybe Region)

rrsName :: Lens' ResourceRecordSet Text

rrsHealthCheckId :: Lens' ResourceRecordSet (Maybe Text)

rrsGeoLocation :: Lens' ResourceRecordSet (Maybe GeoLocation)

rrsFailover :: Lens' ResourceRecordSet (Maybe ResourceRecordSetFailover)

rrsAliasTarget :: Lens' ResourceRecordSet (Maybe AliasTarget)

rtsTags :: Lens' ResourceTagSet (Maybe [Tag])

rtsResourceType :: Lens' ResourceTagSet (Maybe TagResourceType)

rtsResourceId :: Lens' ResourceTagSet (Maybe Text)

tValue :: Lens' Tag (Maybe Text)

tKey :: Lens' Tag (Maybe Text)

updateHealthCheck

Arguments

:: Text
_uhcrHealthCheckId

-> UpdateHealthCheck

Minimum specification for a UpdateHealthCheck request.

data UpdateHealthCheck

Constructors

UpdateHealthCheck

Fields

_uhcrHealthCheckId :: Text

The ID of the health check to update.

_uhcrFailureThreshold :: Maybe Integer

The number of consecutive health checks that an endpoint must pass or fail for Route 53 to change the current status of the endpoint from unhealthy to healthy or vice versa. Valid values are integers between 1 and 10. For more information, see "How Amazon Route 53 Determines Whether an Endpoint Is Healthy" in the Amazon Route 53 Developer Guide. Specify this value only if you want to change it.

_uhcrFullyQualifiedDomainName :: Maybe Text

Fully qualified domain name of the instance to be health checked. Specify this value only if you want to change it.

_uhcrHealthCheckVersion :: Maybe Integer

Optional. When you specify a health check version, Route 53 compares this value with the current value in the health check, which prevents you from updating the health check when the versions don't match. Using HealthCheckVersion lets you prevent overwriting another change to the health check.

_uhcrIPAddress :: Maybe Text

The IP address of the resource that you want to check. Specify this value only if you want to change it.

_uhcrPort :: Maybe Integer

The port on which you want Route 53 to open a connection to perform health checks. Specify this value only if you want to change it.

_uhcrResourcePath :: Maybe Text

The path that you want Amazon Route 53 to request when performing health checks. The path can be any value for which your endpoint will return an HTTP status code of 2xx or 3xx when the endpoint is healthy, for example the file docsroute53-health-check.html. Specify this value only if you want to change it.

_uhcrSearchString :: Maybe Text

If the value of Type is HTTP_STR_MATCH or HTTP_STR_MATCH, the string that you want Route 53 to search for in the response body from the specified resource. If the string appears in the response body, Route 53 considers the resource healthy. Specify this value only if you want to change it.

uhcrSearchString :: Lens' UpdateHealthCheck (Maybe Text)

uhcrResourcePath :: Lens' UpdateHealthCheck (Maybe Text)

uhcrPort :: Lens' UpdateHealthCheck (Maybe Integer)

uhcrIPAddress :: Lens' UpdateHealthCheck (Maybe Text)

uhcrHealthCheckVersion :: Lens' UpdateHealthCheck (Maybe Integer)

uhcrHealthCheckId :: Lens' UpdateHealthCheck Text

uhcrFullyQualifiedDomainName :: Lens' UpdateHealthCheck (Maybe Text)

uhcrFailureThreshold :: Lens' UpdateHealthCheck (Maybe Integer)

data UpdateHealthCheckResponse

Constructors

UpdateHealthCheckResponse

Fields

_uhcsHealthCheck :: HealthCheck

A complex type that contains identifying information about the health check.

uhcsHealthCheck :: Iso' UpdateHealthCheckResponse HealthCheck