Network.AWS.Route53.V2013_04_01.ChangeTagsForResource

Documentation

changeTagsForResource

Arguments

:: Text
_ctfrrResourceId

-> TagResourceType
_ctfrrResourceType

-> ChangeTagsForResource

Minimum specification for a ChangeTagsForResource request.

data ChangeTagsForResource

Constructors

ChangeTagsForResource

Fields

_ctfrrResourceId :: Text

The ID of the resource for which you want to add, change, or delete tags.

_ctfrrResourceType :: TagResourceType

The type of the resource. The resource type for health checks is healthcheck.

_ctfrrRemoveTagKeys :: Maybe [Text]

A list of Tag keys that you want to remove from the specified resource.

_ctfrrAddTags :: Maybe [Tag]

A complex type that contains a list of Tag elements. Each Tag element identifies a tag that you want to add or update for the specified resource.

ctfrrResourceType :: Lens' ChangeTagsForResource TagResourceType

ctfrrResourceId :: Lens' ChangeTagsForResource Text

ctfrrRemoveTagKeys :: Lens' ChangeTagsForResource (Maybe [Text])

ctfrrAddTags :: Lens' ChangeTagsForResource (Maybe [Tag])

data ChangeTagsForResourceResponse

Constructors

ChangeTagsForResourceResponse