Network.AWS.Route53.V2013_04_01.DeleteHostedZone

Description

This action deletes a hosted zone. To delete a hosted zone, send a DELETE request to the 2013-04-01hostedzonehosted zone ID resource. For more information about deleting a hosted zone, see Deleting a Hosted Zone in the Amazon Route 53 Developer Guide. You can delete a hosted zone only if there are no resource record sets other than the default SOA record and NS resource record sets. If your hosted zone contains other resource record sets, you must delete them before you can delete your hosted zone. If you try to delete a hosted zone that contains other resource record sets, Route 53 will deny your request with a HostedZoneNotEmpty error. For information about deleting records from your hosted zone, see ChangeResourceRecordSets.

Request

data DeleteHostedZone

A complex type containing the response information for the delete request.

Request constructor

mkDeleteHostedZone

Arguments

:: Text
dhzId

-> DeleteHostedZone

Smart constructor for the minimum required parameters to construct a valid DeleteHostedZone request.

Request lenses

dhzId :: Lens' DeleteHostedZone Text

The ID of the request. Include this ID in a call to GetChange to track when the change has propagated to all Route 53 DNS servers.

Response

data DeleteHostedZoneResponse

A complex type containing the response information for the request.

Response lenses

dhzrChangeInfo :: Lens' DeleteHostedZoneResponse ChangeInfo

A complex type that contains the ID, the status, and the date and time of your delete request.