Network.AWS.Route53.V2013_04_01.GetChange

Description

This action returns the current status of a change batch request. The status is one of the following values: - PENDING indicates that the changes in this request have not replicated to all Route 53 DNS servers. This is the initial status of all change batch requests. - INSYNC indicates that the changes have replicated to all Amazon Route 53 DNS servers.

Documentation

data GetChange

Constructors

GetChange

Fields

_gcrId :: Text

The ID of the change batch request. The value that you specify here is the value that ChangeResourceRecordSets returned in the Id element when you submitted the request.

gcrId :: Iso' GetChange Text

data GetChangeResponse

Constructors

GetChangeResponse

Fields

_gcsChangeInfo :: ChangeInfo

A complex type that contains information about the specified change batch, including the change batch ID, the status of the change, and the date and time of the request.

gcsChangeInfo :: Iso' GetChangeResponse ChangeInfo