Network.AWS.Route53.V2013_04_01.GetHostedZone

Description

To retrieve the delegation set for a hosted zone, send a GET request to the 2013-04-01hostedzonehosted zone ID resource. The delegation set is the four Route 53 name servers that were assigned to the hosted zone when you created it.

Request

data GetHostedZone

The input for a GetHostedZone request.

Request constructor

mkGetHostedZone

Arguments

:: Text
ghzId

-> GetHostedZone

Smart constructor for the minimum required parameters to construct a valid GetHostedZone request.

Request lenses

ghzId :: Lens' GetHostedZone Text

The ID of the hosted zone for which you want to get a list of the name servers in the delegation set.

Response

data GetHostedZoneResponse

A complex type containing information about the specified hosted zone.

Response lenses

ghzrsHostedZone :: Lens' GetHostedZoneResponse HostedZone

A complex type that contains the information about the specified hosted zone.

ghzrsDelegationSet :: Lens' GetHostedZoneResponse DelegationSet

A complex type that contains information about the name servers for the specified hosted zone.