Network.AWS.Route53.V2013_04_01.ListGeoLocations

Request

data ListGeoLocations

The input for a ListGeoLocations request.

Request constructor

mkListGeoLocations :: ListGeoLocations

Smart constructor for the minimum required parameters to construct a valid ListGeoLocations request.

Request lenses

lglStartContinentCode :: Lens' ListGeoLocations (Maybe Text)

The first continent code in the lexicographic ordering of geo locations that you want the ListGeoLocations request to list. For non-continent geo locations, this should be null. Valid values: AF | AN | AS | EU | OC | NA | SA Constraint: Specifying ContinentCode with either CountryCode or SubdivisionCode returns an InvalidInput error.

lglStartCountryCode :: Lens' ListGeoLocations (Maybe Text)

The first country code in the lexicographic ordering of geo locations that you want the ListGeoLocations request to list. The default geo location uses a * for the country code. All other country codes follow the ISO 3166 two-character code.

lglStartSubdivisionCode :: Lens' ListGeoLocations (Maybe Text)

The first subdivision code in the lexicographic ordering of geo locations that you want the ListGeoLocations request to list. Constraint: Specifying SubdivisionCode without CountryCode returns an InvalidInput error.

lglMaxItems :: Lens' ListGeoLocations (Maybe Text)

The maximum number of geo locations you want in the response body.

Response

data ListGeoLocationsResponse

A complex type that contains information about the geo locations that are returned by the request and information about the response.

Response lenses

lglrsGeoLocationDetailsList :: Lens' ListGeoLocationsResponse [GeoLocationDetails]

A complex type that contains information about the geo locations that are returned by the request.

lglrsIsTruncated :: Lens' ListGeoLocationsResponse Bool

A flag that indicates whether there are more geo locations to be listed. If your results were truncated, you can make a follow-up request for the next page of results by using the values included in the ListGeoLocationsResponse$NextContinentCode, ListGeoLocationsResponse$NextCountryCode and ListGeoLocationsResponse$NextSubdivisionCode elements. Valid Values: true | false.

lglrsNextContinentCode :: Lens' ListGeoLocationsResponse (Maybe Text)

If the results were truncated, the continent code of the next geo location in the list. This element is present only if ListGeoLocationsResponse$IsTruncated is true and the next geo location to list is a continent location.

lglrsNextCountryCode :: Lens' ListGeoLocationsResponse (Maybe Text)

If the results were truncated, the country code of the next geo location in the list. This element is present only if ListGeoLocationsResponse$IsTruncated is true and the next geo location to list is not a continent location.

lglrsNextSubdivisionCode :: Lens' ListGeoLocationsResponse (Maybe Text)

If the results were truncated, the subdivision code of the next geo location in the list. This element is present only if ListGeoLocationsResponse$IsTruncated is true and the next geo location has a subdivision.

lglrsMaxItems :: Lens' ListGeoLocationsResponse Text

The maximum number of records you requested. The maximum value of MaxItems is 100.