Network.AWS.Route53.V2013_04_01.ListHostedZones

Description

To retrieve a list of your hosted zones, send a GET request to the 2013-04-01/hostedzone resource. The response to this request includes a HostedZones element with zero, one, or multiple HostedZone child elements. By default, the list of hosted zones is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. You can use the Marker parameter to control the hosted zone that the list begins with. Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Amazon Route 53 returns only the first 100.

Request

data ListHostedZones

To retrieve a list of your hosted zones, send a GET request to the 2013-04-01/hostedzone resource. The response to this request includes a HostedZones element with zero or more HostedZone child elements. By default, the list of hosted zones is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. You can use the Marker parameter to control the hosted zone that the list begins with. For more information about listing hosted zones, see Listing the Hosted Zones for an AWS Account in the Amazon Route 53 Developer Guide. Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Route 53 returns only the first 100.

Request constructor

mkListHostedZones :: ListHostedZones

Smart constructor for the minimum required parameters to construct a valid ListHostedZones request.

Request lenses

lhzMarker :: Lens' ListHostedZones (Maybe Text)

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

lhzMaxItems :: Lens' ListHostedZones (Maybe Text)

Specify the maximum number of hosted zones to return per page of results.

Response

data ListHostedZonesResponse

A complex type that contains the response for the request.

Response constructor

mkListHostedZonesResponse

Arguments

:: [HostedZone]
lhzrHostedZones

-> Text
lhzrMarker

-> Bool
lhzrIsTruncated

-> Text
lhzrMaxItems

-> ListHostedZonesResponse

Smart constructor for the minimum required parameters to construct a valid ListHostedZonesResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

lhzrHostedZones :: Lens' ListHostedZonesResponse [HostedZone]

A complex type that contains information about the hosted zones associated with the current AWS account.

lhzrMarker :: Lens' ListHostedZonesResponse Text

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

lhzrIsTruncated :: Lens' ListHostedZonesResponse Bool

A flag indicating whether there are more hosted zones to be listed. If your results were truncated, you can make a follow-up request for the next page of results by using the Marker element. Valid Values: true | false.

lhzrNextMarker :: Lens' ListHostedZonesResponse (Maybe Text)

Indicates where to continue listing hosted zones. If ListHostedZonesResponse$IsTruncated is true, make another request to ListHostedZones and include the value of the NextMarker element in the Marker element to get the next page of results.

lhzrMaxItems :: Lens' ListHostedZonesResponse Text

The maximum number of hosted zones to be included in the response body. If the number of hosted zones associated with this AWS account exceeds MaxItems, the value of ListHostedZonesResponse$IsTruncated in the response is true. Call ListHostedZones again and specify the value of ListHostedZonesResponse$NextMarker in the ListHostedZonesRequest$Marker element to get the next page of results.