Network.AWS.Route53.V2013_04_01.Types

Description

Amazon Route 53 is a scalable Domain Name System (DNS) web service. It provides secure and reliable routing to your infrastructure that uses Amazon Web Services (AWS) products, such as Amazon Elastic Compute Cloud (Amazon EC2), Elastic Load Balancing, or Amazon Simple Storage Service (Amazon S3). You can also use Amazon Route 53 to route users to your infrastructure outside of AWS.

Service

data Route53

Supported version (2013-04-01) of the Amazon Route 53 service.

XML

xmlOptions :: Tagged a XMLOptions

ChangeAction

data ChangeAction

Constructors

ChangeActionCreate
CREATE

ChangeActionDelete
DELETE

ChangeActionUpsert
UPSERT

ChangeStatus

data ChangeStatus

Constructors

ChangeStatusInsync
INSYNC

ChangeStatusPending
PENDING

Failover

data Failover

Constructors

Primary
PRIMARY

Secondary
SECONDARY

HealthCheckType

data HealthCheckType

Constructors

HealthCheckTypeHttp
HTTP

HealthCheckTypeHttpStrMatch
HTTP_STR_MATCH

HealthCheckTypeHttps
HTTPS

HealthCheckTypeHttpsStrMatch
HTTPS_STR_MATCH

HealthCheckTypeTcp
TCP

TagResourceType

data TagResourceType

Constructors

TagResourceTypeHealthcheck
healthcheck

DelegationSet

data DelegationSet

A complex type that contains name server information.

mkDelegationSet

Arguments

:: List1 Text
dsNameServers

-> DelegationSet

Smart constructor for the minimum required fields to construct a valid DelegationSet data type to populate a request.

dsNameServers :: Lens' DelegationSet (List1 Text)

A complex type that contains the authoritative name servers for the hosted zone. Use the method provided by your domain registrar to add an NS record to your domain for each NameServer that is assigned to your hosted zone.

HostedZoneConfig

data HostedZoneConfig

A complex type that contains an optional comment about your hosted zone.

mkHostedZoneConfig :: HostedZoneConfig

Smart constructor for the minimum required fields to construct a valid HostedZoneConfig data type to populate a request.

hzcComment :: Lens' HostedZoneConfig (Maybe Text)

An optional comment about your hosted zone. If you don't want to specify a comment, you can omit the HostedZoneConfig and Comment elements from the XML document.

ResourceRecord

data ResourceRecord

A complex type that contains the value of the Value element for the current resource record set.

mkResourceRecord

Arguments

:: Text
rrValue

-> ResourceRecord

Smart constructor for the minimum required fields to construct a valid ResourceRecord data type to populate a request.

rrValue :: Lens' ResourceRecord Text

The value of the Value element for the current resource record set.

AliasTarget

data AliasTarget

Alias resource record sets only: Information about the AWS resource to which you are redirecting traffic.

mkAliasTarget

Arguments

:: Text
atHostedZoneId

-> Text
atDNSName

-> Bool
atEvaluateTargetHealth

-> AliasTarget

Smart constructor for the minimum required fields to construct a valid AliasTarget data type to populate a request.

atHostedZoneId :: Lens' AliasTarget Text

Alias resource record sets only: The value of the hosted zone ID for the AWS resource. For more information and an example, see Creating Alias Resource Record Sets in the Amazon Route 53 Developer Guide.

atDNSName :: Lens' AliasTarget Text

Alias resource record sets only: The external DNS name associated with the AWS Resource. For more information and an example, see Creating Alias Resource Record Sets in the Amazon Route 53 Developer Guide.

atEvaluateTargetHealth :: Lens' AliasTarget Bool

Alias resource record sets only: A boolean value that indicates whether this Resource Record Set should respect the health status of any health checks associated with the ALIAS target record which it is linked to. For more information and an example, see Creating Alias Resource Record Sets in the Amazon Route 53 Developer Guide.

Change

data Change

A complex type that contains the information for each change in a change batch request.

mkChange

Arguments

:: ChangeAction
cAction

-> ResourceRecordSet
cResourceRecordSet

-> Change

Smart constructor for the minimum required fields to construct a valid Change data type to populate a request.

cAction :: Lens' Change ChangeAction

The action to perform. Valid values: CREATE | DELETE | UPSERT.

cResourceRecordSet :: Lens' Change ResourceRecordSet

Information about the resource record set to create or delete.

ChangeBatch

data ChangeBatch

A complex type that contains an optional comment and the Changes element.

mkChangeBatch

Arguments

:: List1 Change
cbChanges

-> ChangeBatch

Smart constructor for the minimum required fields to construct a valid ChangeBatch data type to populate a request.

cbComment :: Lens' ChangeBatch (Maybe Text)

Optional: Any comments you want to include about a change batch request.

cbChanges :: Lens' ChangeBatch (List1 Change)

A complex type that contains one Change element for each resource record set that you want to create or delete.

ChangeInfo

data ChangeInfo

A complex type that contains information about changes made to your hosted zone. This element contains an ID that you use when performing a GetChange action to get detailed information about the change.

mkChangeInfo

Arguments

:: Text
ciId

-> ChangeStatus
ciStatus

-> ISO8601
ciSubmittedAt

-> ChangeInfo

Smart constructor for the minimum required fields to construct a valid ChangeInfo data type to populate a request.

ciId :: Lens' ChangeInfo Text

The ID of the request. Use this ID to track when the change has completed across all Amazon Route 53 DNS servers.

ciStatus :: Lens' ChangeInfo ChangeStatus

The current state of the request. PENDING indicates that this request has not yet been applied to all Amazon Route 53 DNS servers. Valid Values: PENDING | INSYNC.

ciSubmittedAt :: Lens' ChangeInfo ISO8601

The date and time the change was submitted, in the format YYYY-MM-DDThh:mm:ssZ, as specified in the ISO 8601 standard (for example, 2009-11-19T19:37:58Z). The Z after the time indicates that the time is listed in Coordinated Universal Time (UTC), which is synonymous with Greenwich Mean Time in this context.

ciComment :: Lens' ChangeInfo (Maybe Text)

A complex type that describes change information about changes made to your hosted zone. This element contains an ID that you use when performing a GetChange action to get detailed information about the change.

GeoLocation

data GeoLocation

Geo location resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that specifies the geo location for the current resource record set.

mkGeoLocation :: GeoLocation

Smart constructor for the minimum required fields to construct a valid GeoLocation data type to populate a request.

glContinentCode :: Lens' GeoLocation (Maybe Text)

The code for a continent geo location. Note: only continent locations have a continent code. Valid values: AF | AN | AS | EU | OC | NA | SA Constraint: Specifying ContinentCode with either CountryCode or SubdivisionCode returns an InvalidInput error.

glCountryCode :: Lens' GeoLocation (Maybe Text)

The code for a country geo location. The default location uses * for the country code and will match all locations that are not matched by a geo location. The default geo location uses a * for the country code. All other country codes follow the ISO 3166 two-character code.

glSubdivisionCode :: Lens' GeoLocation (Maybe Text)

The code for a country's subdivision (e.g., a province of Canada). A subdivision code is only valid with the appropriate country code. Constraint: Specifying SubdivisionCode without CountryCode returns an InvalidInput error.

GeoLocationDetails

data GeoLocationDetails

A complex type that contains the information about the specified geo location.

mkGeoLocationDetails :: GeoLocationDetails

Smart constructor for the minimum required fields to construct a valid GeoLocationDetails data type to populate a request.

gldContinentCode :: Lens' GeoLocationDetails (Maybe Text)

The code for a continent geo location. Note: only continent locations have a continent code.

gldContinentName :: Lens' GeoLocationDetails (Maybe Text)

The name of the continent. This element is only present if ContinentCode is also present.

gldCountryCode :: Lens' GeoLocationDetails (Maybe Text)

The code for a country geo location. The default location uses * for the country code and will match all locations that are not matched by a geo location. The default geo location uses a * for the country code. All other country codes follow the ISO 3166 two-character code.

gldCountryName :: Lens' GeoLocationDetails (Maybe Text)

The name of the country. This element is only present if CountryCode is also present.

gldSubdivisionCode :: Lens' GeoLocationDetails (Maybe Text)

The code for a country's subdivision (e.g., a province of Canada). A subdivision code is only valid with the appropriate country code.

gldSubdivisionName :: Lens' GeoLocationDetails (Maybe Text)

The name of the subdivision. This element is only present if SubdivisionCode is also present.

HealthCheck

data HealthCheck

A complex type that contains identifying information about the health check.

mkHealthCheck

Arguments

:: Text
hcId

-> Text
hcCallerReference

-> HealthCheckConfig
hcHealthCheckConfig

-> Integer
hcHealthCheckVersion

-> HealthCheck

Smart constructor for the minimum required fields to construct a valid HealthCheck data type to populate a request.

hcId :: Lens' HealthCheck Text

The ID of the specified health check.

hcCallerReference :: Lens' HealthCheck Text

A unique string that identifies the request to create the health check.

hcHealthCheckConfig :: Lens' HealthCheck HealthCheckConfig

A complex type that contains the health check configuration.

hcHealthCheckVersion :: Lens' HealthCheck Integer

The version of the health check. You can optionally pass this value in a call to UpdateHealthCheck to prevent overwriting another change to the health check.

HealthCheckConfig

data HealthCheckConfig

A complex type that contains health check configuration.

mkHealthCheckConfig

Arguments

:: HealthCheckType
hccType

-> HealthCheckConfig

Smart constructor for the minimum required fields to construct a valid HealthCheckConfig data type to populate a request.

hccIPAddress :: Lens' HealthCheckConfig (Maybe Text)

IP Address of the instance being checked.

hccPort :: Lens' HealthCheckConfig (Maybe Integer)

Port on which connection will be opened to the instance to health check. For HTTP and HTTP_STR_MATCH this defaults to 80 if the port is not specified. For HTTPS and HTTPS_STR_MATCH this defaults to 443 if the port is not specified.

hccType :: Lens' HealthCheckConfig HealthCheckType

The type of health check to be performed. Currently supported types are TCP, HTTP, HTTPS, HTTP_STR_MATCH, and HTTPS_STR_MATCH.

hccResourcePath :: Lens' HealthCheckConfig (Maybe Text)

Path to ping on the instance to check the health. Required for HTTP, HTTPS, HTTP_STR_MATCH, and HTTPS_STR_MATCH health checks, HTTP request is issued to the instance on the given port and path.

hccFullyQualifiedDomainName :: Lens' HealthCheckConfig (Maybe Text)

Fully qualified domain name of the instance to be health checked.

hccSearchString :: Lens' HealthCheckConfig (Maybe Text)

A string to search for in the body of a health check response. Required for HTTP_STR_MATCH and HTTPS_STR_MATCH health checks.

hccRequestInterval :: Lens' HealthCheckConfig (Maybe Integer)

The number of seconds between the time that Route 53 gets a response from your endpoint and the time that it sends the next health-check request. Each Route 53 health checker makes requests at this interval. Valid values are 10 and 30. The default value is 30.

hccFailureThreshold :: Lens' HealthCheckConfig (Maybe Integer)

The number of consecutive health checks that an endpoint must pass or fail for Route 53 to change the current status of the endpoint from unhealthy to healthy or vice versa. Valid values are integers between 1 and 10. For more information, see "How Amazon Route 53 Determines Whether an Endpoint Is Healthy" in the Amazon Route 53 Developer Guide.

HostedZone

data HostedZone

A complex type that contains identifying information about the hosted zone.

mkHostedZone

Arguments

:: Text
hzId

-> Text
hzName

-> Text
hzCallerReference

-> HostedZone

Smart constructor for the minimum required fields to construct a valid HostedZone data type to populate a request.

hzId :: Lens' HostedZone Text

The ID of the specified hosted zone.

hzName :: Lens' HostedZone Text

The name of the domain. This must be a fully-specified domain, for example, www.example.com. The trailing dot is optional; Route 53 assumes that the domain name is fully qualified. This means that Route 53 treats www.example.com (without a trailing dot) and www.example.com. (with a trailing dot) as identical. This is the name you have registered with your DNS registrar. You should ask your registrar to change the authoritative name servers for your domain to the set of NameServers elements returned in DelegationSet.

hzCallerReference :: Lens' HostedZone Text

A unique string that identifies the request to create the hosted zone.

hzConfig :: Lens' HostedZone (Maybe HostedZoneConfig)

A complex type that contains the Comment element.

hzResourceRecordSetCount :: Lens' HostedZone (Maybe Integer)

Total number of resource record sets in the hosted zone.

ResourceRecordSet

data ResourceRecordSet

Information about the resource record set to create or delete.

mkResourceRecordSet

Arguments

:: Text
rrsName

-> RecordType
rrsType

-> List1 ResourceRecord
rrsResourceRecords

-> ResourceRecordSet

Smart constructor for the minimum required fields to construct a valid ResourceRecordSet data type to populate a request.

rrsName :: Lens' ResourceRecordSet Text

The domain name of the current resource record set.

rrsType :: Lens' ResourceRecordSet RecordType

The type of the current resource record set.

rrsSetIdentifier :: Lens' ResourceRecordSet (Maybe Text)

Weighted, Latency, Geo, and Failover resource record sets only: An identifier that differentiates among multiple resource record sets that have the same combination of DNS name and type.

rrsWeight :: Lens' ResourceRecordSet (Maybe Integer)

Weighted resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that determines what portion of traffic for the current resource record set is routed to the associated location.

rrsRegion :: Lens' ResourceRecordSet (Maybe Region)

Latency-based resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that specifies the AWS region for the current resource record set.

rrsGeoLocation :: Lens' ResourceRecordSet (Maybe GeoLocation)

Geo location resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that specifies the geo location for the current resource record set.

rrsFailover :: Lens' ResourceRecordSet (Maybe Failover)

Failover resource record sets only: Among resource record sets that have the same combination of DNS name and type, a value that indicates whether the current resource record set is a primary or secondary resource record set. A failover set may contain at most one resource record set marked as primary and one resource record set marked as secondary. A resource record set marked as primary will be returned if any of the following are true: (1) an associated health check is passing, (2) if the resource record set is an alias with the evaluate target health and at least one target resource record set is healthy, (3) both the primary and secondary resource record set are failing health checks or (4) there is no secondary resource record set. A secondary resource record set will be returned if: (1) the primary is failing a health check and either the secondary is passing a health check or has no associated health check, or (2) there is no primary resource record set. Valid values: PRIMARY | SECONDARY.

rrsTTL :: Lens' ResourceRecordSet (Maybe Integer)

The cache time to live for the current resource record set.

rrsResourceRecords :: Lens' ResourceRecordSet (List1 ResourceRecord)

A complex type that contains the resource records for the current resource record set.

rrsAliasTarget :: Lens' ResourceRecordSet (Maybe AliasTarget)

Alias resource record sets only: Information about the AWS resource to which you are redirecting traffic.

rrsHealthCheckId :: Lens' ResourceRecordSet (Maybe Text)

Health Check resource record sets only, not required for alias resource record sets: An identifier that is used to identify health check associated with the resource record set.

ResourceTagSet

data ResourceTagSet

A ResourceTagSet containing tags associated with the specified resource.

mkResourceTagSet :: ResourceTagSet

Smart constructor for the minimum required fields to construct a valid ResourceTagSet data type to populate a request.

rtsResourceType :: Lens' ResourceTagSet (Maybe TagResourceType)

The type of the resource. The resource type for health checks is healthcheck.

rtsResourceId :: Lens' ResourceTagSet (Maybe Text)

The ID for the specified resource.

rtsTags :: Lens' ResourceTagSet (Maybe (List1 Tag))

The tags associated with the specified resource.

Tag

data Tag

A single tag containing a key and value.

mkTag :: Tag

Smart constructor for the minimum required fields to construct a valid Tag data type to populate a request.

tKey :: Lens' Tag (Maybe Text)

The key for a Tag.

tValue :: Lens' Tag (Maybe Text)

The value for a Tag.

Common

data RecordType

Constructors

A

AAAA

CNAME

MX

NS

PTR

SOA

SPF

SRV

TXT