Network.AWS.Route53.V2013_04_01.UpdateHealthCheck

Description

This action updates an existing health check. To update a health check, send a POST request to the 2013-05-27healthcheckhealth check ID resource. The request body must include an XML document with an UpdateHealthCheckRequest element. The response returns an UpdateHealthCheckResponse element, which contains metadata about the health check.

Request

data UpdateHealthCheck

&gt;A complex type that contains information about the request to update a health check.

Request constructor

mkUpdateHealthCheck

Arguments

:: Text
uhcHealthCheckId

-> UpdateHealthCheck

Smart constructor for the minimum required parameters to construct a valid UpdateHealthCheck request.

Request lenses

uhcHealthCheckId :: Lens' UpdateHealthCheck Text

The ID of the health check to update.

uhcHealthCheckVersion :: Lens' UpdateHealthCheck (Maybe Integer)

Optional. When you specify a health check version, Route 53 compares this value with the current value in the health check, which prevents you from updating the health check when the versions don't match. Using HealthCheckVersion lets you prevent overwriting another change to the health check.

uhcIPAddress :: Lens' UpdateHealthCheck (Maybe Text)

The IP address of the resource that you want to check. Specify this value only if you want to change it.

uhcPort :: Lens' UpdateHealthCheck (Maybe Integer)

The port on which you want Route 53 to open a connection to perform health checks. Specify this value only if you want to change it.

uhcResourcePath :: Lens' UpdateHealthCheck (Maybe Text)

The path that you want Amazon Route 53 to request when performing health checks. The path can be any value for which your endpoint will return an HTTP status code of 2xx or 3xx when the endpoint is healthy, for example the file docsroute53-health-check.html. Specify this value only if you want to change it.

uhcFullyQualifiedDomainName :: Lens' UpdateHealthCheck (Maybe Text)

Fully qualified domain name of the instance to be health checked. Specify this value only if you want to change it.

uhcSearchString :: Lens' UpdateHealthCheck (Maybe Text)

If the value of Type is HTTP_STR_MATCH or HTTP_STR_MATCH, the string that you want Route 53 to search for in the response body from the specified resource. If the string appears in the response body, Route 53 considers the resource healthy. Specify this value only if you want to change it.

uhcFailureThreshold :: Lens' UpdateHealthCheck (Maybe Integer)

The number of consecutive health checks that an endpoint must pass or fail for Route 53 to change the current status of the endpoint from unhealthy to healthy or vice versa. Valid values are integers between 1 and 10. For more information, see "How Amazon Route 53 Determines Whether an Endpoint Is Healthy" in the Amazon Route 53 Developer Guide. Specify this value only if you want to change it.

Response

data UpdateHealthCheckResponse

Response lenses

uhcrsHealthCheck :: Lens' UpdateHealthCheckResponse HealthCheck

A complex type that contains identifying information about the health check.