Network.AWS.Route53.AssociateVPCWithHostedZone

Description

Associates an Amazon VPC with a private hosted zone.

Important: To perform the association, the VPC and the private hosted zone must already exist. You can't convert a public hosted zone into a private hosted zone.

Creating a Request

associateVPCWithHostedZone

Arguments

:: ResourceId
avwhzHostedZoneId

-> VPC
avwhzVPC

-> AssociateVPCWithHostedZone

Creates a value of AssociateVPCWithHostedZone with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avwhzComment - Optional: A comment about the association request.

avwhzHostedZoneId - The ID of the private hosted zone that you want to associate an Amazon VPC with. Note that you can't associate a VPC with a hosted zone that doesn't have an existing VPC association.

avwhzVPC - A complex type that contains information about the VPC that you want to associate with a private hosted zone.

data AssociateVPCWithHostedZone

A complex type that contains information about the request to associate a VPC with a private hosted zone.

See: associateVPCWithHostedZone smart constructor.

Request Lenses

avwhzComment :: Lens' AssociateVPCWithHostedZone (Maybe Text)

Optional: A comment about the association request.

avwhzHostedZoneId :: Lens' AssociateVPCWithHostedZone ResourceId

The ID of the private hosted zone that you want to associate an Amazon VPC with. Note that you can't associate a VPC with a hosted zone that doesn't have an existing VPC association.

avwhzVPC :: Lens' AssociateVPCWithHostedZone VPC

A complex type that contains information about the VPC that you want to associate with a private hosted zone.

Destructuring the Response

associateVPCWithHostedZoneResponse

Arguments

:: Int
avwhzrsResponseStatus

-> ChangeInfo
avwhzrsChangeInfo

-> AssociateVPCWithHostedZoneResponse

Creates a value of AssociateVPCWithHostedZoneResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avwhzrsResponseStatus - -- | The response status code.

avwhzrsChangeInfo - A complex type that describes the changes made to your hosted zone.

data AssociateVPCWithHostedZoneResponse

A complex type that contains the response information for the AssociateVPCWithHostedZone request.

See: associateVPCWithHostedZoneResponse smart constructor.

Response Lenses

avwhzrsResponseStatus :: Lens' AssociateVPCWithHostedZoneResponse Int

- | The response status code.

avwhzrsChangeInfo :: Lens' AssociateVPCWithHostedZoneResponse ChangeInfo

A complex type that describes the changes made to your hosted zone.