Network.AWS.Route53.ChangeTagsForResource

Description

Adds, edits, or deletes tags for a health check or a hosted zone.

For information about using tags for cost allocation, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide .

Creating a Request

changeTagsForResource

Arguments

:: TagResourceType
ctfrResourceType

-> Text
ctfrResourceId

-> ChangeTagsForResource

Creates a value of ChangeTagsForResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctfrRemoveTagKeys - A complex type that contains a list of the tags that you want to delete from the specified health check or hosted zone. You can specify up to 10 keys.

ctfrAddTags - A complex type that contains a list of the tags that you want to add to the specified health check or hosted zone and/or the tags that you want to edit Value for. You can add a maximum of 10 tags to a health check or a hosted zone.

ctfrResourceType - The type of the resource. * The resource type for health checks is healthcheck . * The resource type for hosted zones is hostedzone .

ctfrResourceId - The ID of the resource for which you want to add, change, or delete tags.

data ChangeTagsForResource

A complex type that contains information about the tags that you want to add, edit, or delete.

See: changeTagsForResource smart constructor.

Request Lenses

ctfrRemoveTagKeys :: Lens' ChangeTagsForResource (Maybe (NonEmpty Text))

A complex type that contains a list of the tags that you want to delete from the specified health check or hosted zone. You can specify up to 10 keys.

ctfrAddTags :: Lens' ChangeTagsForResource (Maybe (NonEmpty Tag))

A complex type that contains a list of the tags that you want to add to the specified health check or hosted zone and/or the tags that you want to edit Value for. You can add a maximum of 10 tags to a health check or a hosted zone.

ctfrResourceType :: Lens' ChangeTagsForResource TagResourceType

The type of the resource. * The resource type for health checks is healthcheck . * The resource type for hosted zones is hostedzone .

ctfrResourceId :: Lens' ChangeTagsForResource Text

The ID of the resource for which you want to add, change, or delete tags.

Destructuring the Response

changeTagsForResourceResponse

Arguments

:: Int
ctfrrsResponseStatus

-> ChangeTagsForResourceResponse

Creates a value of ChangeTagsForResourceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctfrrsResponseStatus - -- | The response status code.

data ChangeTagsForResourceResponse

Empty response for the request.

See: changeTagsForResourceResponse smart constructor.

Response Lenses

ctfrrsResponseStatus :: Lens' ChangeTagsForResourceResponse Int

- | The response status code.