Network.AWS.Route53.CreateHostedZone

Description

Creates a new public hosted zone, which you use to specify how the Domain Name System (DNS) routes traffic on the Internet for a domain, such as example.com, and its subdomains.

Important: You can't convert a public hosted zones to a private hosted zone or vice versa. Instead, you must create a new hosted zone with the same name and create new resource record sets.

For more information about charges for hosted zones, see Amazon Route 53 Pricing .

Note the following:

You can't create a hosted zone for a top-level domain (TLD).

Amazon Route 53 automatically creates a default SOA record and four NS records for the zone. For more information about SOA and NS records, see NS and SOA Records that Amazon Route 53 Creates for a Hosted Zone in the Amazon Route 53 Developer Guide .

If you want to use the same name servers for multiple hosted zones, you can optionally associate a reusable delegation set with the hosted zone. See the DelegationSetId element.

If your domain is registered with a registrar other than Amazon Route 53, you must update the name servers with your registrar to make Amazon Route 53 your DNS service. For more information, see Configuring Amazon Route 53 as your DNS Service in the Amazon Route 53 Developer Guide .

When you submit a CreateHostedZone request, the initial status of the hosted zone is PENDING . This means that the NS and SOA records are not yet available on all Amazon Route 53 DNS servers. When the NS and SOA records are available, the status of the zone changes to INSYNC .

Creating a Request

createHostedZone

Arguments

:: Text
chzName

-> Text
chzCallerReference

-> CreateHostedZone

Creates a value of CreateHostedZone with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chzDelegationSetId - If you want to associate a reusable delegation set with this hosted zone, the ID that Amazon Route 53 assigned to the reusable delegation set when you created it. For more information about reusable delegation sets, see CreateReusableDelegationSet .

chzVPC - (Private hosted zones only) A complex type that contains information about the Amazon VPC that you're associating with this hosted zone. You can specify only one Amazon VPC when you create a private hosted zone. To associate additional Amazon VPCs with the hosted zone, use AssociateVPCWithHostedZone after you create a hosted zone.

chzHostedZoneConfig - (Optional) A complex type that contains the following optional values: * For public and private hosted zones, an optional comment * For private hosted zones, an optional PrivateZone element If you don't specify a comment or the PrivateZone element, omit HostedZoneConfig and the other elements.

chzName - The name of the domain. For resource record types that include a domain name, specify a fully qualified domain name, for example, www.example.com . The trailing dot is optional; Amazon Route 53 assumes that the domain name is fully qualified. This means that Amazon Route 53 treats www.example.com (without a trailing dot) and www.example.com. (with a trailing dot) as identical. If you're creating a public hosted zone, this is the name you have registered with your DNS registrar. If your domain name is registered with a registrar other than Amazon Route 53, change the name servers for your domain to the set of NameServers that CreateHostedZone returns in DelegationSet .

chzCallerReference - A unique string that identifies the request and that allows failed CreateHostedZone requests to be retried without the risk of executing the operation twice. You must use a unique CallerReference string every time you submit a CreateHostedZone request. CallerReference can be any unique string, for example, a date/time stamp.

data CreateHostedZone

A complex type that contains information about the request to create a hosted zone.

See: createHostedZone smart constructor.

Request Lenses

chzDelegationSetId :: Lens' CreateHostedZone (Maybe ResourceId)

If you want to associate a reusable delegation set with this hosted zone, the ID that Amazon Route 53 assigned to the reusable delegation set when you created it. For more information about reusable delegation sets, see CreateReusableDelegationSet .

chzVPC :: Lens' CreateHostedZone (Maybe VPC)

(Private hosted zones only) A complex type that contains information about the Amazon VPC that you're associating with this hosted zone. You can specify only one Amazon VPC when you create a private hosted zone. To associate additional Amazon VPCs with the hosted zone, use AssociateVPCWithHostedZone after you create a hosted zone.

chzHostedZoneConfig :: Lens' CreateHostedZone (Maybe HostedZoneConfig)

(Optional) A complex type that contains the following optional values: * For public and private hosted zones, an optional comment * For private hosted zones, an optional PrivateZone element If you don't specify a comment or the PrivateZone element, omit HostedZoneConfig and the other elements.

chzName :: Lens' CreateHostedZone Text

The name of the domain. For resource record types that include a domain name, specify a fully qualified domain name, for example, www.example.com . The trailing dot is optional; Amazon Route 53 assumes that the domain name is fully qualified. This means that Amazon Route 53 treats www.example.com (without a trailing dot) and www.example.com. (with a trailing dot) as identical. If you're creating a public hosted zone, this is the name you have registered with your DNS registrar. If your domain name is registered with a registrar other than Amazon Route 53, change the name servers for your domain to the set of NameServers that CreateHostedZone returns in DelegationSet .

chzCallerReference :: Lens' CreateHostedZone Text

A unique string that identifies the request and that allows failed CreateHostedZone requests to be retried without the risk of executing the operation twice. You must use a unique CallerReference string every time you submit a CreateHostedZone request. CallerReference can be any unique string, for example, a date/time stamp.

Destructuring the Response

createHostedZoneResponse

Arguments

:: Int
chzrsResponseStatus

-> HostedZone
chzrsHostedZone

-> ChangeInfo
chzrsChangeInfo

-> DelegationSet
chzrsDelegationSet

-> Text
chzrsLocation

-> CreateHostedZoneResponse

Creates a value of CreateHostedZoneResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chzrsVPC - A complex type that contains information about an Amazon VPC that you associated with this hosted zone.

chzrsResponseStatus - -- | The response status code.

chzrsHostedZone - A complex type that contains general information about the hosted zone.

chzrsChangeInfo - A complex type that contains information about the CreateHostedZone request.

chzrsDelegationSet - A complex type that describes the name servers for this hosted zone.

chzrsLocation - The unique URL representing the new hosted zone.

data CreateHostedZoneResponse

A complex type containing the response information for the hosted zone.

See: createHostedZoneResponse smart constructor.

Response Lenses

chzrsVPC :: Lens' CreateHostedZoneResponse (Maybe VPC)

A complex type that contains information about an Amazon VPC that you associated with this hosted zone.

chzrsResponseStatus :: Lens' CreateHostedZoneResponse Int

- | The response status code.

chzrsHostedZone :: Lens' CreateHostedZoneResponse HostedZone

A complex type that contains general information about the hosted zone.

chzrsChangeInfo :: Lens' CreateHostedZoneResponse ChangeInfo

A complex type that contains information about the CreateHostedZone request.

chzrsDelegationSet :: Lens' CreateHostedZoneResponse DelegationSet

A complex type that describes the name servers for this hosted zone.

chzrsLocation :: Lens' CreateHostedZoneResponse Text

The unique URL representing the new hosted zone.