Network.AWS.Route53.DeleteHealthCheck

Description

This action deletes a health check. To delete a health check, send a DELETE request to the 2013-04-01healthcheckhealth check ID resource. You can delete a health check only if there are no resource record sets associated with this health check. If resource record sets are associated with this health check, you must disassociate them before you can delete your health check. If you try to delete a health check that is associated with resource record sets, Route 53 will deny your request with a HealthCheckInUse error. For information about disassociating the records from your health check, see ChangeResourceRecordSets.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_DeleteHealthCheck.html

Request

data DeleteHealthCheck

Request constructor

deleteHealthCheck

Arguments

:: Text
dhcHealthCheckId

-> DeleteHealthCheck

DeleteHealthCheck constructor.

The fields accessible through corresponding lenses are:

dhcHealthCheckId :: Text

Request lenses

dhcHealthCheckId :: Lens' DeleteHealthCheck Text

The ID of the health check to delete.

Response

data DeleteHealthCheckResponse

Response constructor

deleteHealthCheckResponse :: DeleteHealthCheckResponse

DeleteHealthCheckResponse constructor.