Network.AWS.Route53.DeleteHostedZone

Description

Deletes a hosted zone.

Important: If the name servers for the hosted zone are associated with a domain and if you want to make the domain unavailable on the Internet, we recommend that you delete the name servers from the domain to prevent future DNS queries from possibly being misrouted. If the domain is registered with Amazon Route 53, see UpdateDomainNameservers . If the domain is registered with another registrar, use the method provided by the registrar to delete name servers for the domain.

Some domain registries don't allow you to remove all of the name servers for a domain. If the registry for your domain requires one or more name servers, we recommend that you delete the hosted zone only if you transfer DNS service to another service provider, and you replace the name servers for the domain with name servers from the new provider.

You can delete a hosted zone only if it contains only the default SOA record and NS resource record sets. If the hosted zone contains other resource record sets, you must delete them before you can delete the hosted zone. If you try to delete a hosted zone that contains other resource record sets, the request fails, and Amazon Route 53 returns a HostedZoneNotEmpty error. For information about deleting records from your hosted zone, see ChangeResourceRecordSets .

To verify that the hosted zone has been deleted, do one of the following:

Use the GetHostedZone action to request information about the hosted zone.

Use the ListHostedZones action to get a list of the hosted zones associated with the current AWS account.

Creating a Request

deleteHostedZone

Arguments

:: ResourceId
dhzId

-> DeleteHostedZone

Creates a value of DeleteHostedZone with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhzId - The ID of the hosted zone you want to delete.

data DeleteHostedZone

A request to delete a hosted zone.

See: deleteHostedZone smart constructor.

Request Lenses

dhzId :: Lens' DeleteHostedZone ResourceId

The ID of the hosted zone you want to delete.

Destructuring the Response

deleteHostedZoneResponse

Arguments

:: Int
dhzrsResponseStatus

-> ChangeInfo
dhzrsChangeInfo

-> DeleteHostedZoneResponse

Creates a value of DeleteHostedZoneResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhzrsResponseStatus - -- | The response status code.

dhzrsChangeInfo - A complex type that contains the ID, the status, and the date and time of a request to delete a hosted zone.

data DeleteHostedZoneResponse

A complex type that contains the response to a DeleteHostedZone request.

See: deleteHostedZoneResponse smart constructor.

Response Lenses

dhzrsResponseStatus :: Lens' DeleteHostedZoneResponse Int

- | The response status code.

dhzrsChangeInfo :: Lens' DeleteHostedZoneResponse ChangeInfo

A complex type that contains the ID, the status, and the date and time of a request to delete a hosted zone.