Network.AWS.Route53.DeleteReusableDelegationSet

Description

Deletes a reusable delegation set.

Important: You can delete a reusable delegation set only if it isn't associated with any hosted zones.

To verify that the reusable delegation set is not associated with any hosted zones, submit a GetReusableDelegationSet request and specify the ID of the reusable delegation set that you want to delete.

Creating a Request

deleteReusableDelegationSet

Arguments

:: ResourceId
drdsId

-> DeleteReusableDelegationSet

Creates a value of DeleteReusableDelegationSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drdsId - The ID of the reusable delegation set that you want to delete.

data DeleteReusableDelegationSet

A request to delete a reusable delegation set.

See: deleteReusableDelegationSet smart constructor.

Request Lenses

drdsId :: Lens' DeleteReusableDelegationSet ResourceId

The ID of the reusable delegation set that you want to delete.

Destructuring the Response

deleteReusableDelegationSetResponse

Arguments

:: Int
drdsrsResponseStatus

-> DeleteReusableDelegationSetResponse

Creates a value of DeleteReusableDelegationSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drdsrsResponseStatus - -- | The response status code.

data DeleteReusableDelegationSetResponse

An empty element.

See: deleteReusableDelegationSetResponse smart constructor.

Response Lenses

drdsrsResponseStatus :: Lens' DeleteReusableDelegationSetResponse Int

- | The response status code.