Network.AWS.Route53.DisassociateVPCFromHostedZone

Description

Disassociates a VPC from a Amazon Route 53 private hosted zone.

Important: You can't disassociate a VPC from a private hosted zone when only one VPC is associated with the hosted zone. You also can't convert a private hosted zone into a public hosted zone.

Creating a Request

disassociateVPCFromHostedZone

Arguments

:: ResourceId
dvfhzHostedZoneId

-> VPC
dvfhzVPC

-> DisassociateVPCFromHostedZone

Creates a value of DisassociateVPCFromHostedZone with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvfhzComment - Optional: A comment about the disassociation request.

dvfhzHostedZoneId - The ID of the private hosted zone that you want to disassociate a VPC from.

dvfhzVPC - A complex type that contains information about the VPC that you're disassociating from the specified hosted zone.

data DisassociateVPCFromHostedZone

A complex type that contains information about the VPC that you want to disassociate from a specified private hosted zone.

See: disassociateVPCFromHostedZone smart constructor.

Request Lenses

dvfhzComment :: Lens' DisassociateVPCFromHostedZone (Maybe Text)

Optional: A comment about the disassociation request.

dvfhzHostedZoneId :: Lens' DisassociateVPCFromHostedZone ResourceId

The ID of the private hosted zone that you want to disassociate a VPC from.

dvfhzVPC :: Lens' DisassociateVPCFromHostedZone VPC

A complex type that contains information about the VPC that you're disassociating from the specified hosted zone.

Destructuring the Response

disassociateVPCFromHostedZoneResponse

Arguments

:: Int
dvfhzrsResponseStatus

-> ChangeInfo
dvfhzrsChangeInfo

-> DisassociateVPCFromHostedZoneResponse

Creates a value of DisassociateVPCFromHostedZoneResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvfhzrsResponseStatus - -- | The response status code.

dvfhzrsChangeInfo - A complex type that describes the changes made to the specified private hosted zone.

data DisassociateVPCFromHostedZoneResponse

A complex type that contains the response information for the disassociate request.

See: disassociateVPCFromHostedZoneResponse smart constructor.

Response Lenses

dvfhzrsResponseStatus :: Lens' DisassociateVPCFromHostedZoneResponse Int

- | The response status code.

dvfhzrsChangeInfo :: Lens' DisassociateVPCFromHostedZoneResponse ChangeInfo

A complex type that describes the changes made to the specified private hosted zone.