Network.AWS.Route53.GetGeoLocation

Description

Gets information about whether a specified geographic location is supported for Amazon Route 53 geolocation resource record sets.

Use the following syntax to determine whether a continent is supported for geolocation:

GET 2013-04-01geolocation?ContinentCode=two-letter abbreviation for a continent

Use the following syntax to determine whether a country is supported for geolocation:

GET 2013-04-01geolocation?CountryCode=two-character country code

Use the following syntax to determine whether a subdivision of a country is supported for geolocation:

GET 2013-04-01geolocation?CountryCode=two-character country code &SubdivisionCode=subdivision code

Creating a Request

getGeoLocation :: GetGeoLocation

Creates a value of GetGeoLocation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gglSubdivisionCode - Amazon Route 53 uses the one- to three-letter subdivision codes that are specified in ISO standard 3166-1 alpha-2 . Amazon Route 53 doesn't support subdivision codes for all countries. If you specify SubdivisionCode , you must also specify CountryCode .

gglCountryCode - Amazon Route 53 uses the two-letter country codes that are specified in ISO standard 3166-1 alpha-2 .

gglContinentCode - Amazon Route 53 supports the following continent codes: * AF : Africa * AN : Antarctica * AS : Asia * EU : Europe * OC : Oceania * NA : North America * SA : South America

data GetGeoLocation

A request for information about whether a specified geographic location is supported for Amazon Route 53 geolocation resource record sets.

See: getGeoLocation smart constructor.

Request Lenses

gglSubdivisionCode :: Lens' GetGeoLocation (Maybe Text)

Amazon Route 53 uses the one- to three-letter subdivision codes that are specified in ISO standard 3166-1 alpha-2 . Amazon Route 53 doesn't support subdivision codes for all countries. If you specify SubdivisionCode , you must also specify CountryCode .

gglCountryCode :: Lens' GetGeoLocation (Maybe Text)

Amazon Route 53 uses the two-letter country codes that are specified in ISO standard 3166-1 alpha-2 .

gglContinentCode :: Lens' GetGeoLocation (Maybe Text)

Amazon Route 53 supports the following continent codes: * AF : Africa * AN : Antarctica * AS : Asia * EU : Europe * OC : Oceania * NA : North America * SA : South America

Destructuring the Response

getGeoLocationResponse

Arguments

:: Int
gglrsResponseStatus

-> GeoLocationDetails
gglrsGeoLocationDetails

-> GetGeoLocationResponse

Creates a value of GetGeoLocationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gglrsResponseStatus - -- | The response status code.

gglrsGeoLocationDetails - A complex type that contains the codes and full continent, country, and subdivision names for the specified geolocation code.

data GetGeoLocationResponse

A complex type that contains the response information for the specified geolocation code.

See: getGeoLocationResponse smart constructor.

Response Lenses

gglrsResponseStatus :: Lens' GetGeoLocationResponse Int

- | The response status code.

gglrsGeoLocationDetails :: Lens' GetGeoLocationResponse GeoLocationDetails

A complex type that contains the codes and full continent, country, and subdivision names for the specified geolocation code.