Network.AWS.Route53.GetHealthCheckLastFailureReason

Description

Gets the reason that a specified health check failed most recently.

Creating a Request

getHealthCheckLastFailureReason

Arguments

:: Text
ghclfrHealthCheckId

-> GetHealthCheckLastFailureReason

Creates a value of GetHealthCheckLastFailureReason with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ghclfrHealthCheckId - The ID for the health check for which you want the last failure reason. When you created the health check, CreateHealthCheck returned the ID in the response, in the HealthCheckId element.

data GetHealthCheckLastFailureReason

A request for the reason that a health check failed most recently.

See: getHealthCheckLastFailureReason smart constructor.

Request Lenses

ghclfrHealthCheckId :: Lens' GetHealthCheckLastFailureReason Text

The ID for the health check for which you want the last failure reason. When you created the health check, CreateHealthCheck returned the ID in the response, in the HealthCheckId element.

Destructuring the Response

getHealthCheckLastFailureReasonResponse

Arguments

:: Int
ghclfrrsResponseStatus

-> GetHealthCheckLastFailureReasonResponse

Creates a value of GetHealthCheckLastFailureReasonResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ghclfrrsResponseStatus - -- | The response status code.

ghclfrrsHealthCheckObservations - A list that contains one Observation element for each Amazon Route 53 health checker that is reporting a last failure reason.

data GetHealthCheckLastFailureReasonResponse

A complex type that contains the response to a GetHealthCheckLastFailureReason request.

See: getHealthCheckLastFailureReasonResponse smart constructor.

Response Lenses

ghclfrrsResponseStatus :: Lens' GetHealthCheckLastFailureReasonResponse Int

- | The response status code.

ghclfrrsHealthCheckObservations :: Lens' GetHealthCheckLastFailureReasonResponse [HealthCheckObservation]

A list that contains one Observation element for each Amazon Route 53 health checker that is reporting a last failure reason.