Network.AWS.Route53.GetHostedZone

Description

To retrieve the delegation set for a hosted zone, send a GET request to the '2013-04-01hostedzonehosted zone ID resource. The delegation set is the four Route 53 name servers that were assigned to the hosted zone when you created it.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_GetHostedZone.html

Request

data GetHostedZone

Request constructor

getHostedZone

Arguments

:: Text
ghzId

-> GetHostedZone

GetHostedZone constructor.

The fields accessible through corresponding lenses are:

ghzId :: Text

Request lenses

ghzId :: Lens' GetHostedZone Text

The ID of the hosted zone for which you want to get a list of the name servers in the delegation set.

Response

data GetHostedZoneResponse

Response constructor

getHostedZoneResponse

Arguments

:: HostedZone
ghzrHostedZone

-> NonEmpty VPC
ghzrVPCs

-> GetHostedZoneResponse

GetHostedZoneResponse constructor.

The fields accessible through corresponding lenses are:

ghzrDelegationSet :: Maybe DelegationSet

ghzrHostedZone :: HostedZone

ghzrVPCs :: NonEmpty VPC

Response lenses

ghzrDelegationSet :: Lens' GetHostedZoneResponse (Maybe DelegationSet)

A complex type that contains information about the name servers for the specified hosted zone.

ghzrHostedZone :: Lens' GetHostedZoneResponse HostedZone

A complex type that contains the information about the specified hosted zone.

ghzrVPCs :: Lens' GetHostedZoneResponse (NonEmpty VPC)

A complex type that contains information about VPCs associated with the specified hosted zone.