Network.AWS.Route53.ListHealthChecks

Description

Retrieve a list of the health checks that are associated with the current AWS account.

This operation returns paginated results.

Creating a Request

listHealthChecks :: ListHealthChecks

Creates a value of ListHealthChecks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lhcMarker - If the value of IsTruncated in the previous response was true , you have more health checks. To get another group, submit another ListHealthChecks request. For the value of marker , specify the value of NextMarker from the previous response, which is the ID of the first health check that Amazon Route 53 will return if you submit another request. If the value of IsTruncated in the previous response was false , there are no more health checks to get.

lhcMaxItems - The maximum number of health checks that you want ListHealthChecks to return in response to the current request. Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Amazon Route 53 returns only the first 100 health checks.

data ListHealthChecks

A request to retrieve a list of the health checks that are associated with the current AWS account.

See: listHealthChecks smart constructor.

Request Lenses

lhcMarker :: Lens' ListHealthChecks (Maybe Text)

If the value of IsTruncated in the previous response was true , you have more health checks. To get another group, submit another ListHealthChecks request. For the value of marker , specify the value of NextMarker from the previous response, which is the ID of the first health check that Amazon Route 53 will return if you submit another request. If the value of IsTruncated in the previous response was false , there are no more health checks to get.

lhcMaxItems :: Lens' ListHealthChecks (Maybe Text)

The maximum number of health checks that you want ListHealthChecks to return in response to the current request. Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Amazon Route 53 returns only the first 100 health checks.

Destructuring the Response

listHealthChecksResponse

Arguments

:: Int
lhcrsResponseStatus

-> Text
lhcrsMarker

-> Bool
lhcrsIsTruncated

-> Text
lhcrsMaxItems

-> ListHealthChecksResponse

Creates a value of ListHealthChecksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lhcrsNextMarker - If IsTruncated is true , the value of NextMarker identifies the first health check that Amazon Route 53 returns if you submit another ListHealthChecks request and specify the value of NextMarker in the marker parameter.

lhcrsResponseStatus - -- | The response status code.

lhcrsHealthChecks - A complex type that contains one HealthCheck element for each health check that is associated with the current AWS account.

lhcrsMarker - For the second and subsequent calls to ListHealthChecks , Marker is the value that you specified for the marker parameter in the previous request.

lhcrsIsTruncated - A flag that indicates whether there are more health checks to be listed. If the response was truncated, you can get the next group of health checks by submitting another ListHealthChecks request and specifying the value of NextMarker in the marker parameter.

lhcrsMaxItems - The value that you specified for the maxitems parameter in the call to ListHealthChecks that produced the current response.

data ListHealthChecksResponse

A complex type that contains the response to a ListHealthChecks request.

See: listHealthChecksResponse smart constructor.

Response Lenses

lhcrsNextMarker :: Lens' ListHealthChecksResponse (Maybe Text)

If IsTruncated is true , the value of NextMarker identifies the first health check that Amazon Route 53 returns if you submit another ListHealthChecks request and specify the value of NextMarker in the marker parameter.

lhcrsResponseStatus :: Lens' ListHealthChecksResponse Int

- | The response status code.

lhcrsHealthChecks :: Lens' ListHealthChecksResponse [HealthCheck]

A complex type that contains one HealthCheck element for each health check that is associated with the current AWS account.

lhcrsMarker :: Lens' ListHealthChecksResponse Text

For the second and subsequent calls to ListHealthChecks , Marker is the value that you specified for the marker parameter in the previous request.

lhcrsIsTruncated :: Lens' ListHealthChecksResponse Bool

A flag that indicates whether there are more health checks to be listed. If the response was truncated, you can get the next group of health checks by submitting another ListHealthChecks request and specifying the value of NextMarker in the marker parameter.

lhcrsMaxItems :: Lens' ListHealthChecksResponse Text

The value that you specified for the maxitems parameter in the call to ListHealthChecks that produced the current response.