Network.AWS.Route53.ListHostedZonesByName

Description

Retrieves a list of your hosted zones in lexicographic order. The response includes a HostedZones child element for each hosted zone created by the current AWS account.

ListHostedZonesByName sorts hosted zones by name with the labels reversed. For example:

com.example.www.

Note the trailing dot, which can change the sort order in some circumstances.

If the domain name includes escape characters or Punycode, ListHostedZonesByName alphabetizes the domain name using the escaped or Punycoded value, which is the format that Amazon Route 53 saves in its database. For example, to create a hosted zone for exämple.com, you specify ex344mple.com for the domain name. ListHostedZonesByName alphabetizes it as:

com.ex344mple.

The labels are reversed and alphabetized using the escaped value. For more information about valid domain name formats, including internationalized domain names, see DNS Domain Name Format in the Amazon Route 53 Developer Guide .

Amazon Route 53 returns up to 100 items in each response. If you have a lot of hosted zones, use the MaxItems parameter to list them in groups of up to 100. The response includes values that help navigate from one group of MaxItems hosted zones to the next:

The DNSName and HostedZoneId elements in the response contain the values, if any, specified for the dnsname and hostedzoneid parameters in the request that produced the current response.

The MaxItems element in the response contains the value, if any, that you specified for the maxitems parameter in the request that produced the current response.

If the value of IsTruncated in the response is true, there are more hosted zones associated with the current AWS account.

If IsTruncated is false, this response includes the last hosted zone that is associated with the current account. The NextDNSName element and NextHostedZoneId elements are omitted from the response.

The NextDNSName and NextHostedZoneId elements in the response contain the domain name and the hosted zone ID of the next hosted zone that is associated with the current AWS account. If you want to list more hosted zones, make another call to ListHostedZonesByName , and specify the value of NextDNSName and NextHostedZoneId in the dnsname and hostedzoneid parameters, respectively.

Creating a Request

listHostedZonesByName :: ListHostedZonesByName

Creates a value of ListHostedZonesByName with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lhzbnHostedZoneId - (Optional) For your first request to ListHostedZonesByName , do not include the hostedzoneid parameter. If you have more hosted zones than the value of maxitems , ListHostedZonesByName returns only the first maxitems hosted zones. To get the next group of maxitems hosted zones, submit another request to ListHostedZonesByName and include both dnsname and hostedzoneid parameters. For the value of hostedzoneid , specify the value of the NextHostedZoneId element from the previous response.

lhzbnMaxItems - The maximum number of hosted zones to be included in the response body for this request. If you have more than maxitems hosted zones, then the value of the IsTruncated element in the response is true, and the values of NextDNSName and NextHostedZoneId specify the first hosted zone in the next group of maxitems hosted zones.

lhzbnDNSName - (Optional) For your first request to ListHostedZonesByName , include the dnsname parameter only if you want to specify the name of the first hosted zone in the response. If you don't include the dnsname parameter, Amazon Route 53 returns all of the hosted zones that were created by the current AWS account, in ASCII order. For subsequent requests, include both dnsname and hostedzoneid parameters. For dnsname , specify the value of NextDNSName from the previous response.

data ListHostedZonesByName

Retrieves a list of the public and private hosted zones that are associated with the current AWS account in ASCII order by domain name.

See: listHostedZonesByName smart constructor.

Request Lenses

lhzbnHostedZoneId :: Lens' ListHostedZonesByName (Maybe ResourceId)

(Optional) For your first request to ListHostedZonesByName , do not include the hostedzoneid parameter. If you have more hosted zones than the value of maxitems , ListHostedZonesByName returns only the first maxitems hosted zones. To get the next group of maxitems hosted zones, submit another request to ListHostedZonesByName and include both dnsname and hostedzoneid parameters. For the value of hostedzoneid , specify the value of the NextHostedZoneId element from the previous response.

lhzbnMaxItems :: Lens' ListHostedZonesByName (Maybe Text)

The maximum number of hosted zones to be included in the response body for this request. If you have more than maxitems hosted zones, then the value of the IsTruncated element in the response is true, and the values of NextDNSName and NextHostedZoneId specify the first hosted zone in the next group of maxitems hosted zones.

lhzbnDNSName :: Lens' ListHostedZonesByName (Maybe Text)

(Optional) For your first request to ListHostedZonesByName , include the dnsname parameter only if you want to specify the name of the first hosted zone in the response. If you don't include the dnsname parameter, Amazon Route 53 returns all of the hosted zones that were created by the current AWS account, in ASCII order. For subsequent requests, include both dnsname and hostedzoneid parameters. For dnsname , specify the value of NextDNSName from the previous response.

Destructuring the Response

listHostedZonesByNameResponse

Arguments

:: Int
lhzbnrsResponseStatus

-> Bool
lhzbnrsIsTruncated

-> Text
lhzbnrsMaxItems

-> ListHostedZonesByNameResponse

Creates a value of ListHostedZonesByNameResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lhzbnrsHostedZoneId - The ID that Amazon Route 53 assigned to the hosted zone when you created it.

lhzbnrsNextHostedZoneId - If IsTruncated is true , the value of NextHostedZoneId identifies the first hosted zone in the next group of maxitems hosted zones. Call ListHostedZonesByName again and specify the value of NextDNSName and NextHostedZoneId in the dnsname and hostedzoneid parameters, respectively. This element is present only if IsTruncated is true .

lhzbnrsDNSName - For the second and subsequent calls to ListHostedZonesByName , DNSName is the value that you specified for the dnsname parameter in the request that produced the current response.

lhzbnrsNextDNSName - If IsTruncated is true, the value of NextDNSName is the name of the first hosted zone in the next group of maxitems hosted zones. Call ListHostedZonesByName again and specify the value of NextDNSName and NextHostedZoneId in the dnsname and hostedzoneid parameters, respectively. This element is present only if IsTruncated is true .

lhzbnrsResponseStatus - -- | The response status code.

lhzbnrsHostedZones - A complex type that contains general information about the hosted zone.

lhzbnrsIsTruncated - A flag that indicates whether there are more hosted zones to be listed. If the response was truncated, you can get the next group of maxitems hosted zones by calling ListHostedZonesByName again and specifying the values of NextDNSName and NextHostedZoneId elements in the dnsname and hostedzoneid parameters.

lhzbnrsMaxItems - The value that you specified for the maxitems parameter in the call to ListHostedZonesByName that produced the current response.

data ListHostedZonesByNameResponse

A complex type that contains the response information for the request.

See: listHostedZonesByNameResponse smart constructor.

Response Lenses

lhzbnrsHostedZoneId :: Lens' ListHostedZonesByNameResponse (Maybe ResourceId)

The ID that Amazon Route 53 assigned to the hosted zone when you created it.

lhzbnrsNextHostedZoneId :: Lens' ListHostedZonesByNameResponse (Maybe ResourceId)

If IsTruncated is true , the value of NextHostedZoneId identifies the first hosted zone in the next group of maxitems hosted zones. Call ListHostedZonesByName again and specify the value of NextDNSName and NextHostedZoneId in the dnsname and hostedzoneid parameters, respectively. This element is present only if IsTruncated is true .

lhzbnrsDNSName :: Lens' ListHostedZonesByNameResponse (Maybe Text)

For the second and subsequent calls to ListHostedZonesByName , DNSName is the value that you specified for the dnsname parameter in the request that produced the current response.

lhzbnrsNextDNSName :: Lens' ListHostedZonesByNameResponse (Maybe Text)

If IsTruncated is true, the value of NextDNSName is the name of the first hosted zone in the next group of maxitems hosted zones. Call ListHostedZonesByName again and specify the value of NextDNSName and NextHostedZoneId in the dnsname and hostedzoneid parameters, respectively. This element is present only if IsTruncated is true .

lhzbnrsResponseStatus :: Lens' ListHostedZonesByNameResponse Int

- | The response status code.

lhzbnrsHostedZones :: Lens' ListHostedZonesByNameResponse [HostedZone]

A complex type that contains general information about the hosted zone.

lhzbnrsIsTruncated :: Lens' ListHostedZonesByNameResponse Bool

A flag that indicates whether there are more hosted zones to be listed. If the response was truncated, you can get the next group of maxitems hosted zones by calling ListHostedZonesByName again and specifying the values of NextDNSName and NextHostedZoneId elements in the dnsname and hostedzoneid parameters.

lhzbnrsMaxItems :: Lens' ListHostedZonesByNameResponse Text

The value that you specified for the maxitems parameter in the call to ListHostedZonesByName that produced the current response.