Network.AWS.Route53.ListResourceRecordSets

Description

Lists the resource record sets in a specified hosted zone.

ListResourceRecordSets returns up to 100 resource record sets at a time in ASCII order, beginning at a position specified by the name and type elements. The action sorts results first by DNS name with the labels reversed, for example:

com.example.www.

Note the trailing dot, which can change the sort order in some circumstances.

When multiple records have the same DNS name, the action sorts results by the record type.

You can use the name and type elements to adjust the beginning position of the list of resource record sets returned:

If you do not specify Name or Type * The results begin with the first resource record set that the hosted zone contains.

If you specify Name but not Type * The results begin with the first resource record set in the list whose name is greater than or equal to Name .

If you specify Type but not Name * Amazon Route 53 returns the InvalidInput error.

If you specify both Name and Type * The results begin with the first resource record set in the list whose name is greater than or equal to Name , and whose type is greater than or equal to Type .

This action returns the most current version of the records. This includes records that are PENDING , and that are not yet available on all Amazon Route 53 DNS servers.

To ensure that you get an accurate listing of the resource record sets for a hosted zone at a point in time, do not submit a ChangeResourceRecordSets request while you're paging through the results of a ListResourceRecordSets request. If you do, some pages may display results without the latest changes while other pages display results with the latest changes.

This operation returns paginated results.

Creating a Request

listResourceRecordSets

Arguments

:: ResourceId
lrrsHostedZoneId

-> ListResourceRecordSets

Creates a value of ListResourceRecordSets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrrsStartRecordName - The first name in the lexicographic ordering of resource record sets that you want to list.

lrrsStartRecordType - The type of resource record set to begin the record listing from. Valid values for basic resource record sets: A | AAAA | CNAME | MX | NAPTR | NS | PTR | SOA | SPF | SRV | TXT Values for weighted, latency, geo, and failover resource record sets: A | AAAA | CNAME | MX | NAPTR | PTR | SPF | SRV | TXT Values for alias resource record sets: * CloudFront distribution : A or AAAA * Elastic Beanstalk environment that has a regionalized subdomain : A * ELB load balancer : A | AAAA * Amazon S3 bucket : A Constraint: Specifying type without specifying name returns an InvalidInput error.

lrrsStartRecordIdentifier - Weighted resource record sets only: If results were truncated for a given DNS name and type, specify the value of NextRecordIdentifier from the previous response to get the next resource record set that has the current DNS name and type.

lrrsMaxItems - (Optional) The maximum number of resource records sets to include in the response body for this request. If the response includes more than maxitems resource record sets, the value of the IsTruncated element in the response is true , and the values of the NextRecordName and NextRecordType elements in the response identify the first resource record set in the next group of maxitems resource record sets.

lrrsHostedZoneId - The ID of the hosted zone that contains the resource record sets that you want to list.

data ListResourceRecordSets

A request for the resource record sets that are associated with a specified hosted zone.

See: listResourceRecordSets smart constructor.

Request Lenses

lrrsStartRecordName :: Lens' ListResourceRecordSets (Maybe Text)

The first name in the lexicographic ordering of resource record sets that you want to list.

lrrsStartRecordType :: Lens' ListResourceRecordSets (Maybe RecordType)

The type of resource record set to begin the record listing from. Valid values for basic resource record sets: A | AAAA | CNAME | MX | NAPTR | NS | PTR | SOA | SPF | SRV | TXT Values for weighted, latency, geo, and failover resource record sets: A | AAAA | CNAME | MX | NAPTR | PTR | SPF | SRV | TXT Values for alias resource record sets: * CloudFront distribution : A or AAAA * Elastic Beanstalk environment that has a regionalized subdomain : A * ELB load balancer : A | AAAA * Amazon S3 bucket : A Constraint: Specifying type without specifying name returns an InvalidInput error.

lrrsStartRecordIdentifier :: Lens' ListResourceRecordSets (Maybe Text)

Weighted resource record sets only: If results were truncated for a given DNS name and type, specify the value of NextRecordIdentifier from the previous response to get the next resource record set that has the current DNS name and type.

lrrsMaxItems :: Lens' ListResourceRecordSets (Maybe Text)

(Optional) The maximum number of resource records sets to include in the response body for this request. If the response includes more than maxitems resource record sets, the value of the IsTruncated element in the response is true , and the values of the NextRecordName and NextRecordType elements in the response identify the first resource record set in the next group of maxitems resource record sets.

lrrsHostedZoneId :: Lens' ListResourceRecordSets ResourceId

The ID of the hosted zone that contains the resource record sets that you want to list.

Destructuring the Response

listResourceRecordSetsResponse

Arguments

:: Int
lrrsrsResponseStatus

-> Bool
lrrsrsIsTruncated

-> Text
lrrsrsMaxItems

-> ListResourceRecordSetsResponse

Creates a value of ListResourceRecordSetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrrsrsNextRecordType - If the results were truncated, the type of the next record in the list. This element is present only if IsTruncated is true.

lrrsrsNextRecordName - If the results were truncated, the name of the next record in the list. This element is present only if IsTruncated is true.

lrrsrsNextRecordIdentifier - Weighted, latency, geolocation, and failover resource record sets only : If results were truncated for a given DNS name and type, the value of SetIdentifier for the next resource record set that has the current DNS name and type.

lrrsrsResponseStatus - -- | The response status code.

lrrsrsResourceRecordSets - Information about multiple resource record sets.

lrrsrsIsTruncated - A flag that indicates whether more resource record sets remain to be listed. If your results were truncated, you can make a follow-up pagination request by using the NextRecordName element.

lrrsrsMaxItems - The maximum number of records you requested.

data ListResourceRecordSetsResponse

A complex type that contains list information for the resource record set.

See: listResourceRecordSetsResponse smart constructor.

Response Lenses

lrrsrsNextRecordType :: Lens' ListResourceRecordSetsResponse (Maybe RecordType)

If the results were truncated, the type of the next record in the list. This element is present only if IsTruncated is true.

lrrsrsNextRecordName :: Lens' ListResourceRecordSetsResponse (Maybe Text)

If the results were truncated, the name of the next record in the list. This element is present only if IsTruncated is true.

lrrsrsNextRecordIdentifier :: Lens' ListResourceRecordSetsResponse (Maybe Text)

Weighted, latency, geolocation, and failover resource record sets only : If results were truncated for a given DNS name and type, the value of SetIdentifier for the next resource record set that has the current DNS name and type.

lrrsrsResponseStatus :: Lens' ListResourceRecordSetsResponse Int

- | The response status code.

lrrsrsResourceRecordSets :: Lens' ListResourceRecordSetsResponse [ResourceRecordSet]

Information about multiple resource record sets.

lrrsrsIsTruncated :: Lens' ListResourceRecordSetsResponse Bool

A flag that indicates whether more resource record sets remain to be listed. If your results were truncated, you can make a follow-up pagination request by using the NextRecordName element.

lrrsrsMaxItems :: Lens' ListResourceRecordSetsResponse Text

The maximum number of records you requested.