Network.AWS.Route53.UpdateHealthCheck

Description

Updates an existing health check. Note that some values can't be updated.

For more information about updating health checks, see Creating, Updating, and Deleting Health Checks in the Amazon Route 53 Developer Guide .

Creating a Request

updateHealthCheck

Arguments

:: Text
uhcHealthCheckId

-> UpdateHealthCheck

Creates a value of UpdateHealthCheck with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uhcFailureThreshold - The number of consecutive health checks that an endpoint must pass or fail for Amazon Route 53 to change the current status of the endpoint from unhealthy to healthy or vice versa. For more information, see How Amazon Route 53 Determines Whether an Endpoint Is Healthy in the Amazon Route 53 Developer Guide . If you don't specify a value for FailureThreshold , the default value is three health checks.

uhcIPAddress - The IPv4 or IPv6 IP address for the endpoint that you want Amazon Route 53 to perform health checks on. If you don't specify a value for IPAddress , Amazon Route 53 sends a DNS request to resolve the domain name that you specify in FullyQualifiedDomainName at the interval that you specify in RequestInterval . Using an IP address that is returned by DNS, Amazon Route 53 then checks the health of the endpoint. Use one of the following formats for the value of IPAddress : * IPv4 address : four values between 0 and 255, separated by periods (.), for example, 192.0.2.44 . * IPv6 address : eight groups of four hexadecimal values, separated by colons (:), for example, 2001:0db8:85a3:0000:0000:abcd:0001:2345 . You can also shorten IPv6 addresses as described in RFC 5952, for example, 2001:db8:85a3::abcd:1:2345 . If the endpoint is an EC2 instance, we recommend that you create an Elastic IP address, associate it with your EC2 instance, and specify the Elastic IP address for IPAddress . This ensures that the IP address of your instance never changes. For more information, see the applicable documentation: * Linux: Elastic IP Addresses (EIP) in the Amazon EC2 User Guide for Linux Instances * Windows: Elastic IP Addresses (EIP) in the Amazon EC2 User Guide for Windows Instances For more information, see 'UpdateHealthCheckRequest$FullyQualifiedDomainName' . Constraints: Amazon Route 53 can't check the health of endpoints for which the IP address is in local, private, non-routable, or multicast ranges. For more information about IP addresses for which you can't create health checks, see the following documents: * RFC 5735, Special Use IPv4 Addresses * RFC 6598, IANA-Reserved IPv4 Prefix for Shared Address Space * RFC 5156, Special-Use IPv6 Addresses

uhcEnableSNI - Specify whether you want Amazon Route 53 to send the value of FullyQualifiedDomainName to the endpoint in the client_hello message during TLS negotiation. This allows the endpoint to respond to HTTPS health check requests with the applicable SSLTLS certificate. Some endpoints require that HTTPS requests include the host name in the client_hello message. If you don't enable SNI, the status of the health check will be SSL alert handshake_failure . A health check can also have that status for other reasons. If SNI is enabled and you're still getting the error, check the SSLTLS configuration on your endpoint and confirm that your certificate is valid. The SSL/TLS certificate on your endpoint includes a domain name in the Common Name field and possibly several more in the Subject Alternative Names field. One of the domain names in the certificate should match the value that you specify for FullyQualifiedDomainName . If the endpoint responds to the client_hello message with a certificate that does not include the domain name that you specified in FullyQualifiedDomainName , a health checker will retry the handshake. In the second attempt, the health checker will omit FullyQualifiedDomainName from the client_hello message.

uhcSearchString - If the value of Type is HTTP_STR_MATCH or HTTP_STR_MATCH , the string that you want Amazon Route 53 to search for in the response body from the specified resource. If the string appears in the response body, Amazon Route 53 considers the resource healthy. (You can't change the value of Type when you update a health check.)

uhcHealthThreshold - The number of child health checks that are associated with a CALCULATED health that Amazon Route 53 must consider healthy for the CALCULATED health check to be considered healthy. To specify the child health checks that you want to associate with a CALCULATED health check, use the ChildHealthChecks and ChildHealthCheck elements. Note the following: * If you specify a number greater than the number of child health checks, Amazon Route 53 always considers this health check to be unhealthy. * If you specify 0 , Amazon Route 53 always considers this health check to be healthy.

uhcRegions - A complex type that contains one Region element for each region that you want Amazon Route 53 health checkers to check the specified endpoint from.

uhcResourcePath - The path that you want Amazon Route 53 to request when performing health checks. The path can be any value for which your endpoint will return an HTTP status code of 2xx or 3xx when the endpoint is healthy, for example the file docsroute53-health-check.html. Specify this value only if you want to change it.

uhcInsufficientDataHealthStatus - When CloudWatch has insufficient data about the metric to determine the alarm state, the status that you want Amazon Route 53 to assign to the health check: * Healthy : Amazon Route 53 considers the health check to be healthy. * Unhealthy : Amazon Route 53 considers the health check to be unhealthy. * LastKnownStatus : Amazon Route 53 uses the status of the health check from the last time CloudWatch had sufficient data to determine the alarm state. For new health checks that have no last known status, the default status for the health check is healthy.

uhcHealthCheckVersion - A sequential counter that Amazon Route 53 sets to 1 when you create a health check and increments by 1 each time you update settings for the health check. We recommend that you use GetHealthCheck or ListHealthChecks to get the current value of HealthCheckVersion for the health check that you want to update, and that you include that value in your UpdateHealthCheck request. This prevents Amazon Route 53 from overwriting an intervening update: * If the value in the UpdateHealthCheck request matches the value of HealthCheckVersion in the health check, Amazon Route 53 updates the health check with the new settings. * If the value of HealthCheckVersion in the health check is greater, the health check was changed after you got the version number. Amazon Route 53 does not update the health check, and it returns a HealthCheckVersionMismatch error.

uhcAlarmIdentifier - Undocumented member.

uhcInverted - Specify whether you want Amazon Route 53 to invert the status of a health check, for example, to consider a health check unhealthy when it otherwise would be considered healthy.

uhcFullyQualifiedDomainName - Amazon Route 53 behavior depends on whether you specify a value for IPAddress . If you specify a value for IPAddress : Amazon Route 53 sends health check requests to the specified IPv4 or IPv6 address and passes the value of FullyQualifiedDomainName in the Host header for all health checks except TCP health checks. This is typically the fully qualified DNS name of the endpoint on which you want Amazon Route 53 to perform health checks. When Amazon Route 53 checks the health of an endpoint, here is how it constructs the Host header: * If you specify a value of 80 for Port and HTTP or HTTP_STR_MATCH for Type , Amazon Route 53 passes the value of FullyQualifiedDomainName to the endpoint in the Host header. * If you specify a value of 443 for Port and HTTPS or HTTPS_STR_MATCH for Type , Amazon Route 53 passes the value of FullyQualifiedDomainName to the endpoint in the Host header. * If you specify another value for Port and any value except TCP for Type , Amazon Route 53 passes FullyQualifiedDomainName :Port to the endpoint in the Host header. If you don't specify a value for FullyQualifiedDomainName , Amazon Route 53 substitutes the value of IPAddress in the Host header in each of the above cases. If you don't specify a value for IPAddress : If you don't specify a value for IPAddress , Amazon Route 53 sends a DNS request to the domain that you specify in FullyQualifiedDomainName at the interval you specify in RequestInterval . Using an IPv4 address that is returned by DNS, Amazon Route 53 then checks the health of the endpoint. If you want to check the health of weighted, latency, or failover resource record sets and you choose to specify the endpoint only by FullyQualifiedDomainName , we recommend that you create a separate health check for each endpoint. For example, create a health check for each HTTP server that is serving content for www.example.com. For the value of FullyQualifiedDomainName , specify the domain name of the server (such as us-east-2-www.example.com ), not the name of the resource record sets (www.example.com). Important: In this configuration, if the value of FullyQualifiedDomainName matches the name of the resource record sets and you then associate the health check with those resource record sets, health check results will be unpredictable. In addition, if the value of Type is HTTP , HTTPS , HTTP_STR_MATCH , or HTTPS_STR_MATCH , Amazon Route 53 passes the value of FullyQualifiedDomainName in the Host header, as it does when you specify a value for IPAddress . If the value of Type is TCP , Amazon Route 53 doesn't pass a Host header.

uhcChildHealthChecks - A complex type that contains one ChildHealthCheck element for each health check that you want to associate with a CALCULATED health check.

uhcPort - The port on the endpoint on which you want Amazon Route 53 to perform health checks.

uhcHealthCheckId - The ID for the health check for which you want detailed information. When you created the health check, CreateHealthCheck returned the ID in the response, in the HealthCheckId element.

data UpdateHealthCheck

A complex type that contains information about a request to update a health check.

See: updateHealthCheck smart constructor.

Request Lenses

uhcFailureThreshold :: Lens' UpdateHealthCheck (Maybe Natural)

The number of consecutive health checks that an endpoint must pass or fail for Amazon Route 53 to change the current status of the endpoint from unhealthy to healthy or vice versa. For more information, see How Amazon Route 53 Determines Whether an Endpoint Is Healthy in the Amazon Route 53 Developer Guide . If you don't specify a value for FailureThreshold , the default value is three health checks.

uhcIPAddress :: Lens' UpdateHealthCheck (Maybe Text)

The IPv4 or IPv6 IP address for the endpoint that you want Amazon Route 53 to perform health checks on. If you don't specify a value for IPAddress , Amazon Route 53 sends a DNS request to resolve the domain name that you specify in FullyQualifiedDomainName at the interval that you specify in RequestInterval . Using an IP address that is returned by DNS, Amazon Route 53 then checks the health of the endpoint. Use one of the following formats for the value of IPAddress : * IPv4 address : four values between 0 and 255, separated by periods (.), for example, 192.0.2.44 . * IPv6 address : eight groups of four hexadecimal values, separated by colons (:), for example, 2001:0db8:85a3:0000:0000:abcd:0001:2345 . You can also shorten IPv6 addresses as described in RFC 5952, for example, 2001:db8:85a3::abcd:1:2345 . If the endpoint is an EC2 instance, we recommend that you create an Elastic IP address, associate it with your EC2 instance, and specify the Elastic IP address for IPAddress . This ensures that the IP address of your instance never changes. For more information, see the applicable documentation: * Linux: Elastic IP Addresses (EIP) in the Amazon EC2 User Guide for Linux Instances * Windows: Elastic IP Addresses (EIP) in the Amazon EC2 User Guide for Windows Instances For more information, see 'UpdateHealthCheckRequest$FullyQualifiedDomainName' . Constraints: Amazon Route 53 can't check the health of endpoints for which the IP address is in local, private, non-routable, or multicast ranges. For more information about IP addresses for which you can't create health checks, see the following documents: * RFC 5735, Special Use IPv4 Addresses * RFC 6598, IANA-Reserved IPv4 Prefix for Shared Address Space * RFC 5156, Special-Use IPv6 Addresses

uhcEnableSNI :: Lens' UpdateHealthCheck (Maybe Bool)

Specify whether you want Amazon Route 53 to send the value of FullyQualifiedDomainName to the endpoint in the client_hello message during TLS negotiation. This allows the endpoint to respond to HTTPS health check requests with the applicable SSLTLS certificate. Some endpoints require that HTTPS requests include the host name in the client_hello message. If you don't enable SNI, the status of the health check will be SSL alert handshake_failure . A health check can also have that status for other reasons. If SNI is enabled and you're still getting the error, check the SSLTLS configuration on your endpoint and confirm that your certificate is valid. The SSL/TLS certificate on your endpoint includes a domain name in the Common Name field and possibly several more in the Subject Alternative Names field. One of the domain names in the certificate should match the value that you specify for FullyQualifiedDomainName . If the endpoint responds to the client_hello message with a certificate that does not include the domain name that you specified in FullyQualifiedDomainName , a health checker will retry the handshake. In the second attempt, the health checker will omit FullyQualifiedDomainName from the client_hello message.

uhcSearchString :: Lens' UpdateHealthCheck (Maybe Text)

If the value of Type is HTTP_STR_MATCH or HTTP_STR_MATCH , the string that you want Amazon Route 53 to search for in the response body from the specified resource. If the string appears in the response body, Amazon Route 53 considers the resource healthy. (You can't change the value of Type when you update a health check.)

uhcHealthThreshold :: Lens' UpdateHealthCheck (Maybe Natural)

The number of child health checks that are associated with a CALCULATED health that Amazon Route 53 must consider healthy for the CALCULATED health check to be considered healthy. To specify the child health checks that you want to associate with a CALCULATED health check, use the ChildHealthChecks and ChildHealthCheck elements. Note the following: * If you specify a number greater than the number of child health checks, Amazon Route 53 always considers this health check to be unhealthy. * If you specify 0 , Amazon Route 53 always considers this health check to be healthy.

uhcRegions :: Lens' UpdateHealthCheck (Maybe (NonEmpty HealthCheckRegion))

A complex type that contains one Region element for each region that you want Amazon Route 53 health checkers to check the specified endpoint from.

uhcResourcePath :: Lens' UpdateHealthCheck (Maybe Text)

The path that you want Amazon Route 53 to request when performing health checks. The path can be any value for which your endpoint will return an HTTP status code of 2xx or 3xx when the endpoint is healthy, for example the file docsroute53-health-check.html. Specify this value only if you want to change it.

uhcInsufficientDataHealthStatus :: Lens' UpdateHealthCheck (Maybe InsufficientDataHealthStatus)

When CloudWatch has insufficient data about the metric to determine the alarm state, the status that you want Amazon Route 53 to assign to the health check: * Healthy : Amazon Route 53 considers the health check to be healthy. * Unhealthy : Amazon Route 53 considers the health check to be unhealthy. * LastKnownStatus : Amazon Route 53 uses the status of the health check from the last time CloudWatch had sufficient data to determine the alarm state. For new health checks that have no last known status, the default status for the health check is healthy.

uhcHealthCheckVersion :: Lens' UpdateHealthCheck (Maybe Natural)

A sequential counter that Amazon Route 53 sets to 1 when you create a health check and increments by 1 each time you update settings for the health check. We recommend that you use GetHealthCheck or ListHealthChecks to get the current value of HealthCheckVersion for the health check that you want to update, and that you include that value in your UpdateHealthCheck request. This prevents Amazon Route 53 from overwriting an intervening update: * If the value in the UpdateHealthCheck request matches the value of HealthCheckVersion in the health check, Amazon Route 53 updates the health check with the new settings. * If the value of HealthCheckVersion in the health check is greater, the health check was changed after you got the version number. Amazon Route 53 does not update the health check, and it returns a HealthCheckVersionMismatch error.

uhcAlarmIdentifier :: Lens' UpdateHealthCheck (Maybe AlarmIdentifier)

Undocumented member.

uhcInverted :: Lens' UpdateHealthCheck (Maybe Bool)

Specify whether you want Amazon Route 53 to invert the status of a health check, for example, to consider a health check unhealthy when it otherwise would be considered healthy.

uhcFullyQualifiedDomainName :: Lens' UpdateHealthCheck (Maybe Text)

Amazon Route 53 behavior depends on whether you specify a value for IPAddress . If you specify a value for IPAddress : Amazon Route 53 sends health check requests to the specified IPv4 or IPv6 address and passes the value of FullyQualifiedDomainName in the Host header for all health checks except TCP health checks. This is typically the fully qualified DNS name of the endpoint on which you want Amazon Route 53 to perform health checks. When Amazon Route 53 checks the health of an endpoint, here is how it constructs the Host header: * If you specify a value of 80 for Port and HTTP or HTTP_STR_MATCH for Type , Amazon Route 53 passes the value of FullyQualifiedDomainName to the endpoint in the Host header. * If you specify a value of 443 for Port and HTTPS or HTTPS_STR_MATCH for Type , Amazon Route 53 passes the value of FullyQualifiedDomainName to the endpoint in the Host header. * If you specify another value for Port and any value except TCP for Type , Amazon Route 53 passes FullyQualifiedDomainName :Port to the endpoint in the Host header. If you don't specify a value for FullyQualifiedDomainName , Amazon Route 53 substitutes the value of IPAddress in the Host header in each of the above cases. If you don't specify a value for IPAddress : If you don't specify a value for IPAddress , Amazon Route 53 sends a DNS request to the domain that you specify in FullyQualifiedDomainName at the interval you specify in RequestInterval . Using an IPv4 address that is returned by DNS, Amazon Route 53 then checks the health of the endpoint. If you want to check the health of weighted, latency, or failover resource record sets and you choose to specify the endpoint only by FullyQualifiedDomainName , we recommend that you create a separate health check for each endpoint. For example, create a health check for each HTTP server that is serving content for www.example.com. For the value of FullyQualifiedDomainName , specify the domain name of the server (such as us-east-2-www.example.com ), not the name of the resource record sets (www.example.com). Important: In this configuration, if the value of FullyQualifiedDomainName matches the name of the resource record sets and you then associate the health check with those resource record sets, health check results will be unpredictable. In addition, if the value of Type is HTTP , HTTPS , HTTP_STR_MATCH , or HTTPS_STR_MATCH , Amazon Route 53 passes the value of FullyQualifiedDomainName in the Host header, as it does when you specify a value for IPAddress . If the value of Type is TCP , Amazon Route 53 doesn't pass a Host header.

uhcChildHealthChecks :: Lens' UpdateHealthCheck [Text]

A complex type that contains one ChildHealthCheck element for each health check that you want to associate with a CALCULATED health check.

uhcPort :: Lens' UpdateHealthCheck (Maybe Natural)

The port on the endpoint on which you want Amazon Route 53 to perform health checks.

uhcHealthCheckId :: Lens' UpdateHealthCheck Text

The ID for the health check for which you want detailed information. When you created the health check, CreateHealthCheck returned the ID in the response, in the HealthCheckId element.

Destructuring the Response

updateHealthCheckResponse

Arguments

:: Int
uhcrsResponseStatus

-> HealthCheck
uhcrsHealthCheck

-> UpdateHealthCheckResponse

Creates a value of UpdateHealthCheckResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uhcrsResponseStatus - -- | The response status code.

uhcrsHealthCheck - Undocumented member.

data UpdateHealthCheckResponse

See: updateHealthCheckResponse smart constructor.

Response Lenses

uhcrsResponseStatus :: Lens' UpdateHealthCheckResponse Int

- | The response status code.

uhcrsHealthCheck :: Lens' UpdateHealthCheckResponse HealthCheck

Undocumented member.