Network.AWS.Route53.V2013_04_01.GetHostedZone

Description

To retrieve the delegation set for a hosted zone, send a GET request to the 2013-04-01hostedzonehosted zone ID resource. The delegation set is the four Route 53 name servers that were assigned to the hosted zone when you created it.

Documentation

data GetHostedZone

Constructors

GetHostedZone

Fields

_ghzrId :: Text

The ID of the hosted zone for which you want to get a list of the name servers in the delegation set.

ghzrId :: Iso' GetHostedZone Text

data GetHostedZoneResponse

Constructors

GetHostedZoneResponse

Fields

_ghzsDelegationSet :: DelegationSet

A complex type that contains information about the name servers for the specified hosted zone.

_ghzsHostedZone :: HostedZone

A complex type that contains the information about the specified hosted zone.

ghzsHostedZone :: Lens' GetHostedZoneResponse HostedZone

ghzsDelegationSet :: Lens' GetHostedZoneResponse DelegationSet