Network.AWS.Route53.V2013_04_01.ListGeoLocations

Documentation

listGeoLocations :: ListGeoLocations

Minimum specification for a ListGeoLocations request.

data ListGeoLocations

Constructors

ListGeoLocations

Fields

_lglrStartContinentCode :: Maybe Text

The first continent code in the lexicographic ordering of geo locations that you want the ListGeoLocations request to list. For non-continent geo locations, this should be null. Valid values: AF | AN | AS | EU | OC | NA | SA Constraint: Specifying ContinentCode with either CountryCode or SubdivisionCode returns an InvalidInput error.

_lglrStartCountryCode :: Maybe Text

The first country code in the lexicographic ordering of geo locations that you want the ListGeoLocations request to list. The default geo location uses a * for the country code. All other country codes follow the ISO 3166 two-character code.

_lglrStartSubdivisionCode :: Maybe Text

The first subdivision code in the lexicographic ordering of geo locations that you want the ListGeoLocations request to list. Constraint: Specifying SubdivisionCode without CountryCode returns an InvalidInput error.

_lglrMaxItems :: Maybe Text

The maximum number of geo locations you want in the response body.

lglrStartSubdivisionCode :: Lens' ListGeoLocations (Maybe Text)

lglrStartCountryCode :: Lens' ListGeoLocations (Maybe Text)

lglrStartContinentCode :: Lens' ListGeoLocations (Maybe Text)

lglrMaxItems :: Lens' ListGeoLocations (Maybe Text)

data ListGeoLocationsResponse

Constructors

ListGeoLocationsResponse

Fields

_lglsGeoLocationDetailsList :: [GeoLocationDetails]

A complex type that contains information about the geo locations that are returned by the request.

_lglsMaxItems :: Text

The maximum number of records you requested. The maximum value of MaxItems is 100.

_lglsIsTruncated :: Bool

A flag that indicates whether there are more geo locations to be listed. If your results were truncated, you can make a follow-up request for the next page of results by using the values included in the ListGeoLocationsResponse$NextContinentCode, ListGeoLocationsResponse$NextCountryCode and ListGeoLocationsResponse$NextSubdivisionCode elements. Valid Values: true | false.

_lglsNextContinentCode :: Maybe Text

If the results were truncated, the continent code of the next geo location in the list. This element is present only if ListGeoLocationsResponse$IsTruncated is true and the next geo location to list is a continent location.

_lglsNextCountryCode :: Maybe Text

If the results were truncated, the country code of the next geo location in the list. This element is present only if ListGeoLocationsResponse$IsTruncated is true and the next geo location to list is not a continent location.

_lglsNextSubdivisionCode :: Maybe Text

If the results were truncated, the subdivision code of the next geo location in the list. This element is present only if ListGeoLocationsResponse$IsTruncated is true and the next geo location has a subdivision.

lglsNextSubdivisionCode :: Lens' ListGeoLocationsResponse (Maybe Text)

lglsNextCountryCode :: Lens' ListGeoLocationsResponse (Maybe Text)

lglsNextContinentCode :: Lens' ListGeoLocationsResponse (Maybe Text)

lglsMaxItems :: Lens' ListGeoLocationsResponse Text

lglsIsTruncated :: Lens' ListGeoLocationsResponse Bool

lglsGeoLocationDetailsList :: Lens' ListGeoLocationsResponse [GeoLocationDetails]